Struct twitch_api::client::Body

pub struct Body { /* private fields */ }

Available on crate feature client only.

A stream of Bytes, used when receiving bodies.

A good default HttpBody to use in many applications.

Note: To read the full body, use body::to_bytes or body::aggregate.

Implementations

impl Body

pub fn empty() -> Body

Create an empty Body stream.

Example

use hyper::{Body, Request};

// create a `GET /` request
let get = Request::new(Body::empty());

pub fn channel() -> (Sender, Body)

Create a Body stream with an associated sender half.

Useful when wanting to stream chunks from another thread.

pub fn wrap_stream<S, O, E>(stream: S) -> Body
where S: Stream<Item = Result<O, E>> + Send + 'static, O: Into<Bytes> + 'static, E: Into<Box<dyn Error + Sync + Send>> + 'static,

Available on crate feature stream only.

Wrap a futures Stream in a box inside Body.

Example

let chunks: Vec<Result<_, std::io::Error>> = vec![
    Ok("hello"),
    Ok(" "),
    Ok("world"),
];

let stream = futures_util::stream::iter(chunks);

let body = Body::wrap_stream(stream);

Optional

This function requires enabling the stream feature in your Cargo.toml.