twitch_api::client

Struct TowerService

pub struct TowerService<S, ReqBody>(pub S, _);

Available on crate features tower and client only.

A wrapped tower service

Tuple Fields
0: S
Implementations

impl<S, ReqBody> TowerService<S, ReqBody>

pub fn new(s: S) -> Self
where S: Clone,

Create a new wrapped tower service

Notes

Make sure your service stack can be cloned, one easy way to ensure this is to use tower::buffer