twitch_api

Module eventsub

Available on crate feature eventsub only.

Holds serializable EventSub stuff

Use CreateEventSubSubscription to subscribe to an event according to the EventSub guide. Parse the response payload text with Event::parse_http or Event::parse_websocket

Examples

See examples/ for a more complete example of using eventsub.

Subscribe to a channel’s follow events:

use twitch_api::{
    eventsub::{channel::ChannelFollowV2, Transport, TransportMethod},
    helix,
};

let event = ChannelFollowV2::new("1234", "5678");
let transport = Transport::webhook(
    "https://example.org/eventsub/channelfollow",
    String::from("secretabcd"),
);

let event_information = client
    .create_eventsub_subscription(event, transport, &token)
    .await?;

println!("event id: {:?}", event_information.id);

You’ll now get a http POST request to the url you specified as the callback. You need to respond to this request from your webserver with a 200 OK response with the challenge as the body. After this, you’ll get notifications

use twitch_api::eventsub::{Event, Payload, Message};
pub fn parse_request(
    request: &http::Request<Vec<u8>>,
) -> Result<http::Response<Vec<u8>>, Box<dyn std::error::Error + Send + Sync + 'static>> {
    // First, we verify the response, assuring it's legit.
    if !Event::verify_payload(request, b"secretabcd") {
        return Err(todo!());
    }
    match Event::parse_http(request)? {
        Event::ChannelFollowV2(Payload {
            message: Message::VerificationRequest(ver),
            ..
        }) => {
            // We've verified the request, so we can respond to it with the challenge
            Ok(http::Response::builder()
                .status(200)
                .body(ver.challenge.into_bytes())?)
        },
        Event::ChannelFollowV2(Payload {
            message: Message::Notification(notif),
            ..
        }) => {
            // make sure you save the `Twitch-Eventsub-Message-Id` headers value,
            // twitch may resend notifications, and in those cases you should just return 200 OK.

            // Do whatever you need to do with the event. Preferably send the event to a channel.
            println!("user {:?} followed {:?}", notif.user_name, notif.broadcaster_user_name);
            Ok(http::Response::builder().status(200).body(vec![])?)
        }
        _ => Ok(http::Response::builder().status(200).body(vec![])?),
    }
}

Implemented Subscriptions

automod.* 🟢 6/6

Name	Subscription
Payload

automod.message.hold (v1)	AutomodMessageHoldV1
AutomodMessageHoldV1Payload

automod.message.hold (v2)	AutomodMessageHoldV2
AutomodMessageHoldV2Payload

automod.message.update (v1)	AutomodMessageUpdateV1
AutomodMessageUpdateV1Payload

automod.message.update (v2)	AutomodMessageUpdateV2
AutomodMessageUpdateV2Payload

automod.settings.update (v1)	AutomodSettingsUpdateV1
AutomodSettingsUpdateV1Payload

automod.terms.update (v1)	AutomodTermsUpdateV1
AutomodTermsUpdateV1Payload

channel.* 🟡 66/70

Name	Subscription
Payload

channel.ad_break.begin (v1)	ChannelAdBreakBeginV1
ChannelAdBreakBeginV1Payload

channel.ban (v1)	ChannelBanV1
ChannelBanV1Payload

channel.bits.use (v1)	ChannelBitsUseV1
ChannelBitsUseV1Payload

channel.channel_points_automatic_reward_redemption.add (v1)	ChannelPointsAutomaticRewardRedemptionAddV1
ChannelPointsAutomaticRewardRedemptionAddV1Payload

channel.channel_points_automatic_reward_redemption.add (v2)	-
-

channel.channel_points_custom_reward.add (v1)	ChannelPointsCustomRewardAddV1
ChannelPointsCustomRewardAddV1Payload

channel.channel_points_custom_reward.remove (v1)	ChannelPointsCustomRewardRemoveV1
ChannelPointsCustomRewardRemoveV1Payload

channel.channel_points_custom_reward.update (v1)	ChannelPointsCustomRewardUpdateV1
ChannelPointsCustomRewardUpdateV1Payload

channel.channel_points_custom_reward_redemption.add (v1)	ChannelPointsCustomRewardRedemptionAddV1
ChannelPointsCustomRewardRedemptionAddV1Payload

channel.channel_points_custom_reward_redemption.update (v1)	ChannelPointsCustomRewardRedemptionUpdateV1
ChannelPointsCustomRewardRedemptionUpdateV1Payload

channel.charity_campaign.donate (v1)	ChannelCharityCampaignDonateV1
ChannelCharityCampaignDonateV1Payload

channel.charity_campaign.progress (v1)	ChannelCharityCampaignProgressV1
ChannelCharityCampaignProgressV1Payload

channel.charity_campaign.start (v1)	ChannelCharityCampaignStartV1
ChannelCharityCampaignStartV1Payload

channel.charity_campaign.stop (v1)	ChannelCharityCampaignStopV1
ChannelCharityCampaignStopV1Payload

channel.chat.clear (v1)	ChannelChatClearV1
ChannelChatClearV1Payload

channel.chat.clear_user_messages (v1)	ChannelChatClearUserMessagesV1
ChannelChatClearUserMessagesV1Payload

channel.chat.message (v1)	ChannelChatMessageV1
ChannelChatMessageV1Payload

channel.chat.message_delete (v1)	ChannelChatMessageDeleteV1
ChannelChatMessageDeleteV1Payload

channel.chat.notification (v1)	ChannelChatNotificationV1
ChannelChatNotificationV1Payload

channel.chat.user_message_hold (v1)	ChannelChatUserMessageHoldV1
ChannelChatUserMessageHoldV1Payload

channel.chat.user_message_update (v1)	ChannelChatUserMessageUpdateV1
ChannelChatUserMessageUpdateV1Payload

channel.chat_settings.update (v1)	ChannelChatSettingsUpdateV1
ChannelChatSettingsUpdateV1Payload

channel.cheer (v1)	ChannelCheerV1
ChannelCheerV1Payload

channel.follow (v2)	ChannelFollowV2
ChannelFollowV2Payload

channel.goal.begin (v1)	ChannelGoalBeginV1
ChannelGoalBeginV1Payload

channel.goal.end (v1)	ChannelGoalEndV1
ChannelGoalEndV1Payload

channel.goal.progress (v1)	ChannelGoalProgressV1
ChannelGoalProgressV1Payload

channel.guest_star_guest.update (beta)	ChannelGuestStarGuestUpdateBeta
ChannelGuestStarGuestUpdateBetaPayload

channel.guest_star_session.begin (beta)	ChannelGuestStarSessionBeginBeta
ChannelGuestStarSessionBeginBetaPayload

channel.guest_star_session.end (beta)	ChannelGuestStarSessionEndBeta
ChannelGuestStarSessionEndBetaPayload

channel.guest_star_settings.update (beta)	ChannelGuestStarSettingsUpdateBeta
ChannelGuestStarSettingsUpdateBetaPayload

channel.hype_train.begin (v1)	ChannelHypeTrainBeginV1
ChannelHypeTrainBeginV1Payload

channel.hype_train.begin (v2)	-
-

channel.hype_train.end (v1)	ChannelHypeTrainEndV1
ChannelHypeTrainEndV1Payload

channel.hype_train.end (v2)	-
-

channel.hype_train.progress (v1)	ChannelHypeTrainProgressV1
ChannelHypeTrainProgressV1Payload

channel.hype_train.progress (v2)	-
-

channel.moderate (v1)	ChannelModerateV1
ChannelModerateV1Payload

channel.moderate (v2)	ChannelModerateV2
ChannelModerateV2Payload

channel.moderator.add (v1)	ChannelModeratorAddV1
ChannelModeratorAddV1Payload

channel.moderator.remove (v1)	ChannelModeratorRemoveV1
ChannelModeratorRemoveV1Payload

channel.poll.begin (v1)	ChannelPollBeginV1
ChannelPollBeginV1Payload

channel.poll.end (v1)	ChannelPollEndV1
ChannelPollEndV1Payload

channel.poll.progress (v1)	ChannelPollProgressV1
ChannelPollProgressV1Payload

channel.prediction.begin (v1)	ChannelPredictionBeginV1
ChannelPredictionBeginV1Payload

channel.prediction.end (v1)	ChannelPredictionEndV1
ChannelPredictionEndV1Payload

channel.prediction.lock (v1)	ChannelPredictionLockV1
ChannelPredictionLockV1Payload

channel.prediction.progress (v1)	ChannelPredictionProgressV1
ChannelPredictionProgressV1Payload

channel.raid (v1)	ChannelRaidV1
ChannelRaidV1Payload

channel.shared_chat.begin (v1)	ChannelSharedChatBeginV1
ChannelSharedChatBeginV1Payload

channel.shared_chat.end (v1)	ChannelSharedChatEndV1
ChannelSharedChatEndV1Payload

channel.shared_chat.update (v1)	ChannelSharedChatUpdateV1
ChannelSharedChatUpdateV1Payload

channel.shield_mode.begin (v1)	ChannelShieldModeBeginV1
ChannelShieldModeBeginV1Payload

channel.shield_mode.end (v1)	ChannelShieldModeEndV1
ChannelShieldModeEndV1Payload

channel.shoutout.create (v1)	ChannelShoutoutCreateV1
ChannelShoutoutCreateV1Payload

channel.shoutout.receive (v1)	ChannelShoutoutReceiveV1
ChannelShoutoutReceiveV1Payload

channel.subscribe (v1)	ChannelSubscribeV1
ChannelSubscribeV1Payload

channel.subscription.end (v1)	ChannelSubscriptionEndV1
ChannelSubscriptionEndV1Payload

channel.subscription.gift (v1)	ChannelSubscriptionGiftV1
ChannelSubscriptionGiftV1Payload

channel.subscription.message (v1)	ChannelSubscriptionMessageV1
ChannelSubscriptionMessageV1Payload

channel.suspicious_user.message (v1)	ChannelSuspiciousUserMessageV1
ChannelSuspiciousUserMessageV1Payload

channel.suspicious_user.update (v1)	ChannelSuspiciousUserUpdateV1
ChannelSuspiciousUserUpdateV1Payload

channel.unban (v1)	ChannelUnbanV1
ChannelUnbanV1Payload

channel.unban_request.create (v1)	ChannelUnbanRequestCreateV1
ChannelUnbanRequestCreateV1Payload

channel.unban_request.resolve (v1)	ChannelUnbanRequestResolveV1
ChannelUnbanRequestResolveV1Payload

channel.update (v2)	ChannelUpdateV2
ChannelUpdateV2Payload

channel.vip.add (v1)	ChannelVipAddV1
ChannelVipAddV1Payload

channel.vip.remove (v1)	ChannelVipRemoveV1
ChannelVipRemoveV1Payload

channel.warning.acknowledge (v1)	ChannelWarningAcknowledgeV1
ChannelWarningAcknowledgeV1Payload

channel.warning.send (v1)	ChannelWarningSendV1
ChannelWarningSendV1Payload

conduit.* 🟢 1/1

Name	Subscription
Payload

conduit.shard.disabled (v1)	ConduitShardDisabledV1
ConduitShardDisabledV1Payload

drop.* 🔴 0/1

Name	Subscription
Payload

drop.entitlement.grant (v1)	-
-

extension.* 🔴 0/1

Name	Subscription
Payload

extension.bits_transaction.create (v1)	-
-

stream.* 🟢 2/2

Name	Subscription
Payload

stream.offline (v1)	StreamOfflineV1
StreamOfflineV1Payload

stream.online (v1)	StreamOnlineV1
StreamOnlineV1Payload

user.* 🟢 4/4

Name	Subscription
Payload

user.authorization.grant (v1)	UserAuthorizationGrantV1
UserAuthorizationGrantV1Payload

user.authorization.revoke (v1)	UserAuthorizationRevokeV1
UserAuthorizationRevokeV1Payload

user.update (v1)	UserUpdateV1
UserUpdateV1Payload

user.whisper.message (v1)	UserWhisperMessageV1
UserWhisperMessageV1Payload

Re-exports

pub use event::websocket::*;

Modules

automod

Automod related events

channel

Subscription types regarding channels

conduit

Subscription types regarding conduits.

event

EventSub events and their types

stream

Subscription types regarding streams

user

Subscription types regarding users

Structs

Conduit

General information about a Conduit

ConduitTransport

Conduit transport

ConduitTransportResponse

Conduit transport

EventSubSubscription

General information about an EventSub subscription.

EventSubscriptionInformation

Metadata about the subscription.

Payload

Notification received

Shard

General information about a Shard

ShardError

A structured error that occurred with a shard

ShardResponse

A shard when described by Twitch

VerificationRequest

Verification Request

WebhookTransport

Webhook transport

WebhookTransportResponse

Webhook transport

WebsocketTransport

Websocket transport

WebsocketTransportResponse

Websocket transport

Enums

Event

A notification with an event payload. Enumerates all possible Payloads

EventType

Event Types

Message

Subscription message/payload. Received on events and other messages.

PayloadParseError

Errors that can happen when parsing payload

ShardStatus

The shard status.

Status

Subscription request status

Transport

Transport setting for event notification

TransportMethod

Transport method

TransportResponse

Transport response on event notification

Traits

EventSubscription

An EventSub subscription.