twitch_api::eventsub

Module automod

Available on crate feature eventsub only.

Automod related events

Modules

message

Events for Automod actions on messages

settings

Events for Automod settings

terms

Events for Automod terms (allowed/denied words)

Structs

AutomodMessageHoldV1

automod.message.hold: a message was caught by automod for review.

AutomodMessageHoldV2

automod.message.hold: a message was caught by automod for review.

AutomodMessageHoldV1Payload

automod.message.hold response payload.

AutomodMessageHoldV2Payload

automod.message.hold response payload.

AutomodMessageUpdateV1

automod.message.update: a message in the automod queue had its status changed.

AutomodMessageUpdateV2

automod.message.update: a message in the automod queue had its status changed.

AutomodMessageUpdateV1Payload

automod.message.update response payload.

AutomodMessageUpdateV2Payload

automod.message.hold response payload.

AutomodSettingsUpdateV1

automod.settings.update: a notification is sent when a broadcaster’s automod settings are updated.

AutomodSettingsUpdateV1Payload

automod.settings.update response payload.

AutomodTermsUpdateV1

automod.terms.update: A notification is sent when a broadcaster’s automod terms are updated. Changes to private terms are not sent.

AutomodTermsUpdateV1Payload

automod.terms.update response payload.

Enums

AutomodCategory

A category identified by automod for a message.