twitch_api::eventsub::automod
Module message

Available on crate feature eventsub only.

Events for Automod actions on messages

Modules

hold

A message was caught by automod for review

update

a message in the automod queue had its status changed

Structs

AutomodBlockedTerm

A blocked term that was found in a message

AutomodBlockedTermInfo

Information on which blocked terms were matched in a message

AutomodBlockedTermsInfo

Information about the blocked terms that caused a message to be caught by automod.

AutomodMessage

A message sent in automod events

AutomodMessageBoundary

The bounds of the text that caused the message to be caught.

AutomodMessageEmote

An emote in a message filtered by automod

AutomodMessageHoldBetabeta

automod.message.hold: a message was caught by automod for review.

AutomodMessageHoldBetaPayloadbeta

automod.message.hold response payload.

AutomodMessageHoldV1

automod.message.hold: a message was caught by automod for review.

AutomodMessageHoldV1Payload

automod.message.hold response payload.

AutomodMessageInfo

Information on why a message was caught by automod

AutomodMessageUpdateBetabeta

automod.message.update: a message in the automod queue had its status changed.

AutomodMessageUpdateBetaPayloadbeta

automod.message.hold response payload.

AutomodMessageUpdateV1

automod.message.update: a message in the automod queue had its status changed.

AutomodMessageUpdateV1Payload

automod.message.update response payload.

Enums

AutomodHeldReason

The reason why a message was held

AutomodMessageFragment

A chat message fragment

AutomodMessageStatus

A message’s Automod status

Type Aliases

AutomodMessageCheermote

A cheermote in a message filtered by automod