twitch_api::eventsub::automod::message

Type Alias AutomodMessageCheermote

pub type AutomodMessageCheermote = Cheermote;

Available on crate feature eventsub only.

A cheermote in a message filtered by automod

Aliased Type

struct AutomodMessageCheermote {
    pub prefix: String,
    pub bits: i32,
    pub tier: i32,
}

Fields
prefix: String

The name portion of the Cheermote string that you use in chat to cheer Bits.

The full Cheermote string is the concatenation of {prefix} + {number of Bits}. For example, if the prefix is “Cheer” and you want to cheer 100 Bits, the full Cheermote string is Cheer100. When the Cheermote string is entered in chat, Twitch converts it to the image associated with the Bits tier that was cheered.

bits: i32

The amount of bits cheered.

tier: i32

The tier level of the cheermote.