twitch_api::eventsub::channel

Module channel_points_automatic_reward_redemption

Available on crate feature eventsub only.

A viewer has redeemed an automatic channel points reward in a specified channel.

Modules

add

A viewer has redeemed an automatic channel points reward on the specified channel.

Structs

AutomaticReward

Basic information about the automatic reward that was redeemed, at the time it was redeemed.

ChannelPointsAutomaticRewardRedemptionAddV1

channel.channel_points_automatic_reward_redemption.add:a viewer has redeemed an automatic channel points reward on the specified channel.

ChannelPointsAutomaticRewardRedemptionAddV1Payload

channel.channel_points_automatic_reward_redemption.add response payload.

RedemptionMessage

The user message and emote information from a reward redemption

UnlockedEmote

An emote that was unlocked as part of a reward

Enums

AutomaticRewardType

The type of reward