Module twitch_api::eventsub::channel::charity_campaign

Available on crate feature eventsub only.

Poll on a specific channel has been begun, ended or progressed.

Modules

donate

Sends an event notification when a user donates to the broadcaster’s charity campaign.

progress

Sends an event notification when progress is made towards the campaign’s goal or when the broadcaster changes the fundraising goal.

start

Sends an event notification when a broadcaster starts a charity campaign.

stop

Sends an event notification when a broadcaster stops a charity campaign.

Structs

ChannelCharityCampaignDonateV1

channel.charity_campaign.donate: a user donates to the broadcaster’s charity campaign.

ChannelCharityCampaignDonateV1Payload

channel.charity_campaign.donate response payload.

ChannelCharityCampaignProgressV1

channel.charity_campaign.progress: progress is made towards the campaign’s goal or when the broadcaster changes the fundraising goal.

ChannelCharityCampaignProgressV1Payload

channel.charity_campaign.progress response payload.

ChannelCharityCampaignStartV1

channel.charity_campaign.start: a broadcaster starts a charity campaign.

ChannelCharityCampaignStartV1Payload

channel.charity_campaign.start response payload.

ChannelCharityCampaignStopV1

channel.charity_campaign.stop: a broadcaster stops a charity campaign.

ChannelCharityCampaignStopV1Payload

channel.charity_campaign.stop response payload.