twitch_api::eventsub::channel

Module guest_star_guest

Available on crate features eventsub and beta only.

Events regarding guests of guest star sessions

Modules

update

the host preferences for Guest Star have been updated.

Structs

ChannelGuestStarGuestUpdateBeta

channel.guest_star_guest.update: the host preferences for Guest Star have been updated.

ChannelGuestStarGuestUpdateBetaPayload

channel.guest_star_guest.update response payload.

Enums

GuestState

The current state of a user in a guest star session