twitch_api::eventsub::channel
Module moderate

Available on crate feature eventsub only.

a moderator performs a moderation action in a channel.

Structs

AutomodTerms

Metadata associated with the automod terms changes.

Ban

Metadata associated with the ban command.

ChannelModerateV1

channel.moderate: a moderator performs a moderation action in a channel.

ChannelModerateV2

channel.moderate: a moderator performs a moderation action in a channel.

ChannelModerateV1Payload

channel.moderate response payload.

ChannelModerateV2Payload

channel.moderate response payload.

Delete

Metadata associated with the delete command.

Followers

Metadata associated with the followers command.

Mod

Metadata associated with the mod command.

Raid

Metadata associated with the raid command.

SharedChatBan

Information about the shared_chat_ban event.

SharedChatDelete

Information about the shared_chat_delete event.

SharedChatTimeout

Information about the shared_chat_timeout event.

SharedChatUnban

Information about the shared_chat_unban event.

SharedChatUntimeout

Information about the shared_chat_untimeout event.

Slow

Metadata associated with the slow command.

Timeout

Metadata associated with the timeout command.

Unban

Metadata associated with the unban command.

UnbanRequest

Metadata associated with an unban request.

Unmod

Metadata associated with the unmod command.

Unraid

Metadata associated with the unraid command.

Untimeout

Metadata associated with the untimeout command.

Unvip

Metadata associated with the unvip command.

Vip

Metadata associated with the vip command.

Warn

Metadata associated with the warn command.

Enums

ActionV1

All possible actions in ChannelModerateV1Payload

ActionV2

All possible actions in ChannelModerateV1Payload

AutomodTermAction

An action on an Automod term

AutomodTermList

A list with Automod terms