Module twitch_api::eventsub::channel::poll

Available on crate feature eventsub only.

Poll on a specific channel has been begun, ended or progressed.

Modules

begin

A poll begins on the specified channel.

end

A poll ends on the specified channel.

progress

A user responds to a poll on the specified channel

Structs

BitsVoting

Bits voting settings for a poll

ChannelPointsVoting

Channel Points voting settings

ChannelPollBeginV1

channel.poll.begin: a poll begins on the specified channel.

ChannelPollBeginV1Payload

channel.poll.begin response payload.

ChannelPollEndV1

channel.hype_train.end: a poll ends on the specified channel.

ChannelPollEndV1Payload

channel.hype_train.end response payload.

ChannelPollProgressV1

channel.hype_train.progress: an user responds to a poll on the specified channel

ChannelPollProgressV1Payload

channel.hype_train.progress response payload.