twitch_api::eventsub::channel
Module suspicious_user

Available on crate feature eventsub only.

A user’s moderator privileges on a specified channel are changed.

Modules

message

a chat message has been sent from a suspicious user.

update

a suspicious user has been updated.

Structs

ChannelSuspiciousUserMessageV1

channel.suspicious_user.message: a chat message has been sent from a suspicious user.

ChannelSuspiciousUserMessageV1Payload

channel.suspicious_user.message response payload.

ChannelSuspiciousUserUpdateV1

channel.suspicious_user.update: a suspicious user has been updated.

ChannelSuspiciousUserUpdateV1Payload

channel.suspicious_user.update response payload.

Enums

BanEvasionEvaluation

Likelyhood that a user is a ban evader

LowTrustStatus

A user’s low trust status

SuspiciousUserType

Traits of a suspicious user