Module twitch_api::eventsub::channel::update

Available on crate feature eventsub only.

Channel has updated the category, title, mature flag, or broadcast language.

Structs

ChannelUpdateV1Deprecated

version 1 of channel.update subscription type sends notifications when a broadcaster updates the category, title, mature flag, or broadcast language for their channel.

ChannelUpdateV1PayloadDeprecated

channel.update response payload.

ChannelUpdateV2

channel.update subscription type sends notifications when a broadcaster updates the category, title, mature flag, or broadcast language for their channel.

ChannelUpdateV2Payload

channel.update response payload.