Struct twitch_api::eventsub::Payload

#[non_exhaustive]
pub struct Payload<E: EventSubscription + Clone> {
    pub subscription: EventSubscriptionInformation<E>,
    pub message: Message<E>,
}

Available on crate feature eventsub only.

Notification received

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

subscription: EventSubscriptionInformation<E>

Subscription information.

message: Message<E>

Event information.

Implementations

impl<E: EventSubscription> Payload<E>

pub fn parse_notification(source: &str) -> Result<Payload<E>, PayloadParseError>

Parse string slice as a Payload, this will assume your string is from an eventsub message with type notification

The string should be a

Examples

use twitch_api::eventsub::{channel::ChannelFollowV2, Payload};
let notification = r#"
{
    "subscription": {
        "id": "f1c2a387-161a-49f9-a165-0f21d7a4e1c4",
        "type": "channel.follow",
        "version": "2",
        "status": "enabled",
        "cost": 0,
        "condition": {
           "broadcaster_user_id": "1337",
           "moderator_user_id": "1337"
        },
         "transport": {
            "method": "webhook",
            "callback": "https://example.com/webhooks/callback"
        },
        "created_at": "2019-11-16T10:11:12.634234626Z"
    },
    "event": {
        "user_id": "1234",
        "user_login": "cool_user",
        "user_name": "Cool_User",
        "broadcaster_user_id": "1337",
        "broadcaster_user_login": "cooler_user",
        "broadcaster_user_name": "Cooler_User",
        "followed_at": "2020-07-15T18:16:11.17106713Z"
    }
}
"#;
let payload: Payload<ChannelFollowV2> =
    Payload::parse_notification(notification)?;

pub fn parse_revocation(source: &str) -> Result<Payload<E>, PayloadParseError>

Parse string slice as a Payload with a message of Message::Revocation.

Examples

use twitch_api::eventsub::{channel::ChannelFollowV2, Payload};
let notification = r#"
{
    "subscription": {
        "id": "f1c2a387-161a-49f9-a165-0f21d7a4e1c4",
        "status": "authorization_revoked",
        "type": "channel.follow",
        "cost": 0,
        "version": "2",
        "condition": {
            "broadcaster_user_id": "1337",
            "moderator_user_id": "1337"
        },
        "transport": {
            "method": "webhook",
            "callback": "https://example.com/webhooks/callback"
        },
        "created_at": "2019-11-16T10:11:12.123Z"
    }
}
"#;
let payload: Payload<ChannelFollowV2> =
    Payload::parse_revocation(notification)?;

pub fn parse_verification_request( source: &str ) -> Result<Payload<E>, PayloadParseError>

Parse string slice as a Payload with a message of Message::VerificationRequest.

Examples

use twitch_api::eventsub::{channel::ChannelFollowV2, Payload};
let notification = r#"
{
    "challenge": "pogchamp-kappa-360noscope-vohiyo",
    "subscription": {
        "id": "f1c2a387-161a-49f9-a165-0f21d7a4e1c4",
        "status": "webhook_callback_verification_pending",
        "type": "channel.follow",
        "version": "2",
        "cost": 1,
        "condition": {
            "broadcaster_user_id": "12826",
            "moderator_user_id": "12826"
        },
        "transport": {
            "method": "webhook",
            "callback": "https://example.com/webhooks/callback"
        },
        "created_at": "2019-11-16T10:11:12.123Z"
    }
}
"#;
let payload: Payload<ChannelFollowV2> =
    Payload::parse_verification_request(notification)?;

pub fn parse_http<B>( request: &Request<B> ) -> Result<Payload<E>, PayloadParseError>where B: AsRef<[u8]>,

Parse http post request as a Payload with a specific event.

If you don’t know what event this payload is, use Event::parse_http instead.

If your Request<B> is of another type that doesn’t implement AsRef<[u8]>, try converting it with Request::map

use http::Request;
use twitch_api::eventsub::{Payload, channel::ChannelFollowV2};
// Example of a request with a body that doesn't implement `AsRef<[u8]>`
let original_request: Request<Body> = http::Request::new(Body::new());
// Convert to a request with a body of `Vec<u8>`, which does implement `AsRef<[u8]>`
let converted_request: Request<Vec<u8>> = original_request.map(|r| r.to_bytes().to_owned());
Payload::<ChannelFollowV2>::parse_http(&converted_request)?

impl<E: EventSubscription + Clone> Payload<E>

pub fn get_event_type(&self) -> EventType

Convenience method for getting the event type from the payload.

pub fn get_event_version(&self) -> &'static str

Convenience method for getting the event version from the payload.