twitch_api::eventsub::user

Module authorization

Available on crate feature eventsub only.

Authorization from a user has been granted or revoked to a specific client ID

Modules

grant

A user’s authorization has been granted to your client id.

revoke

A user has revoked authorization for your client id.

Structs

UserAuthorizationGrantV1

user.authorization.grant: a user’s authorization has been granted to your client id.

UserAuthorizationGrantV1Payload

user.authorization.grant response payload.

UserAuthorizationRevokeV1

user.authorization.revoke: a user has revoked authorization for your client id. Use this webhook to meet government requirements for handling user data, such as GDPR, LGPD, or CCPA.

UserAuthorizationRevokeV1Payload

user.authorization.revoke response payload.