Function twitch_api::parse_json

pub fn parse_json<'a, T: Deserialize<'a>>(
    s: &'a str,
    log_ignored: bool
) -> Result<T, DeserError>

Available on crate feature serde_json only.

Parse a string as T, logging ignored fields and giving a more detailed error message on parse errors

The log_ignored argument decides if a trace of ignored value should be emitted

Examples found in repository?

src/eventsub/mod.rs (line 187)

≺ ≻↕

174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232

    pub fn parse_notification(source: &str) -> Result<Payload<E>, PayloadParseError> {
        #[derive(Deserialize)]
        #[cfg_attr(feature = "deny_unknown_fields", serde(deny_unknown_fields))]
        struct Notification<E: EventSubscription> {
            #[serde(bound = "E: EventSubscription")]
            pub subscription: EventSubscriptionInformation<E>,
            #[serde(bound = "E: EventSubscription")]
            pub event: <E as EventSubscription>::Payload,
        }

        let Notification {
            subscription,
            event,
        } = parse_json::<Notification<E>>(source, true)?;

        Ok(Payload {
            subscription,
            message: Message::Notification(event),
        })
    }

    /// Parse string slice as a [`Payload`] with a message of [`Message::Revocation`].
    pub fn parse_revocation(source: &str) -> Result<Payload<E>, PayloadParseError> {
        #[derive(Deserialize)]
        #[cfg_attr(feature = "deny_unknown_fields", serde(deny_unknown_fields))]
        struct Notification<E: EventSubscription> {
            #[serde(bound = "E: EventSubscription")]
            pub subscription: EventSubscriptionInformation<E>,
        }

        let Notification { subscription } = parse_json::<Notification<E>>(source, true)?;

        Ok(Payload {
            subscription,
            message: Message::Revocation(),
        })
    }

    /// Parse string slice as a [`Payload`] with a message of [`Message::VerificationRequest`].
    pub fn parse_verification_request(source: &str) -> Result<Payload<E>, PayloadParseError> {
        #[derive(Deserialize)]
        #[cfg_attr(feature = "deny_unknown_fields", serde(deny_unknown_fields))]
        struct Notification<E: EventSubscription> {
            #[serde(bound = "E: EventSubscription")]
            pub subscription: EventSubscriptionInformation<E>,
            #[serde(bound = "E: EventSubscription")]
            pub challenge: String,
        }

        let Notification {
            subscription,
            challenge,
        } = parse_json::<Notification<E>>(source, true)?;

        Ok(Payload {
            subscription,
            message: Message::VerificationRequest(VerificationRequest { challenge }),
        })
    }

Hide additional examples

src/helix/endpoints/schedule/update_channel_stream_schedule_segment.rs (line 158)

↕

148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173

    fn parse_inner_response(
        request: Option<Self>,
        uri: &http::Uri,
        response: &str,
        status: http::StatusCode,
    ) -> Result<helix::Response<Self, Self::Response>, helix::HelixRequestPatchError>
    where
        Self: Sized,
    {
        let response: helix::InnerResponse<<Self as Request>::Response> =
            helix::parse_json(response, true).map_err(|e| {
                helix::HelixRequestPatchError::DeserializeError(
                    response.to_string(),
                    e,
                    uri.clone(),
                    status,
                )
            })?;
        Ok(helix::Response {
            data: response.data,
            pagination: response.pagination.cursor,
            request,
            total: response.total,
            other: None,
        })
    }

src/helix/endpoints/channels/get_channel_information.rs (line 110)

↕

100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125

    fn parse_inner_response(
        request: Option<Self>,
        uri: &http::Uri,
        response: &str,
        status: http::StatusCode,
    ) -> Result<helix::Response<Self, Self::Response>, helix::HelixRequestGetError>
    where
        Self: Sized,
    {
        let response: helix::InnerResponse<Vec<ChannelInformation>> =
            helix::parse_json(response, true).map_err(|e| {
                helix::HelixRequestGetError::DeserializeError(
                    response.to_string(),
                    e,
                    uri.clone(),
                    status,
                )
            })?;
        Ok(helix::Response {
            data: response.data.into_iter().next(),
            pagination: response.pagination.cursor,
            request,
            total: None,
            other: None,
        })
    }

src/helix/endpoints/search/search_categories.rs (line 109)

↕

99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124

    fn parse_inner_response(
        request: Option<Self>,
        uri: &http::Uri,
        response: &str,
        status: http::StatusCode,
    ) -> Result<helix::Response<Self, Self::Response>, helix::HelixRequestGetError>
    where
        Self: Sized,
    {
        let response: helix::InnerResponse<Option<Self::Response>> =
            helix::parse_json(response, true).map_err(|e| {
                helix::HelixRequestGetError::DeserializeError(
                    response.to_string(),
                    e,
                    uri.clone(),
                    status,
                )
            })?;
        Ok(helix::Response {
            data: response.data.unwrap_or_default(),
            pagination: response.pagination.cursor,
            request,
            total: response.total,
            other: None,
        })
    }

src/tmi/mod.rs (line 94)

↕

71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95

    pub async fn get_chatters(
        &'a self,
        broadcaster: &types::UserNameRef,
    ) -> Result<GetChatters, RequestError<<C as crate::HttpClient<'a>>::Error>> {
        let url = format!(
            "{}{}{}{}",
            crate::TWITCH_TMI_URL.as_str(),
            "group/user/",
            broadcaster.as_str().replace('#', "").to_ascii_lowercase(),
            "/chatters"
        );
        let req = http::Request::builder()
            .uri(url)
            .body(Vec::with_capacity(0).into())?;
        let resp = self
            .client
            .req(req)
            .await
            .map_err(|e| RequestError::RequestError(Box::new(e)))?;
        let (parts, mut body) = resp.into_parts();
        let resp = http::Response::from_parts(parts, hyper::body::to_bytes(&mut body).await?);
        let text = std::str::from_utf8(resp.body())
            .map_err(|e| RequestError::Utf8Error(resp.body().to_vec(), e))?;
        crate::parse_json(text, true).map_err(Into::into)
    }

src/helix/endpoints/users/get_users_follows.rs (line 171)

↕

154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189

    fn parse_inner_response(
        request: Option<Self>,
        uri: &http::Uri,
        response: &str,
        status: http::StatusCode,
    ) -> Result<helix::Response<Self, Self::Response>, helix::HelixRequestGetError>
    where
        Self: Sized,
    {
        #[derive(PartialEq, Deserialize, Debug, Clone)]
        struct InnerResponse {
            data: Vec<FollowRelationship>,
            total: i64,
            #[serde(default)]
            pagination: helix::Pagination,
        }

        let response: InnerResponse = helix::parse_json(response, true).map_err(|e| {
            helix::HelixRequestGetError::DeserializeError(
                response.to_string(),
                e,
                uri.clone(),
                status,
            )
        })?;
        Ok(helix::Response {
            data: UsersFollows {
                total: response.total,
                follow_relationships: response.data,
            },
            pagination: response.pagination.cursor,
            request,
            total: Some(response.total),
            other: None,
        })
    }

Additional examples can be found in:

src/helix/endpoints/raids/start_a_raid.rs

src/helix/endpoints/polls/create_poll.rs

src/helix/endpoints/predictions/create_prediction.rs

src/helix/endpoints/subscriptions/check_user_subscription.rs

src/helix/endpoints/points/create_custom_rewards.rs

src/helix/endpoints/moderation/ban_user.rs

src/helix/request.rs

src/helix/endpoints/bits/get_bits_leaderboard.rs

src/helix/endpoints/chat/get_chat_settings.rs

src/eventsub/event.rs

src/helix/endpoints/eventsub/get_eventsub_subscriptions.rs

src/helix/endpoints/chat/update_chat_settings.rs

src/helix/endpoints/points/update_custom_reward.rs

src/helix/endpoints/points/update_redemption_status.rs

src/helix/endpoints/polls/end_poll.rs

src/helix/endpoints/predictions/end_prediction.rs

src/helix/client/custom.rs

src/helix/endpoints/eventsub/create_eventsub_subscription.rs

src/pubsub/mod.rs