twitch_api::helix::channels
Module get_ad_schedule

Available on crate feature helix only.

Gets ad schedule related information, including snooze, when the last ad was run, when the next ad is scheduled, and if the channel is currently in pre-roll free time. get-ad-schedule

Accessing the endpoint

Request: GetAdScheduleRequest

To use this endpoint, construct a GetAdScheduleRequest with the GetAdScheduleRequest::broadcaster_id() method.

use twitch_api::helix::channels::get_ad_schedule;
let request = get_ad_schedule::GetAdScheduleRequest::broadcaster_id("1234");

Response: AdSchedule

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, channels::get_ad_schedule};
let request = get_ad_schedule::GetAdScheduleRequest::broadcaster_id("1234");
let response: Option<get_ad_schedule::AdSchedule> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetAdScheduleRequest::parse_response(None, &request.get_uri(), response)

Structs

AdSchedule

Return Values for Get Ad Schedule

GetAdScheduleRequest

Query Parameters for Get Ad Schedule