twitch_api::helix::channels
Module get_channel_editors

Available on crate feature helix only.

Gets a list of users who have editor permissions for a specific channel. get-channel-editors

Accessing the endpoint

Request: GetChannelEditorsRequest

To use this endpoint, construct a GetChannelEditorsRequest with the GetChannelEditorsRequest::broadcaster_id() or GetChannelEditorsRequest::builder() method.

use twitch_api::helix::channels::get_channel_editors;
let request =
    get_channel_editors::GetChannelEditorsRequest::broadcaster_id("1234");

Response: Editor

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, channels::get_channel_editors};
let request = get_channel_editors::GetChannelEditorsRequest::broadcaster_id("1234");
let response: Vec<get_channel_editors::Editor> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetChannelEditorsRequest::parse_response(None, &request.get_uri(), response)

Structs

Editor

Return Values for Get Channel Editors

GetChannelEditorsRequest

Query Parameters for Get Channel Editors