Module twitch_api::helix::channels::get_followed_channels

Available on crate feature helix only.

Gets a list of broadcasters that the specified user follows. You can also use this endpoint to see whether a user follows a specific broadcaster. get-followed-channels

Accessing the endpoint

Request: GetFollowedChannels

To use this endpoint, construct a GetFollowedChannels with the GetFollowedChannels::user_id() or GetFollowedChannels::builder() method.

use twitch_api::helix::channels::get_followed_channels;
let request = get_followed_channels::GetFollowedChannels::user_id("1234");

Response: FollowedBroadcaster

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, channels::get_followed_channels};
let request = get_followed_channels::GetFollowedChannels::user_id("1234");
let response: Vec<get_followed_channels::FollowedBroadcaster> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetFollowedChannels::parse_response(None, &request.get_uri(), response)

Structs

FollowedBroadcaster

Return Values for Get Followed Channels

GetFollowedChannels

Query Parameters for Get Followed Channels