twitch_api::helix::channels
Module snooze_next_ad

Available on crate feature helix only.

If available, pushes back the timestamp of the upcoming automatic mid-roll ad by 5 minutes. This endpoint duplicates the snooze functionality in the creator dashboard’s Ads Manager. snooze-next-ad

Accessing the endpoint

Request: SnoozeNextAdRequest

To use this endpoint, construct a SnoozeNextAdRequest with the SnoozeNextAdRequest::broadcaster_id() method.

use twitch_api::helix::channels::snooze_next_ad;
let request = snooze_next_ad::SnoozeNextAdRequest::broadcaster_id("1234");

Response: SnoozedAdSchedule

Send the request to receive the response with HelixClient::req_patch().

use twitch_api::helix::{self, channels::snooze_next_ad};
let request = snooze_next_ad::SnoozeNextAdRequest::broadcaster_id("1234");
let response: snooze_next_ad::SnoozedAdSchedule = client.req_post(request, helix::EmptyBody, &token).await?.data;

You can also get the http::Request with request.create_request(body, &token, &client_id) and parse the http::Response with SnoozeNextAdRequest::parse_response(None, &request.get_uri(), response)

Structs

SnoozeNextAdRequest

Query Parameters for Snooze Next Ad

SnoozedAdSchedule

Return Values for Snooze Next Ad