twitch_api::helix

Module charity

Available on crate feature helix only.

Helix endpoints regarding charities

Implemented endpoints

Charity 🟢 2/2

Endpoint	Helper	Module

Get Charity Campaign	-	get_charity_campaign

Get Charity Campaign Donations	-	get_charity_campaign_donations

Modules

get_charity_campaign

Gets information about the charity campaign that a broadcaster is running, such as their fundraising goal and the amount that’s been donated so far. get-charity-campaign

get_charity_campaign_donations

Gets the list of donations that users have made to the broadcaster’s active charity campaign. get-charity-campaign-donations

Structs

CharityCampaign

Return Values for Get Charity Campaign

CharityCampaignDonation

Return Values for Get Charity Campaign Donations

GetCharityCampaignDonationsRequest

Query Parameters for Get Charity Campaign Donations

GetCharityCampaignRequest

Query Parameters for Get Charity Campaign