twitch_api::helix::charity
Module get_charity_campaign

Available on crate feature helix only.

Gets information about the charity campaign that a broadcaster is running, such as their fundraising goal and the amount that’s been donated so far. get-charity-campaign

Accessing the endpoint

Request: GetCharityCampaignRequest

To use this endpoint, construct a GetCharityCampaignRequest with the GetCharityCampaignRequest::broadcaster_id() method.

use twitch_api::helix::charity::get_charity_campaign;
let request =
    get_charity_campaign::GetCharityCampaignRequest::broadcaster_id(
        "123456",
    );

Response: CharityCampaign

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, charity::get_charity_campaign};
let request = get_charity_campaign::GetCharityCampaignRequest::broadcaster_id("123456");
let response: Option<get_charity_campaign::CharityCampaign> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetCharityCampaignRequest::parse_response(None, &request.get_uri(), response)

Structs

CharityCampaign

Return Values for Get Charity Campaign

GetCharityCampaignRequest

Query Parameters for Get Charity Campaign