Module twitch_api::helix::chat

Available on crate feature helix only.

Helix endpoints regarding chat

Modules

get_channel_chat_badges

Gets a list of custom chat badges that can be used in chat for the specified channel. This includes subscriber badges and Bit badges. get-channel-chat-badges

get_channel_emotes

Gets all custom emotes for a specific Twitch channel including subscriber emotes, Bits tier emotes, and follower emotes. get-channel-emotes

get_emote_sets

Gets all Twitch emotes for one or more specific emote sets. get-emote-sets

get_global_chat_badges

Gets a list of chat badges that can be used in chat for any channel. get-global-chat-badges

get_global_emotes

Gets all global emotes. Global emotes are Twitch-specific emoticons that every user can use in Twitch chat. get-global-emotes

update_chat_settings

Updates the broadcaster’s chat settings.

Structs

BadgeSet

A set of badges

ChannelEmote

A chat emote

ChatBadge

A chat Badge

ChatSettings

Chat settings

GetChannelChatBadgesRequest

Query Parameters for Get Channel Chat Badges

GetChannelEmotesRequest

Query Parameters for Get Channel Emotes

GetEmoteSetsRequest

Query Parameters for Get Channel Emotes

GetGlobalChatBadgesRequest

Query Parameters for Get Global Chat Badges

GetGlobalEmotesRequest

Query Parameters for Get Channel Emotes

GlobalEmote

A chat emote

UpdateChatSettingsBody

Body Parameters for Update Chat Settingss

UpdateChatSettingsRequest

Query Parameters for Update Chat Settingss