twitch_api::helix::chat
Module get_channel_chat_badges

Available on crate feature helix only.

Gets a list of custom chat badges that can be used in chat for the specified channel. This includes subscriber badges and Bit badges. get-channel-chat-badges

Accessing the endpoint

Request: GetChannelChatBadgesRequest

To use this endpoint, construct a GetChannelChatBadgesRequest with the GetChannelChatBadgesRequest::broadcaster_id() method.

use twitch_api::helix::chat::get_channel_chat_badges;
let request =
    get_channel_chat_badges::GetChannelChatBadgesRequest::broadcaster_id(
        "1234",
    );

Response: BadgeSet

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, chat::get_channel_chat_badges};
let request = get_channel_chat_badges::GetChannelChatBadgesRequest::broadcaster_id("1234");
let response: Vec<helix::chat::BadgeSet> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetChannelChatBadgesRequest::parse_response(None, &request.get_uri(), response)

Structs

GetChannelChatBadgesRequest

Query Parameters for Get Channel Chat Badges

Type Aliases

GetChannelChatBadgesResponse

Return Values for Get Channel Chat Badges