Module twitch_api::helix::chat::get_chat_settings

Available on crate feature helix only.

Gets the broadcaster’s chat settings. get-chat-settings

Accessing the endpoint

Authorization

Requires an App access token. However, to include the non_moderator_chat_delay or non_moderator_chat_delay_duration settings in the response, you must specify a User access token with scope set to moderator:read:chat_settings.

Request: GetChatSettingsRequest

To use this endpoint, construct a GetChatSettingsRequest with the GetChatSettingsRequest::broadcaster_id() method.

use twitch_api::{
    helix::{self, chat::get_chat_settings},
    types,
};
let request = get_chat_settings::GetChatSettingsRequest::broadcaster_id("1234567")
    // optional
    .moderator_id("9876543");

Response: ChatSettings

Send the request to receive the response with HelixClient::req_get().

use twitch_api::{helix::{self, chat::get_chat_settings}, types};
let request = get_chat_settings::GetChatSettingsRequest::broadcaster_id("1234567")
    // optional
    .moderator_id("9876543");
let response: helix::chat::ChatSettings = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetEmoteSetsRequest::parse_response(None, &request.get_uri(), response)

Structs

GetChatSettingsRequest

Query Parameters for Get Chat Settings