Module twitch_api::helix::chat::get_chatters

Available on crate feature helix only.

Gets the list of users that are connected to the specified broadcaster’s chat session. get-chatters

Notes

there is a delay between when users join and leave a chat and when the list is updated accordingly.

Accessing the endpoint

Request: GetChattersRequest

To use this endpoint, construct a GetChattersRequest with the GetChattersRequest::new() method.

use twitch_api::helix::chat::get_chatters;
let request = get_chatters::GetChattersRequest::new("1234", "4321");

Response: Chatter

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, chat::get_chatters};
let request = get_chatters::GetChattersRequest::new("1234", "4321");
let response: Vec<helix::chat::Chatter> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetChattersRequest::parse_response(None, &request.get_uri(), response)

Structs

Chatter

Return Values for Get Chatters

GetChattersRequest

Query Parameters for Get Chatters