Type Alias twitch_api::helix::chat::get_global_emotes::GetChannelEmotesResponse

pub type GetChannelEmotesResponse = GlobalEmote;

Available on crate feature helix only.

Return Values for Get Channel Emotes

get-global-emotes

Aliased Type

struct GetChannelEmotesResponse {
    pub id: EmoteId,
    pub name: String,
    pub images: Image,
    pub format: Vec<EmoteAnimationSetting, Global>,
    pub scale: Vec<EmoteScale, Global>,
    pub theme_mode: Vec<EmoteThemeMode, Global>,
}

Fields
id: EmoteId

ID of the emote.

name: String

Name of the emote a viewer types into Twitch chat for the image to appear.

images: Image

Object of image URLs for the emote.

format: Vec<EmoteAnimationSetting, Global>

The formats that the emote is available in.

scale: Vec<EmoteScale, Global>

The sizes that the emote is available in.

theme_mode: Vec<EmoteThemeMode, Global>

The background themes that the emote is available in.