twitch_api::helix::chat

Module get_shared_chat_session

Available on crate feature helix only.

Retrieves the active shared chat session for a channel. get-shared-chat-session

Accessing the endpoint

Request: GetSharedChatSessionRequest

To use this endpoint, construct a GetSharedChatSessionRequest with the GetSharedChatSessionRequest::broadcaster_id method.

use twitch_api::helix::chat::get_shared_chat_session;
let request =
    get_shared_chat_session::GetSharedChatSessionRequest::broadcaster_id(
        "12345",
    );

Response: SharedChatSession

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, chat::get_shared_chat_session};
let request = get_shared_chat_session::GetSharedChatSessionRequest::broadcaster_id("12345");
let response: Option<helix::chat::SharedChatSession> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetSharedChatSessionRequest::parse_response(None, &request.get_uri(), response)

Structs

GetSharedChatSessionRequest

Query Parameters for Get Shared Chat Session

SharedChatParticipant

A participant in a shared chat session

SharedChatSession

Return Values for Get Shared Chat Session