twitch_api::helix::chat
Module get_user_emotes

Available on crate feature helix only.

Retrieves emotes available to the user across all channels. get-user-emotes

Accessing the endpoint

Request: GetUserEmotesRequest

To use this endpoint, construct a GetUserEmotesRequest with the GetUserEmotesRequest::user_id() method.

use twitch_api::helix::chat::get_user_emotes;
let request = get_user_emotes::GetUserEmotesRequest::user_id("1234");

Response: UserEmote

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, chat::get_user_emotes};
let request = get_user_emotes::GetUserEmotesRequest::user_id("1234");
let response: Vec<helix::chat::get_user_emotes::UserEmote> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetUserEmotesRequest::parse_response(None, &request.get_uri(), response)

Structs

GetUserEmotesRequest

Query Parameters for Get User Emotes

UserEmote

Return Values for Get User Emotes