Module twitch_api::helix::chat::send_chat_announcement

Available on crate feature helix only.

Sends an announcement to the broadcaster’s chat room. send-chat-announcement

Accessing the endpoint

Request: SendChatAnnouncementRequest

To use this endpoint, construct a SendChatAnnouncementRequest with the SendChatAnnouncementRequest::new() method.

use twitch_api::helix::chat::send_chat_announcement;
let request = send_chat_announcement::SendChatAnnouncementRequest::new("1234", "5678");

Body: SendChatAnnouncementBody

We also need to provide a body to the request containing what we want to change.

let body =
    send_chat_announcement::SendChatAnnouncementBody::new("Hello chat!", "purple").unwrap();

Response: SendChatAnnouncementResponse

Send the request to receive the response with HelixClient::req_post().

use twitch_api::helix::{self, chat::send_chat_announcement};
let request = send_chat_announcement::SendChatAnnouncementRequest::new("1234", "5678");
let body = send_chat_announcement::SendChatAnnouncementBody::new(
    "Hello chat!",
    "purple",
).unwrap();
let response: helix::chat::SendChatAnnouncementResponse = client.req_post(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with SendChatAnnouncementRequest::parse_response(None, &request.get_uri(), response)

Structs

SendChatAnnouncementBody

Body Parameters for Send Chat Announcement

SendChatAnnouncementRequest

Query Parameters for Send Chat Announcement

Enums

SendChatAnnouncementResponse

Return Values for Send Chat Announcement