twitch_api::helix::clips::create_clip

Type Alias CreatedClip

pub type CreatedClip = CreatedClip;

Available on crate feature helix only.

Alias for super::CreatedClip

Aliased Type

pub struct CreatedClip {
    pub id: ClipId,
    pub edit_url: String,
}

Fields
id: ClipId

An ID that uniquely identifies the clip.

edit_url: String

A URL that you can use to edit the clip’s title, identify the part of the clip to publish, and publish the clip.

The URL is valid for up to 24 hours or until the clip is published, whichever comes first.