twitch_api::helix::moderation

Module add_suspicious_status_to_chat_user

Available on crate feature helix only.

Adds a suspicious user status to a chatter on the broadcaster’s channel. add-suspicious-status-to-chat-user

Accessing the endpoint

Request: AddSuspiciousStatusToChatUserRequest

To use this endpoint, construct a AddSuspiciousStatusToChatUserRequest with the AddSuspiciousStatusToChatUserRequest::new() method.

use twitch_api::helix::moderation::add_suspicious_status_to_chat_user;
let request = add_suspicious_status_to_chat_user::AddSuspiciousStatusToChatUserRequest::new("1234", "5678");

Body: AddSuspiciousStatusToChatUserBody

We also need to provide a body to the request containing what we want to change.

let body = add_suspicious_status_to_chat_user::AddSuspiciousStatusToChatUserBody::new("9876", AddedSuspiciousUserStatus::Restricted);

Response: SuspiciousUserInfo

Send the request to receive the response with HelixClient::req_post().

use twitch_api::helix::{self, moderation::{add_suspicious_status_to_chat_user, AddedSuspiciousUserStatus, SuspiciousUserInfo}};
let request = add_suspicious_status_to_chat_user::AddSuspiciousStatusToChatUserRequest::new("1234", "5678");
let body = add_suspicious_status_to_chat_user::AddSuspiciousStatusToChatUserBody::new("9876", AddedSuspiciousUserStatus::Restricted);
let response: SuspiciousUserInfo = client.req_post(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with AddSuspiciousStatusToChatUserRequest::parse_response(None, &request.get_uri(), response)

Structs

AddSuspiciousStatusToChatUserBody

Body Parameters for Add Suspicious Status to Chat User

AddSuspiciousStatusToChatUserRequest

Query Parameters for Add Suspicious Status to Chat User

Enums

AddedSuspiciousUserStatus

A user’s suspicious status when adding.