Module twitch_api::helix::moderation::check_automod_status

Available on crate feature helix only.

Determines whether a string message meets the channel’s AutoMod requirements. check-automod-status

Accessing the endpoint

Request: CheckAutoModStatusRequest

To use this endpoint, construct a CheckAutoModStatusRequest with the CheckAutoModStatusRequest::builder() method.

use twitch_api::helix::moderation::check_automod_status;
let request = check_automod_status::CheckAutoModStatusRequest::builder()
    .broadcaster_id("1234")
    .build();

Body: CheckAutoModStatusBody

We also need to provide a body to the request containing what we want to change.

let body = check_automod_status::CheckAutoModStatusBody::builder()
    .msg_id("test1")
    .msg_text("automod please approve this!")
    .user_id("1234")
    .build();

Response: CheckAutoModStatus

Send the request to receive the response with HelixClient::req_post().

use twitch_api::helix::{self, moderation::check_automod_status};
let request = check_automod_status::CheckAutoModStatusRequest::builder()
    .broadcaster_id("1234")
    .build();
let body = vec![check_automod_status::CheckAutoModStatusBody::builder()
    .msg_id("test1")
    .msg_text("automod please approve this!")
    .user_id("1234")
    .build()];
let response: Vec<check_automod_status::CheckAutoModStatus> = client.req_post(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with CheckAutoModStatusRequest::parse_response(None, &request.get_uri(), response)

Structs

CheckAutoModStatus

Return Values for Check AutoMod Status

CheckAutoModStatusBody

Body Parameters for Check AutoMod Status

CheckAutoModStatusRequest

Query Parameters for Check AutoMod Status