Module twitch_api::helix::moderation::get_blocked_terms

Available on crate feature helix only.

Gets the broadcaster’s list of non-private, blocked words or phrases. These are the terms that the broadcaster or moderator added manually, or that were denied by AutoMod. get-blocked-terms

Accessing the endpoint

Request: GetBlockedTerms

To use this endpoint, construct a GetBlockedTerms with the GetBlockedTerms::new() method.

use twitch_api::helix::moderation::get_blocked_terms;
let request = get_blocked_terms::GetBlockedTerms::new("1234", "5678");

Response: BlockedTerm

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, moderation::get_blocked_terms};
let request = get_blocked_terms::GetBlockedTerms::new("1234", "5678");
let response: Vec<helix::moderation::BlockedTerm> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetBlockedTerms::parse_response(None, &request.get_uri(), response)

Structs

GetBlockedTerms

Query Parameters for Get Blocked Terms

Type Definitions

GetBlockedTermsResponse

Return Values for Get Blocked Terms