twitch_api::helix::moderation::get_blocked_terms
Type Alias GetBlockedTermsResponse

pub type GetBlockedTermsResponse = BlockedTerm;

Available on crate feature helix only.

Return Values for Get Blocked Terms

get-blocked-terms

Aliased Type

struct GetBlockedTermsResponse {
    pub broadcaster_id: UserId,
    pub created_at: Timestamp,
    pub expires_at: Option<Timestamp>,
    pub id: BlockedTermId,
    pub moderator_id: UserId,
    pub text: String,
    pub updated_at: Timestamp,
}

Fields
broadcaster_id: UserId

The broadcaster that owns the list of blocked terms.

created_at: Timestamp

The UTC date and time (in RFC3339 format) of when the term was blocked.

expires_at: Option<Timestamp>

The UTC date and time (in RFC3339 format) of when the blocked term is set to expire. After the block expires, user’s will be able to use the term in the broadcaster’s chat room.

This field is null if the term was added manually or was permanently blocked by AutoMod.

id: BlockedTermId

An ID that uniquely identifies this blocked term.

moderator_id: UserId

The moderator that blocked the word or phrase from being used in the broadcaster’s chat room.

text: String

The blocked word or phrase.

updated_at: Timestamp

The UTC date and time (in RFC3339 format) of when the term was updated.

When the term is added, this timestamp is the same as created_at. The timestamp changes as AutoMod continues to deny the term.