twitch_api::helix::moderation

Module remove_blocked_term

Available on crate feature helix only.

Removes the word or phrase that the broadcaster is blocking users from using in their chat room. remove-blocked-term

Accessing the endpoint

Request: RemoveBlockedTermRequest

To use this endpoint, construct a RemoveBlockedTermRequest with the RemoveBlockedTermRequest::new() method.

use twitch_api::helix::moderation::remove_blocked_term;
let request = remove_blocked_term::RemoveBlockedTermRequest::new(
    "1234", "5678", "DEADBEEF",
);

Response: RemoveBlockedTerm

Send the request to receive the response with HelixClient::req_delete().

use twitch_api::helix::{self, moderation::remove_blocked_term};
let request = remove_blocked_term::RemoveBlockedTermRequest::new("1234", "5678", "DEADBEEF");
let response: remove_blocked_term::RemoveBlockedTerm = client.req_delete(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with RemoveBlockedTermRequest::parse_response(None, &request.get_uri(), response)

Structs

RemoveBlockedTermRequest

Query Parameters for Remove Blocked Terms

Enums

RemoveBlockedTerm

Return Values for Remove Blocked Terms