twitch_api::helix::moderation
Module resolve_unban_request

Available on crate feature helix only.

Resolves an unban request by approving or denying it.

resolve-unban-requests

Accessing the endpoint

Request: ResolveUnbanRequest

To use this endpoint, construct an ResolveUnbanRequest with the ResolveUnbanRequest::new() method.

use twitch_api::helix::moderation::resolve_unban_request;
let request = resolve_unban_request::ResolveUnbanRequest::approve(
    "123",
    "456",
    "123-456-789",
)
.resolution_text("something");

Response: UnbanRequest

Send the request to receive the response with HelixClient::req_patch().

use twitch_api::helix::{self, moderation::resolve_unban_request};
let request = resolve_unban_request::ResolveUnbanRequest::approve(
    "123",
    "456",
    "123-456-789",
);
let response: helix::moderation::UnbanRequest = client.req_patch(request, helix::EmptyBody, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with ResolveUnbanRequest::parse_response(None, &request.get_uri(), response)

Structs

ResolveUnbanRequest

Query Parameters for Resolve Unban Request

UnbanRequest

Return Values for Get Unban Requests

Enums

UnbanRequestStatus

The status of an unban request