twitch_api::helix::moderation
Module update_shield_mode_status

Available on crate feature helix only.

Activates or deactivates the broadcaster’s Shield Mode.

update-shield-mode-status

Accessing the endpoint

Request: UpdateShieldModeStatusRequest

To use this endpoint, construct an UpdateShieldModeStatusRequest with the UpdateShieldModeStatusRequest::new() method.

use twitch_api::helix::moderation::update_shield_mode_status;
let request = update_shield_mode_status::UpdateShieldModeStatusRequest::new(
    "123", "456",
);

Body: UpdateShieldModeStatusBody

We also need to provide a body to the request.

let body =
    update_shield_mode_status::UpdateShieldModeStatusBody::is_active(false);

Response: ShieldModeStatus

Send the request to receive the response with HelixClient::req_patch().

use twitch_api::helix::{self, moderation::update_shield_mode_status};
let request = update_shield_mode_status::UpdateShieldModeStatusRequest::new(
    "123",
    "456"
);
let body =
    update_shield_mode_status::UpdateShieldModeStatusBody::is_active(false);
let response: helix::moderation::ShieldModeStatus = client.req_put(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with UpdateShieldModeStatusRequest::parse_response(None, &request.get_uri(), response)

Structs

ShieldModeStatus

Return Values for Get Shield Mode Status

UpdateShieldModeStatusBody

Body Parameters for Update Shield Mode Status

UpdateShieldModeStatusRequest

Query Parameters for Update Shield Mode Status