twitch_api::helix::pagination

Trait PaginationData

pub trait PaginationData<R: Request> {
    // Required methods
    fn new(
        cursor: Option<Cursor>,
        request: Option<R>,
        total: Option<i64>,
    ) -> Self;
    fn total(&self) -> Option<i64>;
}

Available on crate feature helix only.

Data used to keep track of the pagination for a request.

Required Methods

fn new(cursor: Option<Cursor>, request: Option<R>, total: Option<i64>) -> Self

Create the state for a response and its request

fn total(&self) -> Option<i64>

Get the total number of items in the data

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.