Type Alias twitch_api::helix::polls::create_poll::CreatePollResponse

pub type CreatePollResponse = Poll;

Available on crate feature helix only.

Return Values for Create Poll

create-poll

Aliased Type

struct CreatePollResponse {

    pub id: PollId,
    pub broadcaster_id: UserId,
    pub broadcaster_name: Nickname,
    pub broadcaster_login: DisplayName,
    pub title: String,
    pub choices: Vec<PollChoice>,
    pub bits_voting_enabled: bool,
    pub bits_per_vote: i64,
    pub channel_points_voting_enabled: bool,
    pub channel_points_per_vote: i64,
    pub status: PollStatus,
    pub duration: i64,
    pub started_at: Timestamp,
    pub ended_at: Option<Timestamp>,

}

Fields
id: PollId

ID of the poll.

broadcaster_id: UserId

ID of the broadcaster.

broadcaster_name: Nickname

Name of the broadcaster.

broadcaster_login: DisplayName

Login of the broadcaster.

title: String

Question displayed for the poll.

choices: Vec<PollChoice>

Array of the poll choices.

bits_voting_enabled: bool
👎Deprecated since 0.7.0: the bits options for polls has been removed

Indicates if Bits can be used for voting.

bits_per_vote: i64
👎Deprecated since 0.7.0: the bits options for polls has been removed

Number of Bits required to vote once with Bits.

channel_points_voting_enabled: bool

Indicates if Channel Points can be used for voting.

channel_points_per_vote: i64

Number of Channel Points required to vote once with Channel Points.

status: PollStatus

Poll status. Valid values are:

duration: i64

Total duration for the poll (in seconds).

started_at: Timestamp

UTC timestamp for the poll’s start time.

ended_at: Option<Timestamp>

UTC timestamp for the poll’s end time. Set to null if the poll is active.