Module twitch_api::helix::predictions::end_prediction

Available on crate feature helix only.

End a prediction that is currently active.

end-prediction

Accessing the endpoint

Request: EndPredictionRequest

To use this endpoint, construct an EndPredictionRequest with the EndPredictionRequest::new() method.

use twitch_api::helix::predictions::end_prediction;
let request = end_prediction::EndPredictionRequest::new();

Body: EndPredictionBody

We also need to provide a body to the request containing what we want to change.

let body = end_prediction::EndPredictionBody::builder()
    .broadcaster_id("141981764")
    .id("ed961efd-8a3f-4cf5-a9d0-e616c590cd2a")
    .status(end_prediction::PredictionStatus::Resolved)
    .build();

Response: EndPrediction

Send the request to receive the response with HelixClient::req_patch().

use twitch_api::helix::{self, predictions::end_prediction};
let request = end_prediction::EndPredictionRequest::new();
let body = end_prediction::EndPredictionBody::builder()
    .broadcaster_id("141981764")
    .id("ed961efd-8a3f-4cf5-a9d0-e616c590cd2a")
    .status(end_prediction::PredictionStatus::Resolved)
    .build();
let response: end_prediction::EndPrediction = client.req_patch(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with EndPredictionRequest::parse_response(None, &request.get_uri(), response)

Structs

EndPredictionBody

Body Parameters for End Prediction

EndPredictionRequest

Query Parameters for End Prediction

PredictionId

A prediction ID

Enums

EndPrediction

Return Values for Update CustomReward

PredictionStatus

Status of the Prediction