Module twitch_api::helix::predictions::get_predictions

Available on crate feature helix only.

Get information about all predictions or specific predictions for a Twitch channel. Prediction information is available for 90 days. get-predictions

Request: GetPredictionsRequest

To use this endpoint, construct a GetPredictionsRequest with the GetPredictionsRequest::broadcaster_id() method.

use twitch_api::helix::predictions::get_predictions;
let request = get_predictions::GetPredictionsRequest::broadcaster_id("1234")
    .ids(vec!["ed961efd-8a3f-4cf5-a9d0-e616c590cd2a".into()]);

Response: Prediction

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, predictions::get_predictions};
let request = get_predictions::GetPredictionsRequest::broadcaster_id("1234")
    .ids(vec!["ed961efd-8a3f-4cf5-a9d0-e616c590cd2a".into()]);
let response: Vec<get_predictions::Prediction> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetPredictionsRequest::parse_response(None, &request.get_uri(), response)

Structs

GetPredictionsRequest

Query Parameters for Get predictions

Prediction

Return Values for Get predictions

PredictionOutcome

Outcome for the Prediction

PredictionOutcomeId

A prediction choice ID

Enums

PredictionStatus

Status of the Prediction