twitch_api::helix::schedule
Module delete_channel_stream_schedule_segment

Available on crate feature helix only.

Delete a single scheduled broadcast or a recurring scheduled broadcast for a channel’s stream schedule. delete-channel-stream-schedule-segment

Notes

Accessing the endpoint

Request: DeleteChannelStreamScheduleSegmentRequest

To use this endpoint, construct a DeleteChannelStreamScheduleSegmentRequest with the DeleteChannelStreamScheduleSegmentRequest::new() method.

use twitch_api::helix::schedule::delete_channel_stream_schedule_segment;
let request = delete_channel_stream_schedule_segment::DeleteChannelStreamScheduleSegmentRequest::new(
    "1234",
    "eyJzZWdtZW50SUQiOiI4Y2EwN2E2NC0xYTZkLTRjYWItYWE5Ni0xNjIyYzNjYWUzZDkiLCJpc29ZZWFyIjoyMDIxLCJpc29XZWVrIjoyMX0="
);

Response: DeleteChannelStreamScheduleSegment

Send the request to receive the response with HelixClient::req_delete().

use twitch_api::helix::{self, schedule::delete_channel_stream_schedule_segment};
let request = delete_channel_stream_schedule_segment::DeleteChannelStreamScheduleSegmentRequest::new(
    "1234",
    "eyJzZWdtZW50SUQiOiI4Y2EwN2E2NC0xYTZkLTRjYWItYWE5Ni0xNjIyYzNjYWUzZDkiLCJpc29ZZWFyIjoyMDIxLCJpc29XZWVrIjoyMX0="
);
let response: delete_channel_stream_schedule_segment::DeleteChannelStreamScheduleSegment = client.req_delete(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with DeleteChannelStreamScheduleSegmentRequest::parse_response(None, &request.get_uri(), response)

Structs

DeleteChannelStreamScheduleSegmentRequest

Query Parameters for Delete Channel Stream Schedule Segment

Enums

DeleteChannelStreamScheduleSegment

Return Values for Delete Channel Stream Schedule Segment