Module twitch_api::helix::schedule::get_channel_stream_schedule

Available on crate feature helix only.

Gets all scheduled broadcasts or specific scheduled broadcasts from a channel’s stream schedule. Scheduled broadcasts are defined as “stream segments” in the API. get-channel-stream-schedule

Notes

See also get_channel_schedule

Request: GetChannelStreamScheduleRequest

To use this endpoint, construct a GetChannelStreamScheduleRequest with the GetChannelStreamScheduleRequest::builder() method.

use twitch_api::helix::schedule::get_channel_stream_schedule;
let request = get_channel_stream_schedule::GetChannelStreamScheduleRequest::builder()
    .broadcaster_id("1234")
    .build();

Response: ScheduledBroadcasts

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, schedule::get_channel_stream_schedule};
let request = get_channel_stream_schedule::GetChannelStreamScheduleRequest::builder()
    .broadcaster_id("1234")
    .build();
let response: helix::schedule::ScheduledBroadcasts = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetChannelStreamScheduleRequest::parse_response(None, &request.get_uri(), response)

Structs

GetChannelStreamScheduleRequest

Query Parameters for Get Channel Stream Schedule

Type Definitions

GetChannelStreamScheduleResponse

Return Values for Get Channel Stream Schedule