twitch_api::helix::schedule::get_channel_stream_schedule

Type Alias GetChannelStreamScheduleResponse

pub type GetChannelStreamScheduleResponse = ScheduledBroadcasts;

Available on crate feature helix only.

Return Values for Get Channel Stream Schedule

get-channel-stream-schedule

Aliased Type

struct GetChannelStreamScheduleResponse {
    pub segments: Vec<Segment>,
    pub broadcaster_id: UserId,
    pub broadcaster_name: DisplayName,
    pub broadcaster_login: Nickname,
    pub vacation: Option<Vacation>,
}

Fields
segments: Vec<Segment>

Scheduled broadcasts for this stream schedule.

broadcaster_id: UserId

User ID of the broadcaster.

broadcaster_name: DisplayName

Display name of the broadcaster.

broadcaster_login: Nickname

Login of the broadcaster.

vacation: Option<Vacation>

If Vacation Mode is enabled, this includes start and end dates for the vacation. If Vacation Mode is disabled, value is set to null.