Module twitch_api::helix::schedule::update_channel_stream_schedule_segment

Available on crate feature helix only.

Update a single scheduled broadcast or a recurring scheduled broadcast for a channel’s stream schedule. update-channel-stream-schedule-segment

Accessing the endpoint

Request: UpdateChannelStreamScheduleSegmentRequest

To use this endpoint, construct a UpdateChannelStreamScheduleSegmentRequest with the UpdateChannelStreamScheduleSegmentRequest::new() method.

use twitch_api::helix::schedule::update_channel_stream_schedule_segment;
let request = update_channel_stream_schedule_segment::UpdateChannelStreamScheduleSegmentRequest::new(
    "141981764",
    "eyJzZWdtZW50SUQiOiJlNGFjYzcyNC0zNzFmLTQwMmMtODFjYS0yM2FkYTc5NzU5ZDQiLCJpc29ZZWFyIjoyMDIxLCJpc29XZWVrIjoyNn0=",
);

Body: UpdateChannelStreamScheduleSegmentBody

We also need to provide a body to the request containing what we want to change.

let body =
    update_channel_stream_schedule_segment::UpdateChannelStreamScheduleSegmentBody::builder()
        .duration(Some("120".into()))
        .build();

Response: UpdateChannelStreamScheduleSegmentResponse

Send the request to receive the response with HelixClient::req_post().

use twitch_api::helix::{self, schedule::update_channel_stream_schedule_segment};
let request = update_channel_stream_schedule_segment::UpdateChannelStreamScheduleSegmentRequest::new(
    "141981764",
    "eyJzZWdtZW50SUQiOiJlNGFjYzcyNC0zNzFmLTQwMmMtODFjYS0yM2FkYTc5NzU5ZDQiLCJpc29ZZWFyIjoyMDIxLCJpc29XZWVrIjoyNn0=",
);
let body = update_channel_stream_schedule_segment::UpdateChannelStreamScheduleSegmentBody::builder()
    .duration(Some("120".into()))
    .build();
let response: update_channel_stream_schedule_segment::UpdateChannelStreamScheduleSegmentResponse = client.req_patch(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&body, &token, &client_id) and parse the http::Response with UpdateChannelStreamScheduleSegmentRequest::parse_response(None, &request.get_uri(), response)

Structs

UpdateChannelStreamScheduleSegmentBody

Body Parameters for Update Channel Stream Schedule Segment

UpdateChannelStreamScheduleSegmentRequest

Query Parameters for Update Channel Stream Schedule Segment

Type Definitions

UpdateChannelStreamScheduleSegmentResponse

Return Values for Update Channel Stream Schedule Segment