Struct twitch_api::helix::search::SearchCategoriesRequest

pub struct SearchCategoriesRequest<'a> {
    pub query: Cow<'a, str>,
    pub after: Option<Cow<'a, CursorRef>>,
    pub before: Option<Cow<'a, CursorRef>>,
    pub first: Option<usize>,
}

Available on crate feature helix only.

Query Parameters for Search Categories

search-categories

Fields
query: Cow<'a, str>

URI encoded search query

after: Option<Cow<'a, CursorRef>>

Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response. The cursor value specified here is from the pagination response field of a prior query.

before: Option<Cow<'a, CursorRef>>

Cursor for backward pagination: tells the server where to start fetching the next set of results, in a multi-page response. The cursor value specified here is from the pagination response field of a prior query.

first: Option<usize>

Number of values to be returned per page. Limit: 100. Default: 20.

Implementations

impl<'a> SearchCategoriesRequest<'a>

pub fn builder() -> SearchCategoriesRequestBuilder<'a, ((), (), (), ())>

Create a builder for building SearchCategoriesRequest. On the builder, call .query(...), .after(...)(optional), .before(...)(optional), .first(...)(optional) to set the values of the fields. Finally, call .build() to create the instance of SearchCategoriesRequest.

impl<'a> SearchCategoriesRequest<'a>

pub fn query(query: impl Into<Cow<'a, str>>) -> Self

Search categories with the following query.

pub fn first(self, first: usize) -> Self

Set amount of results returned per page.