Module twitch_api::helix::streams::get_followed_streams

Available on crate feature helix only.

Gets information about active streams belonging to channels that the authenticated user follows. get-followed-streams

Accessing the endpoint

Request: GetFollowedStreamsRequest

To use this endpoint, construct a GetFollowedStreamsRequest with the GetFollowedStreamsRequest::user_id() method.

use twitch_api::helix::streams::get_followed_streams;
let request =
    get_followed_streams::GetFollowedStreamsRequest::user_id("1234");

Response: Stream

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, streams::get_followed_streams};
let request = get_followed_streams::GetFollowedStreamsRequest::user_id("1234");
let response: Vec<get_followed_streams::GetFollowedStreamsResponse> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetFollowedStreamsRequest::parse_response(None, &request.get_uri(), response)

Structs

GetFollowedStreamsRequest

Query Parameters for Get Followed Streams

Type Aliases

GetFollowedStreamsResponse

Return Values for Get Followed Streams