twitch_api::helix::streams::get_stream_tags
Type Alias Tag

pub type Tag = TwitchTag;

Available on crate feature helix only.

Return Values for Get Stream Tags

get-stream-tags

Aliased Type

struct Tag {
    pub id: TagId,
    pub is_auto: AutoGenerated,
    pub localization_names: HashMap<TagLanguage, String>,
    pub localization_descriptions: HashMap<TagLanguage, String>,
}

Fields
id: TagId
👎Deprecated: Twitch-defined tags have been deprecated. See https://discuss.dev.twitch.tv/t/adding-customizable-tags-to-the-twitch-api/42921

ID of the tag.

is_auto: AutoGenerated
👎Deprecated: Twitch-defined tags have been deprecated. See https://discuss.dev.twitch.tv/t/adding-customizable-tags-to-the-twitch-api/42921

true if the tag is auto-generated; otherwise, false . An auto-generated tag is one automatically applied by Twitch (e.g., a language tag based on the broadcaster’s settings); these cannot be added or removed by the user.

localization_names: HashMap<TagLanguage, String>
👎Deprecated: Twitch-defined tags have been deprecated. See https://discuss.dev.twitch.tv/t/adding-customizable-tags-to-the-twitch-api/42921

All localized names of the tag.

localization_descriptions: HashMap<TagLanguage, String>
👎Deprecated: Twitch-defined tags have been deprecated. See https://discuss.dev.twitch.tv/t/adding-customizable-tags-to-the-twitch-api/42921

All localized descriptions of the tag.