twitch_api::helix

Module tags

Available on crate feature helix only.

Helix endpoints regarding tags

Examples

let client = HelixClient::new();
let req = GetAllStreamTagsRequest::default();

println!("{:?}", &client.req_get(req, &token).await?.data.first());

Implemented endpoints

Tags ๐ข 2/2

Endpoint	Helper	Module

Get All Stream Tags	-	get_all_stream_tags

Get Stream Tags	-	get_stream_tags

Modules

get_all_stream_tags

Gets the list of all stream tags defined by Twitch, optionally filtered by tag ID(s). get-all-stream-tags

get_stream_tags

Gets the list of tags for a specified stream (channel). get-stream-tags

Structs

GetAllStreamTagsRequestDeprecated

Query Parameters for Get All Stream Tags

GetStreamTagsRequestDeprecated

Query Parameters for Get Stream Tags

TagLanguageDeprecated

Language code, formatted as 2 letter language by ISO 639-1, a dash (-) and 2 letter region by ISO 3166-1

TagLanguageRefDeprecated

Language code, formatted as 2 letter language by ISO 639-1, a dash (-) and 2 letter region by ISO 3166-1

TwitchTagDeprecated

A stream tag as defined by Twitch.

Enums

AutoGeneratedDeprecated

Tag is auto-generated or not.

Type Aliases

Tag

Return Values for Get All Stream Tags