twitch_api::helix::tags
Module get_all_stream_tags

Available on crate feature helix only.

Gets the list of all stream tags defined by Twitch, optionally filtered by tag ID(s). get-all-stream-tags

Accessing the endpoint

Request: GetAllStreamTagsRequest

To use this endpoint, construct a GetAllStreamTagsRequest with the GetAllStreamTagsRequest::default() method.

use twitch_api::helix::tags::get_all_stream_tags;
let request =
    get_all_stream_tags::GetAllStreamTagsRequest::default().first(100);

Response: Tag

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, tags::get_all_stream_tags};
let request = get_all_stream_tags::GetAllStreamTagsRequest::default();
let response: Vec<get_all_stream_tags::Tag> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetAllStreamTagsRequest::parse_response(None, &request.get_uri(), response)

Structs

GetAllStreamTagsRequestDeprecated

Query Parameters for Get All Stream Tags

Type Aliases

Tag

Return Values for Get All Stream Tags