twitch_api::helix::users

Module get_user_active_extensions

Available on crate feature helix only.

Gets the active extensions that the broadcaster has installed for each configuration. get-user-active-extensions

Request: GetUserActiveExtensionsRequest

To use this endpoint, construct a GetUserActiveExtensionsRequest with the GetUserActiveExtensionsRequest::new() method.

use twitch_api::helix::users::get_user_active_extensions;
let request =
    get_user_active_extensions::GetUserActiveExtensionsRequest::new();

Response: ExtensionConfiguration

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, users::get_user_active_extensions};
let request = get_user_active_extensions::GetUserActiveExtensionsRequest::new();
let response: get_user_active_extensions::ExtensionConfiguration = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetUserActiveExtensionsRequest::parse_response(None, &request.get_uri(), response)

Structs

ActiveExtension

An active extension slot

ActivePositionedExtension

An active extension slot where the extension can be positioned

ExtensionConfiguration

Return Values for Get User Active Extensions

GetUserActiveExtensionsRequest

Query Parameters for Get User Active Extensions