Module twitch_api::helix::users::get_user_block_list

Available on crate feature helix only.

Gets a specified user’s block list. The list is sorted by when the block occurred in descending order (i.e. most recent block first). get-user-block-list

Request: GetUserBlockListRequest

To use this endpoint, construct a GetUserBlockListRequest with the GetUserBlockListRequest::broadcaster_id() method.

use twitch_api::helix::users::get_user_block_list;
let request = get_user_block_list::GetUserBlockListRequest::broadcaster_id("1234");

Response: UserBlock

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, users::get_user_block_list};
let request = get_user_block_list::GetUserBlockListRequest::broadcaster_id("1234");
let response: Vec<get_user_block_list::UserBlock> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetUserBlockListRequest::parse_response(None, &request.get_uri(), response)

Structs

GetUserBlockListRequest

Query Parameters for Get Users Block List

UserBlock

Return Values for Get Users Block List