twitch_api::helix::users

Module update_user

Available on crate feature helix only.

Updates the specified user’s information.

update-user

The user ID in the OAuth token identifies the user whose information you want to update.

Accessing the endpoint

Request: UpdateUserRequest

To use this endpoint, construct an UpdateUserRequest with the UpdateUserRequest::description() method.

use twitch_api::helix::users::update_user;
let mut request =
    update_user::UpdateUserRequest::description("my description");

Response: User

Send the request to receive the response with HelixClient::req_patch().

use twitch_api::helix::{self, users::update_user};
let mut request = update_user::UpdateUserRequest::description("my description");
let body = helix::EmptyBody;
let response: helix::users::User = client.req_put(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with UpdateUserRequest::parse_response(None, &request.get_uri(), response)

Structs

UpdateUserRequest

Query Parameters for Update User