twitch_api::helix::users

Module update_user_extensions

Available on crate feature helix only.

Updates the specified user’s information.

update-user-extensions

The user ID in the OAuth token identifies the user whose information you want to update.

Accessing the endpoint

Request: UpdateUserExtensionsRequest

To use this endpoint, construct an UpdateUserExtensionsRequest with the UpdateUserExtensionsRequest::new() method.

use twitch_api::helix::users::update_user_extensions;
let mut request =
    update_user_extensions::UpdateUserExtensionsRequest::new();

Body: UpdateUserExtensionsBody

We also need to provide a body to the request.

let body = update_user_extensions::UpdateUserExtensionsBody::new(
    update_user_extensions::ExtensionSpecification::new().panel(
        HashMap::from_iter([(
            Cow::Borrowed("1"),
            users::ExtensionSlot::Inactive,
        )]),
    ),
);

Response: ExtensionConfiguration

Send the request to receive the response with HelixClient::req_patch().

use twitch_api::helix::{self, users::update_user_extensions};
let mut request = update_user_extensions::UpdateUserExtensionsRequest::new();
let body = update_user_extensions::UpdateUserExtensionsBody::new(
    update_user_extensions::ExtensionSpecification::new()
        .panel(HashMap::from_iter([(Cow::Borrowed("1"), helix::users::ExtensionSlot::Inactive)]))
);
let response: helix::users::ExtensionConfiguration = client.req_put(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with UpdateUserExtensionsRequest::parse_response(None, &request.get_uri(), response)

Structs

ActiveExtension

An active extension slot

ActivePositionedExtension

An active extension slot where the extension can be positioned

ExtensionSpecification

Inner body for Update User Extensions

UpdateUserExtensionsBody

Body for Update User Extensions

UpdateUserExtensionsRequest

Query Parameters for Update User Extensions