twitch_api::pubsub
Enum Topics

#[non_exhaustive]
pub enum Topics {

    AutoModQueue(AutoModQueue),
    CommunityPointsChannelV1(CommunityPointsChannelV1),
    ChannelBitsEventsV2(ChannelBitsEventsV2),
    ChannelBitsBadgeUnlocks(ChannelBitsBadgeUnlocks),
    ChannelCheerEventsPublicV1(ChannelCheerEventsPublicV1),
    ChannelSubGiftsV1(ChannelSubGiftsV1),
    ChatModeratorActions(ChatModeratorActions),
    ChannelPointsChannelV1(ChannelPointsChannelV1),
    ChannelSubscribeEventsV1(ChannelSubscribeEventsV1),
    VideoPlayback(VideoPlayback),
    VideoPlaybackById(VideoPlaybackById),
    HypeTrainEventsV1(HypeTrainEventsV1),
    HypeTrainEventsV1Rewards(HypeTrainEventsV1Rewards),
    Following(Following),
    Raid(Raid),
    UserModerationNotifications(UserModerationNotifications),

}

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature pubsub only.

All possible topics

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AutoModQueue(AutoModQueue)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

AutoMod flags a message as potentially inappropriate, and when a moderator takes action on a message.

CommunityPointsChannelV1(CommunityPointsChannelV1)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature unsupported only.

A user redeems an reward using channel points.

ChannelBitsEventsV2(ChannelBitsEventsV2)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Anyone cheers in a specified channel.

ChannelBitsBadgeUnlocks(ChannelBitsBadgeUnlocks)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Anyone shares a bit badge in a specified channel.

ChannelCheerEventsPublicV1(ChannelCheerEventsPublicV1)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature unsupported only.

A user redeems a cheer with shared rewards.

ChannelSubGiftsV1(ChannelSubGiftsV1)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature unsupported only.

A user gifts subs.

ChatModeratorActions(ChatModeratorActions)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

A moderator performs an action in the channel.

ChannelPointsChannelV1(ChannelPointsChannelV1)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

A user redeems an reward using channel points.

ChannelSubscribeEventsV1(ChannelSubscribeEventsV1)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

A subscription event happens in channel

VideoPlayback(VideoPlayback)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature unsupported only.

Statistics about stream

VideoPlaybackById(VideoPlaybackById)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature unsupported only.

Statistics about stream

HypeTrainEventsV1(HypeTrainEventsV1)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature unsupported only.

A user redeems an reward using channel points.

HypeTrainEventsV1Rewards(HypeTrainEventsV1Rewards)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature unsupported only.

A user redeems an reward using channel points.

Following(Following)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature unsupported only.

A user follows the channel

Raid(Raid)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature unsupported only.

A user raids the channel

UserModerationNotifications(UserModerationNotifications)

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

A user’s message held by AutoMod has been approved or denied.