Struct twitch_api::tmi::TmiClient

pub struct TmiClient<'a, C: HttpClient<'a>> { /* private fields */ }

Available on crate features client and tmi only.

Client for the twitch TMI endpoint, almost entirely undocumented and certainly not supported.

Examples

let client = TmiClient::new();
println!("{:?}", client.get_chatters("justinfan10".into()).await?);

Most clients will be able to use the 'static lifetime

pub struct MyStruct {
    twitch: TmiClient<'static, surf::Client>,
    token: twitch_oauth2::AppAccessToken,
}
// etc

See HttpClient for implemented http clients, you can also define your own if needed.

Implementations

impl<'a, C: HttpClient<'a>> TmiClient<'a, C>

pub fn with_client(client: C) -> TmiClient<'a, C>

Create a new client with an existing client

pub fn new() -> TmiClient<'a, C>where
C: ClientDefault<'a>,

Create a new TmiClient with a default HttpClient

pub fn clone_client(&self) -> Cwhere
C: Clone,

Retrieve a clone of the HttpClient inside this TmiClient

pub fn get_client(&self) -> &C

Retrieve a reference of the HttpClient inside this TmiClient

pub async fn get_chatters(
&'a self,
broadcaster: &UserNameRef
) -> Result<GetChatters, RequestError<<C as HttpClient<'a>>::Error>>

Get all the chatters in the chat

Notes

This function will aside from url sanitize the broadcasters username, will also remove any # and make it lowercase ascii

Examples found in repository?

examples/get_streams_and_chatters.rs (line 33)

≺ ≻↕

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38

async fn main() {
    let _ = dotenv::dotenv();
    let mut args = std::env::args().skip(1);
    let client: TwitchClient<surf::Client> = TwitchClient::new();
    let token = UserToken::from_existing(
        &client,
        std::env::var("TWITCH_TOKEN")
            .ok()
            .or_else(|| args.next())
            .map(AccessToken::new)
            .expect("Please set env: TWITCH_TOKEN or pass token as first argument"),
        None,
        None,
    )
    .await
    .unwrap();

    let streams: Vec<String> = args.collect();
    let req = GetStreamsRequest::builder().build();

    let response = client.helix.req_get(req, &token).await.unwrap();

    // Note: This will fetch chatters in the current most viewed stream, might spam your console a bit.
    println!("GetStreams:\n\t{:?}", response.data);
    if let Some(stream) = streams.get(0) {
        println!(
            "{:?}",
            client.tmi.get_chatters(stream.as_str().into()).await
        );
    } else if let Some(stream) = response.data.get(0).map(|stream| &stream.user_login) {
        println!("{:?}", client.tmi.get_chatters(stream).await);
    }
}