twitch_api::client
Trait ClientDefault

pub trait ClientDefault<'a>: Clone + Sized {
    type Error: Error + Send + Sync + 'static;

    // Required method
    fn default_client_with_name(
        product: Option<HeaderValue>,
    ) -> Result<Self, Self::Error>;

    // Provided method
    fn default_client() -> Self { ... }
}

Available on crate feature client only.

A specific client default for setting some sane defaults for API calls and oauth2 usage

Required Associated Types

type Error: Error + Send + Sync + 'static

Errors that can happen when assembling the client

Required Methods

fn default_client_with_name( product: Option<HeaderValue>, ) -> Result<Self, Self::Error>

Constructs Self with sane defaults for API calls and oauth2 and setting user-agent to include another product

Specifically, one should

Set User-Agent to {product} twitch_api/{version_of_twitch_api} (+https://github.com/twitch-rs) (According to RFC7231) See TWITCH_API_USER_AGENT for the product of this crate

Disallow redirects

Notes

When the product name is none, this function should never fail. This should be ensured with tests.

Provided Methods

fn default_client() -> Self

Construct Self with sane defaults for API calls and oauth2.

Object Safety

This trait is not object safe.