Module twitch_api::eventsub

Available on crate feature eventsub only.

Holds serializable EventSub stuff

Use CreateEventSubSubscription to subscribe to an event according to the EventSub guide. Parse the response payload text with Event::parse_http or Event::parse_websocket

Examples

See examples/ for a more complete example of using eventsub.

Subscribe to a channel’s follow events:

use twitch_api::eventsub::{channel::ChannelFollowV1, Transport, TransportMethod};
use twitch_api::helix::{self, eventsub::{
    CreateEventSubSubscriptionBody, CreateEventSubSubscriptionRequest,
}};

let event = ChannelFollowV1::builder()
    .broadcaster_user_id("1234")
    .build();
let transport = Transport::webhook(
    "https://example.org/eventsub/channelfollow",
    String::from("secretabcd"),
);

let request = CreateEventSubSubscriptionRequest::default();
let body = CreateEventSubSubscriptionBody::builder()
    .subscription(event)
    .transport(transport)
    .build();

let event_information = client.req_post(request, body, &token).await?.data;

println!("event id: {:?}", event_information.id);

You’ll now get a http POST request to the url you specified as the callback. You need to respond to this request from your webserver with a 200 OK response with the challenge as the body. After this, you’ll get notifications

use twitch_api::eventsub::{Event, Payload, Message};
pub fn parse_request(
    request: &http::Request<Vec<u8>>,
) -> Result<http::Response<Vec<u8>>, Box<dyn std::error::Error + Send + Sync + 'static>> {
    // First, we verify the response, assuring it's legit.
    if !Event::verify_payload(request, b"secretabcd") {
        return Err(todo!());
    }
    match Event::parse_http(request)? {
        Event::ChannelFollowV1(Payload {
            message: Message::VerificationRequest(ver),
            ..
        }) => {
            // We've verified the request, so we can respond to it with the challenge
            Ok(http::Response::builder()
                .status(200)
                .body(ver.challenge.into_bytes())?)
        },
        Event::ChannelFollowV1(Payload {
            message: Message::Notification(notif),
            ..
        }) => {
            // make sure you save the `Twitch-Eventsub-Message-Id` headers value,
            // twitch may resend notifications, and in those cases you should just return 200 OK.

            // Do whatever you need to do with the event. Preferably send the event to a channel.
            println!("user {:?} followed {:?}", notif.user_name, notif.broadcaster_user_name);
            Ok(http::Response::builder().status(200).body(vec![])?)
        }
        _ => Ok(http::Response::builder().status(200).body(vec![])?),
    }
}

Re-exports

pub use event::websocket::*;

unsupported

Modules

channel

Subscription types regarding channels

event

EventSub events and their types

stream

Subscription types regarding streams

user

Subscription types regarding users

Structs

EventSubSubscription

General information about an EventSub subscription.

EventSubscriptionInformation

Metadata about the subscription.

Payload

Notification received

VerificationRequest

Verification Request

WebhookTransport

Webhook transport

WebhookTransportResponse

Webhook transport

WebsocketTransportunsupported

Websocket transport

WebsocketTransportResponseunsupported

Websocket transport

Enums

Event

A notification with an event payload. Enumerates all possible Payloads

EventType

Event Types

Message

Subscription message/payload. Received on events and other messages.

PayloadParseError

Errors that can happen when parsing payload

Status

Subscription request status

Transport

Transport setting for event notification

TransportMethod

Transport method

TransportResponse

Transport response on event notification

Traits

EventSubscription

An EventSub subscription.