twitch_api
Module eventsub

Available on crate feature eventsub only.

Holds serializable EventSub stuff

Use CreateEventSubSubscription to subscribe to an event according to the EventSub guide. Parse the response payload text with Event::parse_http or Event::parse_websocket

Examples

See examples/ for a more complete example of using eventsub.

Subscribe to a channel’s follow events:

use twitch_api::{
    eventsub::{channel::ChannelFollowV2, Transport, TransportMethod},
    helix,
};

let event = ChannelFollowV2::new("1234", "5678");
let transport = Transport::webhook(
    "https://example.org/eventsub/channelfollow",
    String::from("secretabcd"),
);

let event_information = client
    .create_eventsub_subscription(event, transport, &token)
    .await?;

println!("event id: {:?}", event_information.id);

You’ll now get a http POST request to the url you specified as the callback. You need to respond to this request from your webserver with a 200 OK response with the challenge as the body. After this, you’ll get notifications

use twitch_api::eventsub::{Event, Payload, Message};
pub fn parse_request(
    request: &http::Request<Vec<u8>>,
) -> Result<http::Response<Vec<u8>>, Box<dyn std::error::Error + Send + Sync + 'static>> {
    // First, we verify the response, assuring it's legit.
    if !Event::verify_payload(request, b"secretabcd") {
        return Err(todo!());
    }
    match Event::parse_http(request)? {
        Event::ChannelFollowV2(Payload {
            message: Message::VerificationRequest(ver),
            ..
        }) => {
            // We've verified the request, so we can respond to it with the challenge
            Ok(http::Response::builder()
                .status(200)
                .body(ver.challenge.into_bytes())?)
        },
        Event::ChannelFollowV2(Payload {
            message: Message::Notification(notif),
            ..
        }) => {
            // make sure you save the `Twitch-Eventsub-Message-Id` headers value,
            // twitch may resend notifications, and in those cases you should just return 200 OK.

            // Do whatever you need to do with the event. Preferably send the event to a channel.
            println!("user {:?} followed {:?}", notif.user_name, notif.broadcaster_user_name);
            Ok(http::Response::builder().status(200).body(vec![])?)
        }
        _ => Ok(http::Response::builder().status(200).body(vec![])?),
    }
}

Implemented Subscriptions

automod.* 🔴 0/4

Name	Subscription
Payload

automod.message.hold	-
-

automod.message.update	-
-

automod.settings.update	-
-

automod.terms.update	-
-

channel.* 🟡 42/65

Name	Subscription
Payload

channel.ad_break.begin	ChannelAdBreakBeginV1
ChannelAdBreakBeginV1Payload

channel.ban	ChannelBanV1
ChannelBanV1Payload

channel.channel_points_automatic_reward_redemption.add	-
-

channel.channel_points_custom_reward.add	ChannelPointsCustomRewardAddV1
ChannelPointsCustomRewardAddV1Payload

channel.channel_points_custom_reward.remove	ChannelPointsCustomRewardRemoveV1
ChannelPointsCustomRewardRemoveV1Payload

channel.channel_points_custom_reward.update	ChannelPointsCustomRewardUpdateV1
ChannelPointsCustomRewardUpdateV1Payload

channel.channel_points_custom_reward_redemption.add	ChannelPointsCustomRewardRedemptionAddV1
ChannelPointsCustomRewardRedemptionAddV1Payload

channel.channel_points_custom_reward_redemption.update	ChannelPointsCustomRewardRedemptionUpdateV1
ChannelPointsCustomRewardRedemptionUpdateV1Payload

channel.charity_campaign.donate	ChannelCharityCampaignDonateV1
ChannelCharityCampaignDonateV1Payload

channel.charity_campaign.progress	ChannelCharityCampaignProgressV1
ChannelCharityCampaignProgressV1Payload

channel.charity_campaign.start	ChannelCharityCampaignStartV1
ChannelCharityCampaignStartV1Payload

channel.charity_campaign.stop	ChannelCharityCampaignStopV1
ChannelCharityCampaignStopV1Payload

channel.chat.clear	ChannelChatClearV1
ChannelChatClearV1Payload

channel.chat.clear_user_messages	ChannelChatClearUserMessagesV1
ChannelChatClearUserMessagesV1Payload

channel.chat.message	ChannelChatMessageV1
ChannelChatMessageV1Payload

channel.chat.message_delete	ChannelChatMessageDeleteV1
ChannelChatMessageDeleteV1Payload

channel.chat.notification	ChannelChatNotificationV1
ChannelChatNotificationV1Payload

channel.chat.user_message_hold	-
-

channel.chat.user_message_update	-
-

channel.chat_settings.update	-
-

channel.cheer	ChannelCheerV1
ChannelCheerV1Payload

channel.follow	ChannelFollowV2
ChannelFollowV2Payload

channel.goal.begin	ChannelGoalBeginV1
ChannelGoalBeginV1Payload

channel.goal.end	ChannelGoalEndV1
ChannelGoalEndV1Payload

channel.goal.progress	ChannelGoalProgressV1
ChannelGoalProgressV1Payload

channel.guest_star_guest.update	-
-

channel.guest_star_session.begin	-
-

channel.guest_star_session.end	-
-

channel.guest_star_settings.update	-
-

channel.hype_train.begin	ChannelHypeTrainBeginV1
ChannelHypeTrainBeginV1Payload

channel.hype_train.end	ChannelHypeTrainEndV1
ChannelHypeTrainEndV1Payload

channel.hype_train.progress	ChannelHypeTrainProgressV1
ChannelHypeTrainProgressV1Payload

channel.moderate	-
-

channel.moderate	-
-

channel.moderator.add	-
-

channel.moderator.remove	-
-

channel.poll.begin	ChannelPollBeginV1
ChannelPollBeginV1Payload

channel.poll.end	ChannelPollEndV1
ChannelPollEndV1Payload

channel.poll.progress	ChannelPollProgressV1
ChannelPollProgressV1Payload

channel.prediction.begin	ChannelPredictionBeginV1
ChannelPredictionBeginV1Payload

channel.prediction.end	ChannelPredictionEndV1
ChannelPredictionEndV1Payload

channel.prediction.lock	ChannelPredictionLockV1
ChannelPredictionLockV1Payload

channel.prediction.progress	ChannelPredictionProgressV1
ChannelPredictionProgressV1Payload

channel.raid	ChannelRaidV1
ChannelRaidV1Payload

channel.shared_chat.begin	-
-

channel.shared_chat.end	-
-

channel.shared_chat.update	-
-

channel.shield_mode.begin	ChannelShieldModeBeginV1
ChannelShieldModeBeginV1Payload

channel.shield_mode.end	ChannelShieldModeEndV1
ChannelShieldModeEndV1Payload

channel.shoutout.create	ChannelShoutoutCreateV1
ChannelShoutoutCreateV1Payload

channel.shoutout.receive	ChannelShoutoutReceiveV1
ChannelShoutoutReceiveV1Payload

channel.subscribe	ChannelSubscribeV1
ChannelSubscribeV1Payload

channel.subscription.end	ChannelSubscriptionEndV1
ChannelSubscriptionEndV1Payload

channel.subscription.gift	ChannelSubscriptionGiftV1
ChannelSubscriptionGiftV1Payload

channel.subscription.message	ChannelSubscriptionMessageV1
ChannelSubscriptionMessageV1Payload

channel.suspicious_user.message	-
-

channel.suspicious_user.update	-
-

channel.unban	ChannelUnbanV1
ChannelUnbanV1Payload

channel.unban_request.create	-
-

channel.unban_request.resolve	-
-

channel.update	ChannelUpdateV2
ChannelUpdateV2Payload

channel.vip.add	-
-

channel.vip.remove	-
-

channel.warning.acknowledge	-
-

channel.warning.send	-
-

conduit.* 🟢 1/1

Name	Subscription
Payload

conduit.shard.disabled	ConduitShardDisabledV1
ConduitShardDisabledV1Payload

drop.* 🔴 0/1

Name	Subscription
Payload

drop.entitlement.grant	-
-

extension.* 🔴 0/1

Name	Subscription
Payload

extension.bits_transaction.create	-
-

stream.* 🟢 2/2

Name	Subscription
Payload

stream.offline	StreamOfflineV1
StreamOfflineV1Payload

stream.online	StreamOnlineV1
StreamOnlineV1Payload

user.* 🟡 3/4

Name	Subscription
Payload

user.authorization.grant	UserAuthorizationGrantV1
UserAuthorizationGrantV1Payload

user.authorization.revoke	UserAuthorizationRevokeV1
UserAuthorizationRevokeV1Payload

user.update	UserUpdateV1
UserUpdateV1Payload

user.whisper.message	-
-

Re-exports

pub use event::websocket::*;

Modules

channel

Subscription types regarding channels

conduit

Subscription types regarding conduits.

event

EventSub events and their types

stream

Subscription types regarding streams

user

Subscription types regarding users

Structs

Conduit

General information about a Conduit

ConduitTransport

Conduit transport

ConduitTransportResponse

Conduit transport

EventSubSubscription

General information about an EventSub subscription.

EventSubscriptionInformation

Metadata about the subscription.

Payload

Notification received

Shard

General information about a Shard

ShardError

A structured error that occurred with a shard

ShardResponse

A shard when described by Twitch

VerificationRequest

Verification Request

WebhookTransport

Webhook transport

WebhookTransportResponse

Webhook transport

WebsocketTransport

Websocket transport

WebsocketTransportResponse

Websocket transport

Enums

Event

A notification with an event payload. Enumerates all possible Payloads

EventType

Event Types

Message

Subscription message/payload. Received on events and other messages.

PayloadParseError

Errors that can happen when parsing payload

ShardStatus

The shard status.

Status

Subscription request status

Transport

Transport setting for event notification

TransportMethod

Transport method

TransportResponse

Transport response on event notification

Traits

EventSubscription

An EventSub subscription.