twitch_api::eventsub

Module channel

Available on crate feature eventsub only.

Subscription types regarding channels

Modules

ad_break

Ad break on channel has begun

ban

A viewer is banned from the specified channel.

channel_points_automatic_reward_redemption

A viewer has redeemed an automatic channel points reward in a specified channel.

channel_points_custom_reward

Custom channel points rewards on specific channel has been changed, removed or updated.

channel_points_custom_reward_redemption

A viewer has redeemed a custom channel points reward or a redemption of a channel points custom reward has been updated for the specified channel.

charity_campaign

Poll on a specific channel has been begun, ended or progressed.

chat

Chat events

chat_settings

A broadcaster’s chat settings are updated

cheer

A user cheers on the specified channel.

follow

A specified channel receives a follow.

goal

A broadcaster has started, progressed or ended a goal.

guest_star_guestbeta

Events regarding guests of guest star sessions

guest_star_sessionbeta

Events regarding guest star sessions

guest_star_settingsbeta

Events regarding settings of guest star sessions

hypetrain

A hype train has started, progressed or ended.

moderate

a moderator performs a moderation action in a channel.

moderator

A user’s moderator privileges on a specified channel are changed.

poll

Poll on a specific channel has been begun, ended or progressed.

prediction

Prediction on the specified channel begins, progresses, locks or ends.

raid

A a broadcaster raids another broadcaster’s channel.

shared_chat

Events related to shared chat

shield_mode

Shield mode on the specified channel begins or ends.

shoutout

Subscription for when a Shoutout has happened

subscribe

A specified channel receives a subscriber. This does not include resubscribes.

subscription

Subscription on a specified channel has changed

suspicious_user

A user’s moderator privileges on a specified channel are changed.

unban

A viewer is unbanned from the specified channel.

unban_request

An unban request in a specified channel is changed.

update

Channel has updated the category, title, mature flag, or broadcast language.

vip

A user’s VIP status on a specified channel is changed.

warning

Notifications for warnings in a channel.

Structs

ChannelAdBreakBeginV1

channel.ad_break.begin: a user runs a midroll commercial break, either manually or automatically via ads manager.

ChannelAdBreakBeginV1Payload

channel.ad_break.begin response payload.

ChannelBanV1

channel.ban: a viewer is banned from the specified channel.

ChannelBanV1Payload

channel.ban response payload.

ChannelCharityCampaignDonateV1

channel.charity_campaign.donate: a user donates to the broadcaster’s charity campaign.

ChannelCharityCampaignDonateV1Payload

channel.charity_campaign.donate response payload.

ChannelCharityCampaignProgressV1

channel.charity_campaign.progress: progress is made towards the campaign’s goal or when the broadcaster changes the fundraising goal.

ChannelCharityCampaignProgressV1Payload

channel.charity_campaign.progress response payload.

ChannelCharityCampaignStartV1

channel.charity_campaign.start: a broadcaster starts a charity campaign.

ChannelCharityCampaignStartV1Payload

channel.charity_campaign.start response payload.

ChannelCharityCampaignStopV1

channel.charity_campaign.stop: a broadcaster stops a charity campaign.

ChannelCharityCampaignStopV1Payload

channel.charity_campaign.stop response payload.

ChannelChatClearUserMessagesV1

channel.chat.clear_user_messages: a moderator or bot clears all messages for a specific user.

ChannelChatClearUserMessagesV1Payload

channel.chat.clear_user_messages response payload.

ChannelChatClearV1

channel.chat.clear: a moderator or bot clears all messages from the chat room.

ChannelChatClearV1Payload

channel.chat.clear response payload.

ChannelChatMessageDeleteV1

channel.chat.message_delete: a moderator removes a specific message.

ChannelChatMessageDeleteV1Payload

channel.chat.message_delete response payload.

ChannelChatMessageV1

channel.chat.message: a user sends a message to a specific chat room.

ChannelChatMessageV1Payload

channel.chat.message response payload.

ChannelChatNotificationV1

channel.chat.notification: an event that appears in chat occurs, such as someone subscribing to the channel or a subscription is gifted.

ChannelChatNotificationV1Payload

channel.chat.notification response payload.

ChannelChatSettingsUpdateV1

channel.chat_settings.update: a broadcaster’s chat settings are updated.

ChannelChatSettingsUpdateV1Payload

channel.chat_settings.update response payload.

ChannelChatUserMessageHoldV1

channel.chat.user_message_hold: a user’s message is caught by automod.

ChannelChatUserMessageHoldV1Payload

channel.chat.user_message_hold response payload.

ChannelChatUserMessageUpdateV1

channel.chat.user_message_update: a user’s message’s automod status is updated.

ChannelChatUserMessageUpdateV1Payload

channel.chat.user_message_update response payload.

ChannelCheerV1

channel.cheer: a user cheers on the specified channel.

ChannelCheerV1Payload

channel.cheer response payload.

ChannelFollowV1Deprecated

channel.follow v1: a specified channel receives a follow.

ChannelFollowV2

channel.follow v2: a specified channel receives a follow.

ChannelFollowV1Payload

channel.follow response payload.

ChannelFollowV2Payload

channel.follow response payload.

ChannelGoalBeginV1

channel.goal.begin: a specified broadcaster begins a goal.

ChannelGoalBeginV1Payload

channel.goal.begin response payload.

ChannelGoalEndV1

channel.goal.end: a specified broadcaster ends a goal.

ChannelGoalEndV1Payload

channel.goal.end response payload.

ChannelGoalProgressV1

channel.goal.progress: progress is made towards the specified broadcaster’s goal.

ChannelGoalProgressV1Payload

channel.goal.progress response payload.

ChannelGuestStarGuestUpdateBetabeta

channel.guest_star_guest.update: the host preferences for Guest Star have been updated.

ChannelGuestStarGuestUpdateBetaPayloadbeta

channel.guest_star_guest.update response payload.

ChannelGuestStarSessionBeginBetabeta

channel.guest_star_session.begin: the host begins a new Guest Star session.

ChannelGuestStarSessionBeginBetaPayloadbeta

channel.guest_star_session.begin response payload.

ChannelGuestStarSessionEndBetabeta

channel.guest_star_session.end: a running Guest Star session is ended by the host, or automatically by the system.

ChannelGuestStarSessionEndBetaPayloadbeta

channel.guest_star_session.end response payload.

ChannelGuestStarSettingsUpdateBetabeta

channel.guest_star_settings.update: the host preferences for Guest Star have been updated.

ChannelGuestStarSettingsUpdateBetaPayloadbeta

channel.guest_star_settings.update response payload.

ChannelHypeTrainBeginV1

channel.hype_train.begin: a hype train begins on the specified channel.

ChannelHypeTrainBeginV1Payload

channel.hype_train.begin response payload.

ChannelHypeTrainEndV1

channel.hype_train.end: a hype train ends on the specified channel.

ChannelHypeTrainEndV1Payload

channel.hype_train.end response payload.

ChannelHypeTrainProgressV1

channel.hype_train.progress: a hype train makes progress on the specified channel.

ChannelHypeTrainProgressV1Payload

channel.hype_train.progress response payload.

ChannelModerateV1

channel.moderate: a moderator performs a moderation action in a channel.

ChannelModerateV2

channel.moderate: a moderator performs a moderation action in a channel.

ChannelModerateV1Payload

channel.moderate response payload.

ChannelModerateV2Payload

channel.moderate response payload.

ChannelModeratorAddV1

channel.moderator.add: a user is given moderator privileges on a specified channel.

ChannelModeratorAddV1Payload

channel.moderator.add response payload.

ChannelModeratorRemoveV1

channel.moderator.remove: a user has moderator privileges removed on a specified channel.

ChannelModeratorRemoveV1Payload

channel.moderator.remove response payload.

ChannelPointsAutomaticRewardRedemptionAddV1

channel.channel_points_automatic_reward_redemption.add:a viewer has redeemed an automatic channel points reward on the specified channel.

ChannelPointsAutomaticRewardRedemptionAddV1Payload

channel.channel_points_automatic_reward_redemption.add response payload.

ChannelPointsCustomRewardAddV1

channel.channel_points_custom_reward.add: a custom channel points reward has been created for the specified channel.

ChannelPointsCustomRewardAddV1Payload

channel.channel_points_custom_reward.add response payload.

ChannelPointsCustomRewardRedemptionAddV1

channel.channel_points_custom_reward_redemption.add: a viewer has redeemed a custom channel points reward on the specified channel.

ChannelPointsCustomRewardRedemptionAddV1Payload

channel.channel_points_custom_reward_redemption.add response payload.

ChannelPointsCustomRewardRedemptionUpdateV1

channel.channel_points_custom_reward_redemption.update: a redemption of a channel points custom reward has been updated for the specified channel.

ChannelPointsCustomRewardRedemptionUpdateV1Payload

channel.channel_points_custom_reward_redemption.update response payload.

ChannelPointsCustomRewardRemoveV1

channel.channel_points_custom_reward.remove: a custom channel points reward has been removed from the specified channel.

ChannelPointsCustomRewardRemoveV1Payload

channel.channel_points_custom_reward.remove response payload.

ChannelPointsCustomRewardUpdateV1

channel.channel_points_custom_reward.update: a custom channel points reward has been updated for the specified channel.

ChannelPointsCustomRewardUpdateV1Payload

channel.channel_points_custom_reward.update response payload.

ChannelPollBeginV1

channel.poll.begin: a poll begins on the specified channel.

ChannelPollBeginV1Payload

channel.poll.begin response payload.

ChannelPollEndV1

channel.poll.end: a poll ends on the specified channel.

ChannelPollEndV1Payload

channel.poll.end response payload.

ChannelPollProgressV1

channel.poll.progress: an user responds to a poll on the specified channel

ChannelPollProgressV1Payload

channel.poll.progress response payload.

ChannelPredictionBeginV1

channel.prediction.begin: a Prediction begins on the specified channel

ChannelPredictionBeginV1Payload

channel.prediction.begin response payload.

ChannelPredictionEndV1

channel.prediction.end: a Prediction ends on the specified channel.

ChannelPredictionEndV1Payload

channel.prediction.end response payload.

ChannelPredictionLockV1

channel.prediction.lock: an user responds to a prediction on the specified channel

ChannelPredictionLockV1Payload

channel.prediction.lock response payload.

ChannelPredictionProgressV1

channel.prediction.progress: an user responds to a prediction on the specified channel

ChannelPredictionProgressV1Payload

channel.prediction.progress response payload.

ChannelRaidV1

channel.raid: a a broadcaster raids another broadcaster’s channel.

ChannelRaidV1Payload

channel.raid response payload.

ChannelSharedChatBeginV1

channel.shared_chat.begin: a channel becomes active in an active shared chat session.

ChannelSharedChatBeginV1Payload

channel.shared_chat.begin response payload.

ChannelSharedChatEndV1

channel.shared_chat.end: a channel leaves a shared chat session or the session ends.

ChannelSharedChatEndV1Payload

channel.shared_chat.end response payload.

ChannelSharedChatUpdateV1

channel.shared_chat.update: the active shared chat session the channel is in changed.

ChannelSharedChatUpdateV1Payload

channel.shared_chat.update response payload.

ChannelShieldModeBeginV1

channel.shield_mode.begin: an user responds to a prediction on the specified channel

ChannelShieldModeBeginV1Payload

channel.shield_mode.begin response payload.

ChannelShieldModeEndV1

channel.shield_mode.end: an user responds to a prediction on the specified channel

ChannelShieldModeEndV1Payload

channel.shield_mode.end response payload.

ChannelShoutoutCreateV1

channel.shoutout.create: a Prediction begins on the specified channel

ChannelShoutoutCreateV1Payload

channel.shoutout.create response payload.

ChannelShoutoutReceiveV1

channel.shoutout.receive: a Prediction begins on the specified channel

ChannelShoutoutReceiveV1Payload

channel.shoutout.receive response payload.

ChannelSubscribeV1

channel.subscribe: a specified channel receives a subscriber. This does not include resubscribes.

ChannelSubscribeV1Payload

channel.subscribe response payload.

ChannelSubscriptionEndV1

channel.subscription.end: a subscription to the specified channel expires.

ChannelSubscriptionEndV1Payload

channel.subscription.end response payload.

ChannelSubscriptionGiftV1

channel.subscription.gift: a subscription to the specified channel expires.

ChannelSubscriptionGiftV1Payload

channel.subscription.gift response payload.

ChannelSubscriptionMessageV1

channel.subscription.message: a subscription to the specified channel expires.

ChannelSubscriptionMessageV1Payload

channel.subscription.message response payload.

ChannelSuspiciousUserMessageV1

channel.suspicious_user.message: a chat message has been sent from a suspicious user.

ChannelSuspiciousUserMessageV1Payload

channel.suspicious_user.message response payload.

ChannelSuspiciousUserUpdateV1

channel.suspicious_user.update: a suspicious user has been updated.

ChannelSuspiciousUserUpdateV1Payload

channel.suspicious_user.update response payload.

ChannelUnbanRequestCreateV1

channel.unban_request.create: a user creates an unban request.

ChannelUnbanRequestCreateV1Payload

channel.unban_request.create response payload.

ChannelUnbanRequestResolveV1

channel.unban_request.resolve: an unban request has been resolved.

ChannelUnbanRequestResolveV1Payload

channel.unban_request.resolve response payload.

ChannelUnbanV1

channel.unban: a viewer is unbanned from the specified channel.

ChannelUnbanV1Payload

channel.unban response payload.

ChannelUpdateV1Deprecated

version 1 of channel.update subscription type sends notifications when a broadcaster updates the category, title, mature flag, or broadcast language for their channel.

ChannelUpdateV2

channel.update subscription type sends notifications when a broadcaster updates the category, title, mature flag, or broadcast language for their channel.

ChannelUpdateV1PayloadDeprecated

channel.update response payload.

ChannelUpdateV2Payload

channel.update response payload.

ChannelVipAddV1

channel.vip.add: a VIP is added to the channel.

ChannelVipAddV1Payload

channel.vip.add response payload.

ChannelVipRemoveV1

channel.vip.remove: a user has vip privileges removed on a specified channel.

ChannelVipRemoveV1Payload

channel.vip.remove response payload.

ChannelWarningAcknowledgeV1

channel.warning.acknowledge: a warning is acknowledged by a user.

ChannelWarningAcknowledgeV1Payload

channel.warning.acknowledge response payload.

ChannelWarningSendV1

channel.warning.send: a warning is sent to a user.

ChannelWarningSendV1Payload

channel.warning.send response payload.