twitch_api::eventsub::channel
Module chat

Available on crate feature eventsub only.

Chat events

Modules

clear

A moderator or bot clears all messages from the chat room.

clear_user_messages

A moderator or bot clears all messages for a specific user.

message

Any user sends a message to a specific chat room.

message_delete

A moderator removes a specific message.

notification

An event that appears in chat occurs, such as someone subscribing to the channel or a subscription is gifted.

Structs

Badge

A badge

ChannelChatClearUserMessagesV1

channel.chat.clear_user_messages: a moderator or bot clears all messages for a specific user.

ChannelChatClearUserMessagesV1Payload

channel.chat.clear_user_messages response payload.

ChannelChatClearV1

channel.chat.clear: a moderator or bot clears all messages from the chat room.

ChannelChatClearV1Payload

channel.chat.clear response payload.

ChannelChatMessageDeleteV1

channel.chat.message_delete: a moderator removes a specific message.

ChannelChatMessageDeleteV1Payload

channel.chat.message_delete response payload.

ChannelChatMessageV1

channel.chat.message: a user sends a message to a specific chat room.

ChannelChatMessageV1Payload

channel.chat.message response payload.

ChannelChatNotificationV1

channel.chat.notification: an event that appears in chat occurs, such as someone subscribing to the channel or a subscription is gifted.

ChannelChatNotificationV1Payload

channel.chat.notification response payload.

Cheermote

A cheermote fragment

Emote

An emote fragment

Mention

A user mention fragment

Message

A message

Enums

Fragment

A chat message fragment