twitch_api::eventsub::channel::chat
Module notification

Available on crate feature eventsub only.

An event that appears in chat occurs, such as someone subscribing to the channel or a subscription is gifted.

Structs

Announcement

A announcement notification

BitsBadgeTier

A bits badge tier upgrade notification

ChannelChatNotificationV1

channel.chat.notification: an event that appears in chat occurs, such as someone subscribing to the channel or a subscription is gifted.

ChannelChatNotificationV1Payload

channel.chat.notification response payload.

CharityDonation

A charity donation notification

CommunitySubGift

A gift notification for multiple gifted subscriptions. Followed by CommunitySubGift::total amount of SubGifts.

GiftPaidUpgrade

A gift notification for a paid upgrade of a previously gifted subscription.

PayItForward

A pay it forward notification

PrimePaidUpgrade

A notification for a paid upgrade of a previous Twitch Prime channel subscription.

Raid

A raid notification

Resubscription

A resubcription notification

SubGift

A subscription gift notification

Subscription

A subscription notification

Unraid

A unraid notification

Enums

Chatter

Information about the user that triggered this notification

Gifter

A gifter

Notification

All possible notifications in ChannelChatNotificationV1Payload