twitch_api::eventsub
Enum EventType

#[non_exhaustive]
pub enum EventType {

    ChannelAdBreakBegin,
    ChannelChatClear,
    ChannelChatClearUserMessages,
    ChannelChatMessage,
    ChannelChatMessageDelete,
    ChannelChatNotification,
    ChannelCharityCampaignDonate,
    ChannelCharityCampaignProgress,
    ChannelCharityCampaignStart,
    ChannelCharityCampaignStop,
    ChannelUpdate,
    ChannelFollow,
    ChannelSubscribe,
    ChannelCheer,
    ChannelBan,
    ChannelUnban,
    ChannelPointsCustomRewardAdd,
    ChannelPointsCustomRewardUpdate,
    ChannelPointsCustomRewardRemove,
    ChannelPointsCustomRewardRedemptionAdd,
    ChannelPointsCustomRewardRedemptionUpdate,
    ChannelPollBegin,
    ChannelPollProgress,
    ChannelPollEnd,
    ChannelPredictionBegin,
    ChannelPredictionProgress,
    ChannelPredictionLock,
    ChannelPredictionEnd,
    ChannelShoutoutCreate,
    ChannelShoutoutReceive,
    ChannelRaid,
    ChannelSubscriptionEnd,
    ChannelSubscriptionGift,
    ChannelSubscriptionMessage,
    ChannelShieldModeBegin,
    ChannelShieldModeEnd,
    ChannelGoalBegin,
    ChannelGoalProgress,
    ChannelGoalEnd,
    ChannelHypeTrainBegin,
    ChannelHypeTrainProgress,
    ChannelHypeTrainEnd,
    ConduitShardDisabled,
    StreamOnline,
    StreamOffline,
    UserUpdate,
    UserAuthorizationRevoke,
    UserAuthorizationGrant,

}

Available on crate feature eventsub only.

Event Types

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ChannelAdBreakBegin

channel.ad_break.begin: a user runs a midroll commercial break, either manually or automatically via ads manager.

ChannelChatClear

channel.chat.clear: a moderator or bot clears all messages from the chat room.

ChannelChatClearUserMessages

channel.chat.clear_user_messages: a moderator or bot clears all messages for a specific user.

ChannelChatMessage

channel.chat.message: any user sends a message to a specific chat room.

ChannelChatMessageDelete

channel.chat.message_delete: a moderator removes a specific message.

ChannelChatNotification

channel.chat.notification: an event that appears in chat occurs, such as someone subscribing to the channel or a subscription is gifted.

ChannelCharityCampaignDonate

channel.charity_campaign.donate: a user donates to the broadcaster’s charity campaign.

ChannelCharityCampaignProgress

channel.charity_campaign.progress: progress is made towards the campaign’s goal or when the broadcaster changes the fundraising goal.

ChannelCharityCampaignStart

channel.charity_campaign.start: a broadcaster starts a charity campaign.

ChannelCharityCampaignStop

channel.charity_campaign.stop: a broadcaster stops a charity campaign.

ChannelUpdate

channel.update: subscription type sends notifications when a broadcaster updates the category, title, mature flag, or broadcast language for their channel.

ChannelFollow

channel.follow: a specified channel receives a follow.

ChannelSubscribe

channel.subscribe: a specified channel receives a subscriber. This does not include resubscribes.

ChannelCheer

channel.cheer: a user cheers on the specified channel.

ChannelBan

channel.ban: a viewer is banned from the specified channel.

ChannelUnban

channel.unban: a viewer is unbanned from the specified channel.

ChannelPointsCustomRewardAdd

channel.channel_points_custom_reward.add: a custom channel points reward has been created for the specified channel.

ChannelPointsCustomRewardUpdate

channel.channel_points_custom_reward.update: a custom channel points reward has been updated for the specified channel.

ChannelPointsCustomRewardRemove

channel.channel_points_custom_reward.remove: a custom channel points reward has been removed from the specified channel.

ChannelPointsCustomRewardRedemptionAdd

channel.channel_points_custom_reward_redemption.add: a viewer has redeemed a custom channel points reward on the specified channel.

ChannelPointsCustomRewardRedemptionUpdate

channel.channel_points_custom_reward_redemption.update: a redemption of a channel points custom reward has been updated for the specified channel.

ChannelPollBegin

channel.poll.begin: a poll begins on the specified channel.

ChannelPollProgress

channel.poll.progress: a user responds to a poll on the specified channel.

ChannelPollEnd

channel.poll.end: a poll ends on the specified channel.

ChannelPredictionBegin

channel.prediction.begin: a Prediction begins on the specified channel

ChannelPredictionProgress

channel.prediction.progress: a user participates in a Prediction on the specified channel.

ChannelPredictionLock

channel.prediction.lock: a Prediction is locked on the specified channel.

ChannelPredictionEnd

channel.prediction.end: a Prediction ends on the specified channel.

ChannelShoutoutCreate

channel.shoutout.create: a specified broadcaster sends a Shoutout.

ChannelShoutoutReceive

channel.shoutout.receive: a specified broadcaster receives a Shoutout.

ChannelRaid

channel.raid: a broadcaster raids another broadcaster’s channel.

ChannelSubscriptionEnd

channel.subscription.end: a subscription to the specified channel expires.

ChannelSubscriptionGift

channel.subscription.gift: a user gives one or more gifted subscriptions in a channel.

ChannelSubscriptionMessage

channel.subscription.message: a user sends a resubscription chat message in a specific channel

ChannelShieldModeBegin

channel.shield_mode.begin: a channel activates shield mode

ChannelShieldModeEnd

channel.shield_mode.end: a channel deactivates shield mode

ChannelGoalBegin

channel.goal.begin: a goal begins on the specified channel.

ChannelGoalProgress

channel.goal.progress: a goal makes progress on the specified channel.

ChannelGoalEnd

channel.goal.end: a goal ends on the specified channel.

ChannelHypeTrainBegin

channel.hype_train.begin: a hype train begins on the specified channel.

ChannelHypeTrainProgress

channel.hype_train.progress: a hype train makes progress on the specified channel.

ChannelHypeTrainEnd

channel.hype_train.end: a hype train ends on the specified channel.

ConduitShardDisabled

conduit.shard.disabled: sends a notification when eventsub disables a shard due to the status of the underlying transport changing.

StreamOnline

stream.online: the specified broadcaster starts a stream.

StreamOffline

stream.offline: the specified broadcaster stops a stream.

UserUpdate

user.update: user updates their account.

UserAuthorizationRevoke

user.authorization.revoke: a user has revoked authorization for your client id. Use this webhook to meet government requirements for handling user data, such as GDPR, LGPD, or CCPA.

UserAuthorizationGrant

user.authorization.grant: a user’s authorization has been granted to your client id.

Implementations

impl EventType

pub const fn to_str(&self) -> &'static str

Get the event string of this event.

fn main() {
    assert_eq!(EventType::ChannelUpdate.to_str(), "channel.update");
    assert_eq!(EventType::ChannelUnban.to_str(), "channel.unban");
}