twitch_api::eventsub::event

Enum Event

#[non_exhaustive]
pub enum Event {

    AutomodMessageHoldV1(Payload<AutomodMessageHoldV1>),
    AutomodMessageHoldV2(Payload<AutomodMessageHoldV2>),
    AutomodMessageUpdateV1(Payload<AutomodMessageUpdateV1>),
    AutomodMessageUpdateV2(Payload<AutomodMessageUpdateV2>),
    AutomodSettingsUpdateV1(Payload<AutomodSettingsUpdateV1>),
    AutomodTermsUpdateV1(Payload<AutomodTermsUpdateV1>),
    ChannelAdBreakBeginV1(Payload<ChannelAdBreakBeginV1>),
    ChannelChatClearV1(Payload<ChannelChatClearV1>),
    ChannelChatClearUserMessagesV1(Payload<ChannelChatClearUserMessagesV1>),
    ChannelChatMessageV1(Payload<ChannelChatMessageV1>),
    ChannelChatMessageDeleteV1(Payload<ChannelChatMessageDeleteV1>),
    ChannelChatNotificationV1(Payload<ChannelChatNotificationV1>),
    ChannelChatUserMessageHoldV1(Payload<ChannelChatUserMessageHoldV1>),
    ChannelChatUserMessageUpdateV1(Payload<ChannelChatUserMessageUpdateV1>),
    ChannelChatSettingsUpdateV1(Payload<ChannelChatSettingsUpdateV1>),
    ChannelCharityCampaignDonateV1(Payload<ChannelCharityCampaignDonateV1>),
    ChannelCharityCampaignProgressV1(Payload<ChannelCharityCampaignProgressV1>),
    ChannelCharityCampaignStartV1(Payload<ChannelCharityCampaignStartV1>),
    ChannelCharityCampaignStopV1(Payload<ChannelCharityCampaignStopV1>),
    ChannelUpdateV1(Payload<ChannelUpdateV1>),
    ChannelUpdateV2(Payload<ChannelUpdateV2>),
    ChannelFollowV1(Payload<ChannelFollowV1>),
    ChannelFollowV2(Payload<ChannelFollowV2>),
    ChannelSubscribeV1(Payload<ChannelSubscribeV1>),
    ChannelCheerV1(Payload<ChannelCheerV1>),
    ChannelBanV1(Payload<ChannelBanV1>),
    ChannelUnbanV1(Payload<ChannelUnbanV1>),
    ChannelUnbanRequestCreateV1(Payload<ChannelUnbanRequestCreateV1>),
    ChannelUnbanRequestResolveV1(Payload<ChannelUnbanRequestResolveV1>),
    ChannelVipAddV1(Payload<ChannelVipAddV1>),
    ChannelVipRemoveV1(Payload<ChannelVipRemoveV1>),
    ChannelWarningAcknowledgeV1(Payload<ChannelWarningAcknowledgeV1>),
    ChannelWarningSendV1(Payload<ChannelWarningSendV1>),
    ChannelPointsAutomaticRewardRedemptionAddV1(Payload<ChannelPointsAutomaticRewardRedemptionAddV1>),
    ChannelPointsCustomRewardAddV1(Payload<ChannelPointsCustomRewardAddV1>),
    ChannelPointsCustomRewardUpdateV1(Payload<ChannelPointsCustomRewardUpdateV1>),
    ChannelPointsCustomRewardRemoveV1(Payload<ChannelPointsCustomRewardRemoveV1>),
    ChannelPointsCustomRewardRedemptionAddV1(Payload<ChannelPointsCustomRewardRedemptionAddV1>),
    ChannelPointsCustomRewardRedemptionUpdateV1(Payload<ChannelPointsCustomRewardRedemptionUpdateV1>),
    ChannelPollBeginV1(Payload<ChannelPollBeginV1>),
    ChannelPollProgressV1(Payload<ChannelPollProgressV1>),
    ChannelPollEndV1(Payload<ChannelPollEndV1>),
    ChannelPredictionBeginV1(Payload<ChannelPredictionBeginV1>),
    ChannelPredictionProgressV1(Payload<ChannelPredictionProgressV1>),
    ChannelPredictionLockV1(Payload<ChannelPredictionLockV1>),
    ChannelPredictionEndV1(Payload<ChannelPredictionEndV1>),
    ChannelRaidV1(Payload<ChannelRaidV1>),
    ChannelSharedChatBeginV1(Payload<ChannelSharedChatBeginV1>),
    ChannelSharedChatEndV1(Payload<ChannelSharedChatEndV1>),
    ChannelSharedChatUpdateV1(Payload<ChannelSharedChatUpdateV1>),
    ChannelShieldModeBeginV1(Payload<ChannelShieldModeBeginV1>),
    ChannelShieldModeEndV1(Payload<ChannelShieldModeEndV1>),
    ChannelShoutoutCreateV1(Payload<ChannelShoutoutCreateV1>),
    ChannelShoutoutReceiveV1(Payload<ChannelShoutoutReceiveV1>),
    ChannelSuspiciousUserMessageV1(Payload<ChannelSuspiciousUserMessageV1>),
    ChannelSuspiciousUserUpdateV1(Payload<ChannelSuspiciousUserUpdateV1>),
    ChannelGoalBeginV1(Payload<ChannelGoalBeginV1>),
    ChannelGoalProgressV1(Payload<ChannelGoalProgressV1>),
    ChannelGoalEndV1(Payload<ChannelGoalEndV1>),
    ChannelGuestStarSessionBeginBeta(Payload<ChannelGuestStarSessionBeginBeta>),
    ChannelGuestStarSessionEndBeta(Payload<ChannelGuestStarSessionEndBeta>),
    ChannelGuestStarSettingsUpdateBeta(Payload<ChannelGuestStarSettingsUpdateBeta>),
    ChannelGuestStarGuestUpdateBeta(Payload<ChannelGuestStarGuestUpdateBeta>),
    ChannelHypeTrainBeginV1(Payload<ChannelHypeTrainBeginV1>),
    ChannelHypeTrainProgressV1(Payload<ChannelHypeTrainProgressV1>),
    ChannelHypeTrainEndV1(Payload<ChannelHypeTrainEndV1>),
    ChannelModerateV1(Payload<ChannelModerateV1>),
    ChannelModerateV2(Payload<ChannelModerateV2>),
    ChannelModeratorAddV1(Payload<ChannelModeratorAddV1>),
    ChannelModeratorRemoveV1(Payload<ChannelModeratorRemoveV1>),
    ConduitShardDisabledV1(Payload<ConduitShardDisabledV1>),
    StreamOnlineV1(Payload<StreamOnlineV1>),
    StreamOfflineV1(Payload<StreamOfflineV1>),
    UserUpdateV1(Payload<UserUpdateV1>),
    UserAuthorizationGrantV1(Payload<UserAuthorizationGrantV1>),
    UserAuthorizationRevokeV1(Payload<UserAuthorizationRevokeV1>),
    UserWhisperMessageV1(Payload<UserWhisperMessageV1>),
    ChannelSubscriptionEndV1(Payload<ChannelSubscriptionEndV1>),
    ChannelSubscriptionGiftV1(Payload<ChannelSubscriptionGiftV1>),
    ChannelSubscriptionMessageV1(Payload<ChannelSubscriptionMessageV1>),

}

Available on crate feature eventsub only.

A notification with an event payload. Enumerates all possible Payloads

Parse with Event::parse or parse the whole http request your server receives with Payload::parse_http

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AutomodMessageHoldV1(Payload<AutomodMessageHoldV1>)

Automod Message Hold V1 Event

AutomodMessageHoldV2(Payload<AutomodMessageHoldV2>)

Automod Message Hold V2 Event

AutomodMessageUpdateV1(Payload<AutomodMessageUpdateV1>)

Automod Message Update V1 Event

AutomodMessageUpdateV2(Payload<AutomodMessageUpdateV2>)

Automod Message Update V2 Event

AutomodSettingsUpdateV1(Payload<AutomodSettingsUpdateV1>)

Automod Settings Update V1 Event

AutomodTermsUpdateV1(Payload<AutomodTermsUpdateV1>)

Automod Terms Update V1 Event

ChannelAdBreakBeginV1(Payload<ChannelAdBreakBeginV1>)

Channel Ad Break Begin V1 Event

ChannelChatClearV1(Payload<ChannelChatClearV1>)

Channel Chat Clear V1 Event

ChannelChatClearUserMessagesV1(Payload<ChannelChatClearUserMessagesV1>)

Channel Chat ClearUserMessages V1 Event

ChannelChatMessageV1(Payload<ChannelChatMessageV1>)

Channel Chat Message V1 Event

ChannelChatMessageDeleteV1(Payload<ChannelChatMessageDeleteV1>)

Channel Chat MessageDelete V1 Event

ChannelChatNotificationV1(Payload<ChannelChatNotificationV1>)

Channel Chat Notification V1 Event

ChannelChatUserMessageHoldV1(Payload<ChannelChatUserMessageHoldV1>)

Channel Chat UserMessageHold V1 Event

ChannelChatUserMessageUpdateV1(Payload<ChannelChatUserMessageUpdateV1>)

Channel Chat UserMessageUpdate V1 Event

ChannelChatSettingsUpdateV1(Payload<ChannelChatSettingsUpdateV1>)

Channel ChatSettings Update V1 Event

ChannelCharityCampaignDonateV1(Payload<ChannelCharityCampaignDonateV1>)

Channel Charity Campaign Donate V1 Event

ChannelCharityCampaignProgressV1(Payload<ChannelCharityCampaignProgressV1>)

Channel Charity Campaign Progress V1 Event

ChannelCharityCampaignStartV1(Payload<ChannelCharityCampaignStartV1>)

Channel Charity Campaign Start V1 Event

ChannelCharityCampaignStopV1(Payload<ChannelCharityCampaignStopV1>)

Channel Charity Campaign Stop V1 Event

ChannelUpdateV1(Payload<ChannelUpdateV1>)

👎Deprecated: use Event::ChannelUpdateV2 instead

Channel Update V1 Event

ChannelUpdateV2(Payload<ChannelUpdateV2>)

Channel Update V2 Event

ChannelFollowV1(Payload<ChannelFollowV1>)

👎Deprecated: use Event::ChannelFollowV2 instead

Channel Follow V1 Event

ChannelFollowV2(Payload<ChannelFollowV2>)

Channel Follow V2 Event

ChannelSubscribeV1(Payload<ChannelSubscribeV1>)

Channel Subscribe V1 Event

ChannelCheerV1(Payload<ChannelCheerV1>)

Channel Cheer V1 Event

ChannelBanV1(Payload<ChannelBanV1>)

Channel Ban V1 Event

ChannelUnbanV1(Payload<ChannelUnbanV1>)

Channel Unban V1 Event

ChannelUnbanRequestCreateV1(Payload<ChannelUnbanRequestCreateV1>)

Channel UnbanRequest Create V1 Event

ChannelUnbanRequestResolveV1(Payload<ChannelUnbanRequestResolveV1>)

Channel UnbanRequest Resolve V1 Event

ChannelVipAddV1(Payload<ChannelVipAddV1>)

Channel VIP Add V1 Event

ChannelVipRemoveV1(Payload<ChannelVipRemoveV1>)

Channel VIP Remove V1 Event

ChannelWarningAcknowledgeV1(Payload<ChannelWarningAcknowledgeV1>)

Channel Warning Acknowledge V1 Event

ChannelWarningSendV1(Payload<ChannelWarningSendV1>)

Channel Warning Send V1 Event

ChannelPointsAutomaticRewardRedemptionAddV1(Payload<ChannelPointsAutomaticRewardRedemptionAddV1>)

Channel Points Automatic Reward Redemption Add V1 Event

ChannelPointsCustomRewardAddV1(Payload<ChannelPointsCustomRewardAddV1>)

Channel Points Custom Reward Add V1 Event

ChannelPointsCustomRewardUpdateV1(Payload<ChannelPointsCustomRewardUpdateV1>)

Channel Points Custom Reward Update V1 Event

ChannelPointsCustomRewardRemoveV1(Payload<ChannelPointsCustomRewardRemoveV1>)

Channel Points Custom Reward Remove V1 Event

ChannelPointsCustomRewardRedemptionAddV1(Payload<ChannelPointsCustomRewardRedemptionAddV1>)

Channel Points Custom Reward Redemption Add V1 Event

ChannelPointsCustomRewardRedemptionUpdateV1(Payload<ChannelPointsCustomRewardRedemptionUpdateV1>)

Channel Points Custom Reward Redemption Update V1 Event

ChannelPollBeginV1(Payload<ChannelPollBeginV1>)

Channel Poll Begin V1 Event

ChannelPollProgressV1(Payload<ChannelPollProgressV1>)

Channel Poll Progress V1 Event

ChannelPollEndV1(Payload<ChannelPollEndV1>)

Channel Poll End V1 Event

ChannelPredictionBeginV1(Payload<ChannelPredictionBeginV1>)

Channel Prediction Begin V1 Event

ChannelPredictionProgressV1(Payload<ChannelPredictionProgressV1>)

Channel Prediction Progress V1 Event

ChannelPredictionLockV1(Payload<ChannelPredictionLockV1>)

Channel Prediction Lock V1 Event

ChannelPredictionEndV1(Payload<ChannelPredictionEndV1>)

Channel Prediction End V1 Event

ChannelRaidV1(Payload<ChannelRaidV1>)

Channel Raid V1 Event

ChannelSharedChatBeginV1(Payload<ChannelSharedChatBeginV1>)

Channel SharedChat Begin V1 Event

ChannelSharedChatEndV1(Payload<ChannelSharedChatEndV1>)

Channel SharedChat End V1 Event

ChannelSharedChatUpdateV1(Payload<ChannelSharedChatUpdateV1>)

Channel SharedChat Update V1 Event

ChannelShieldModeBeginV1(Payload<ChannelShieldModeBeginV1>)

Channel ShieldMode Begin V1 Event

ChannelShieldModeEndV1(Payload<ChannelShieldModeEndV1>)

Channel ShieldMode End V1 Event

ChannelShoutoutCreateV1(Payload<ChannelShoutoutCreateV1>)

Channel Shoutout Create V1 Event

ChannelShoutoutReceiveV1(Payload<ChannelShoutoutReceiveV1>)

Channel Shoutout Receive V1 Event

ChannelSuspiciousUserMessageV1(Payload<ChannelSuspiciousUserMessageV1>)

Channel SuspicousUser Message V1 Event

ChannelSuspiciousUserUpdateV1(Payload<ChannelSuspiciousUserUpdateV1>)

Channel SuspicousUser Update V1 Event

ChannelGoalBeginV1(Payload<ChannelGoalBeginV1>)

Channel Goal Begin V1 Event

ChannelGoalProgressV1(Payload<ChannelGoalProgressV1>)

Channel Goal Progress V1 Event

ChannelGoalEndV1(Payload<ChannelGoalEndV1>)

Channel Goal End V1 Event

ChannelGuestStarSessionBeginBeta(Payload<ChannelGuestStarSessionBeginBeta>)

Available on crate feature beta only.

Channel GuestStarSession Begin V1 Event

ChannelGuestStarSessionEndBeta(Payload<ChannelGuestStarSessionEndBeta>)

Available on crate feature beta only.

Channel GuestStarSession End V1 Event

ChannelGuestStarSettingsUpdateBeta(Payload<ChannelGuestStarSettingsUpdateBeta>)

Available on crate feature beta only.

Channel GuestStarSettings Update V1 Event

ChannelGuestStarGuestUpdateBeta(Payload<ChannelGuestStarGuestUpdateBeta>)

Available on crate feature beta only.

Channel GuestStarGuest Update V1 Event

ChannelHypeTrainBeginV1(Payload<ChannelHypeTrainBeginV1>)

Channel Hype Train Begin V1 Event

ChannelHypeTrainProgressV1(Payload<ChannelHypeTrainProgressV1>)

Channel Hype Train Progress V1 Event

ChannelHypeTrainEndV1(Payload<ChannelHypeTrainEndV1>)

Channel Hype Train End V1 Event

ChannelModerateV1(Payload<ChannelModerateV1>)

Channel Moderate V1 Event

ChannelModerateV2(Payload<ChannelModerateV2>)

Channel Moderate V2 Event

ChannelModeratorAddV1(Payload<ChannelModeratorAddV1>)

Channel Moderator Add V1 Event

ChannelModeratorRemoveV1(Payload<ChannelModeratorRemoveV1>)

Channel Moderator Remove V1 Event

ConduitShardDisabledV1(Payload<ConduitShardDisabledV1>)

Conduit Shard Disabled V1 Event

StreamOnlineV1(Payload<StreamOnlineV1>)

StreamOnline V1 Event

StreamOfflineV1(Payload<StreamOfflineV1>)

StreamOffline V1 Event

UserUpdateV1(Payload<UserUpdateV1>)

User Update V1 Event

UserAuthorizationGrantV1(Payload<UserAuthorizationGrantV1>)

User Authorization Grant V1 Event

UserAuthorizationRevokeV1(Payload<UserAuthorizationRevokeV1>)

User Authorization Revoke V1 Event

UserWhisperMessageV1(Payload<UserWhisperMessageV1>)

User Whisper Message V1 Event

ChannelSubscriptionEndV1(Payload<ChannelSubscriptionEndV1>)

Channel Subscription End V1 Event

ChannelSubscriptionGiftV1(Payload<ChannelSubscriptionGiftV1>)

Channel Subscription Gift V1 Event

ChannelSubscriptionMessageV1(Payload<ChannelSubscriptionMessageV1>)

Channel Subscription Message V1 Event

Implementations

impl Event

pub fn parse(source: &str) -> Result<Event, PayloadParseError>

Parse string slice as an Event. Consider using Event::parse_http instead.

pub fn is_notification(&self) -> bool

Returns true if the message in the Payload is Notification.

pub fn is_revocation(&self) -> bool

Returns true if the message in the Payload is Revocation.

pub fn is_verification_request(&self) -> bool

Returns true if the message in the Payload is VerificationRequest.

pub fn get_verification_request(&self) -> Option<&VerificationRequest>

If this event is a VerificationRequest, return the VerificationRequest message, including the message.

pub fn subscription(&self) -> Result<EventSubSubscription, Error>

Make a EventSubSubscription from this notification.

pub fn verify_payload<B>(request: &Request<B>, secret: &[u8]) -> bool
where B: AsRef<[u8]>,

Available on crate feature hmac only.

Verify that this event is authentic using HMAC-SHA256.

HMAC key is secret, HMAC message is a concatenation of Twitch-Eventsub-Message-Id header, Twitch-Eventsub-Message-Timestamp header and the request body. HMAC signature is Twitch-Eventsub-Message-Signature header.

impl Event

pub fn parse_http<B>(request: &Request<B>) -> Result<Event, PayloadParseError>
where B: AsRef<[u8]>,

Parse a http payload as an Event

Create the webhook via CreateEventSubSubscription according to the Eventsub WebHooks guide

pub fn parse_websocket( frame: &str, ) -> Result<EventsubWebsocketData<'_>, PayloadParseError>

Parse a websocket frame as an EventsubWebsocketData

Create the websocket via CreateEventSubSubscription according to the Eventsub WebSocket guide

Examples

use twitch_api::eventsub::{Event, EventsubWebsocketData};
let notification = r#"
{
    "metadata": {
        "message_id": "befa7b53-d79d-478f-86b9-120f112b044e",
        "message_type": "notification",
        "message_timestamp": "2019-11-16T10:11:12.123Z",
        "subscription_type": "channel.follow",
        "subscription_version": "1"
    },
    "payload": {
        "subscription": {
            "id": "f1c2a387-161a-49f9-a165-0f21d7a4e1c4",
            "status": "enabled",
            "type": "channel.follow",
            "version": "1",
            "cost": 1,
            "condition": {
                "broadcaster_user_id": "12826"
            },
            "transport": {
                "method": "websocket",
                "session_id": "AQoQexAWVYKSTIu4ec_2VAxyuhAB"
            },
            "created_at": "2019-11-16T10:11:12.123Z"
        },
        "event": {
            "user_id": "1337",
            "user_login": "awesome_user",
            "user_name": "Awesome_User",
            "broadcaster_user_id": "12826",
            "broadcaster_user_login": "twitch",
            "broadcaster_user_name": "Twitch",
            "followed_at": "2020-07-15T18:16:11.17106713Z"
        }
    }
}
"#;
let event: EventsubWebsocketData<'_> =
    Event::parse_websocket(notification)?;