twitch_api::eventsub::event
Module websocket

Available on crate feature eventsub only.

Types for Eventsub websocket

Structs

KeepaliveMetadata

Defines the message that the EventSub WebSocket server sends your client to indicate that the WebSocket connection is healthy.

NotificationMetadata

Defines a message that the EventSub WebSocket server sends your client when an event that you subscribe to occurs.

ReconnectMetadata

Defines the message that the EventSub WebSocket server sends if the server must drop the connection.

ReconnectPayload

Defines the message that the EventSub WebSocket server sends if the server must drop the connection.

RevocationMetadata

Defines the message that the EventSub WebSocket server sends if the user no longer exists or they revoked the authorization token that the subscription relied on.

SessionData

Session information sent with the welcome

WelcomeMetadata

Defines the first message that the EventSub WebSocket server sends after your client connects to it.

WelcomePayload

Defines the first message that the EventSub WebSocket server sends after your client connects to it.

Enums

EventsubWebsocketData

EventsubWebsocketMetadata

Metadata for a websocket event