Function twitch_api::parse_json

pub fn parse_json<'a, T: Deserialize<'a>>(
    s: &'a str,
    log_ignored: bool
) -> Result<T, DeserError>

Available on crate feature serde_json only.

Parse a string as T, logging ignored fields and giving a more detailed error message on parse errors

The log_ignored argument decides if a trace of ignored value should be emitted

Examples found in repository?

src/eventsub/mod.rs (line 224)

211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312
313
314
315
316
317
318
319
320
321
322
323
324
325
326

    pub fn parse_notification(source: &str) -> Result<Payload<E>, PayloadParseError> {
        #[derive(Deserialize)]
        #[cfg_attr(feature = "deny_unknown_fields", serde(deny_unknown_fields))]
        struct Notification<E: EventSubscription> {
            #[serde(bound = "E: EventSubscription")]
            pub subscription: EventSubscriptionInformation<E>,
            #[serde(bound = "E: EventSubscription")]
            pub event: <E as EventSubscription>::Payload,
        }

        let Notification {
            subscription,
            event,
        } = parse_json::<Notification<E>>(source, true)?;

        Ok(Payload {
            subscription,
            message: Message::Notification(event),
        })
    }

    /// Parse string slice as a [`Payload`] with a message of [`Message::Revocation`].
    ///
    /// # Examples
    ///
    /// ```rust
    /// use twitch_api::eventsub::{channel::ChannelFollowV1, Payload};
    /// let notification = r#"
    /// {
    ///     "subscription": {
    ///         "id": "f1c2a387-161a-49f9-a165-0f21d7a4e1c4",
    ///         "status": "authorization_revoked",
    ///         "type": "channel.follow",
    ///         "cost": 1,
    ///         "version": "1",
    ///         "condition": {
    ///             "broadcaster_user_id": "12826"
    ///         },
    ///         "transport": {
    ///             "method": "webhook",
    ///             "callback": "https://example.com/webhooks/callback"
    ///         },
    ///         "created_at": "2019-11-16T10:11:12.123Z"
    ///     }
    /// }
    /// "#;
    /// let payload: Payload<ChannelFollowV1> =
    ///     Payload::parse_revocation(notification)?;
    /// # Ok::<(), Box<dyn std::error::Error>>(())
    /// ```
    pub fn parse_revocation(source: &str) -> Result<Payload<E>, PayloadParseError> {
        #[derive(Deserialize)]
        #[cfg_attr(feature = "deny_unknown_fields", serde(deny_unknown_fields))]
        struct Notification<E: EventSubscription> {
            #[serde(bound = "E: EventSubscription")]
            pub subscription: EventSubscriptionInformation<E>,
        }

        let Notification { subscription } = parse_json::<Notification<E>>(source, true)?;

        Ok(Payload {
            subscription,
            message: Message::Revocation(),
        })
    }

    /// Parse string slice as a [`Payload`] with a message of [`Message::VerificationRequest`].
    ///
    /// # Examples
    ///
    /// ```rust
    /// use twitch_api::eventsub::{channel::ChannelFollowV1, Payload};
    /// let notification = r#"
    /// {
    ///     "challenge": "pogchamp-kappa-360noscope-vohiyo",
    ///     "subscription": {
    ///         "id": "f1c2a387-161a-49f9-a165-0f21d7a4e1c4",
    ///         "status": "webhook_callback_verification_pending",
    ///         "type": "channel.follow",
    ///         "version": "1",
    ///         "cost": 1,
    ///         "condition": {
    ///             "broadcaster_user_id": "12826"
    ///         },
    ///         "transport": {
    ///             "method": "webhook",
    ///             "callback": "https://example.com/webhooks/callback"
    ///         },
    ///         "created_at": "2019-11-16T10:11:12.123Z"
    ///     }
    /// }
    /// "#;
    /// let payload: Payload<ChannelFollowV1> =
    ///     Payload::parse_verification_request(notification)?;
    /// # Ok::<(), Box<dyn std::error::Error>>(())
    /// ```
    pub fn parse_verification_request(source: &str) -> Result<Payload<E>, PayloadParseError> {
        #[derive(Deserialize)]
        #[cfg_attr(feature = "deny_unknown_fields", serde(deny_unknown_fields))]
        struct Notification<E: EventSubscription> {
            #[serde(bound = "E: EventSubscription")]
            pub subscription: EventSubscriptionInformation<E>,
            #[serde(bound = "E: EventSubscription")]
            pub challenge: String,
        }

        let Notification {
            subscription,
            challenge,
        } = parse_json::<Notification<E>>(source, true)?;

        Ok(Payload {
            subscription,
            message: Message::VerificationRequest(VerificationRequest { challenge }),
        })
    }

Hide additional examples

src/helix/endpoints/schedule/update_channel_stream_schedule_segment.rs (line 158)

148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173

    fn parse_inner_response(
        request: Option<Self>,
        uri: &http::Uri,
        response: &str,
        status: http::StatusCode,
    ) -> Result<helix::Response<Self, Self::Response>, helix::HelixRequestPatchError>
    where
        Self: Sized,
    {
        let response: helix::InnerResponse<<Self as Request>::Response> =
            helix::parse_json(response, true).map_err(|e| {
                helix::HelixRequestPatchError::DeserializeError(
                    response.to_string(),
                    e,
                    uri.clone(),
                    status,
                )
            })?;
        Ok(helix::Response {
            data: response.data,
            pagination: response.pagination.cursor,
            request,
            total: response.total,
            other: None,
        })
    }

src/helix/endpoints/search/search_categories.rs (line 109)

99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124

    fn parse_inner_response(
        request: Option<Self>,
        uri: &http::Uri,
        response: &str,
        status: http::StatusCode,
    ) -> Result<helix::Response<Self, Self::Response>, helix::HelixRequestGetError>
    where
        Self: Sized,
    {
        let response: helix::InnerResponse<Option<Self::Response>> =
            helix::parse_json(response, true).map_err(|e| {
                helix::HelixRequestGetError::DeserializeError(
                    response.to_string(),
                    e,
                    uri.clone(),
                    status,
                )
            })?;
        Ok(helix::Response {
            data: response.data.unwrap_or_default(),
            pagination: response.pagination.cursor,
            request,
            total: response.total,
            other: None,
        })
    }

src/helix/endpoints/charity/get_charity_campaign.rs (line 114)

104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129

    fn parse_inner_response(
        request: Option<Self>,
        uri: &http::Uri,
        response: &str,
        status: http::StatusCode,
    ) -> Result<helix::Response<Self, <Self as Request>::Response>, helix::HelixRequestGetError>
    where
        Self: Sized,
    {
        let response: helix::InnerResponse<Vec<_>> =
            crate::parse_json(response, true).map_err(|e| {
                helix::HelixRequestGetError::DeserializeError(
                    response.to_string(),
                    e,
                    uri.clone(),
                    status,
                )
            })?;
        Ok(helix::Response {
            data: response.data.into_iter().next(),
            pagination: response.pagination.cursor,
            request,
            total: response.total,
            other: response.other,
        })
    }

src/tmi/mod.rs (line 94)

71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95

    pub async fn get_chatters(
        &'a self,
        broadcaster: &types::UserNameRef,
    ) -> Result<GetChatters, RequestError<<C as crate::HttpClient>::Error>> {
        let url = format!(
            "{}{}{}{}",
            crate::TWITCH_TMI_URL.as_str(),
            "group/user/",
            broadcaster.as_str().replace('#', "").to_ascii_lowercase(),
            "/chatters"
        );
        let req = http::Request::builder()
            .uri(url)
            .body(Vec::with_capacity(0).into())?;
        let resp = self
            .client
            .req(req)
            .await
            .map_err(|e| RequestError::RequestError(Box::new(e)))?;
        let (parts, mut body) = resp.into_parts();
        let resp = http::Response::from_parts(parts, hyper::body::to_bytes(&mut body).await?);
        let text = std::str::from_utf8(resp.body())
            .map_err(|e| RequestError::Utf8Error(resp.body().to_vec(), e))?;
        crate::parse_json(text, true).map_err(Into::into)
    }

src/helix/endpoints/users/get_users_follows.rs (line 171)

154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189

    fn parse_inner_response(
        request: Option<Self>,
        uri: &http::Uri,
        response: &str,
        status: http::StatusCode,
    ) -> Result<helix::Response<Self, Self::Response>, helix::HelixRequestGetError>
    where
        Self: Sized,
    {
        #[derive(PartialEq, Deserialize, Debug, Clone)]
        struct InnerResponse {
            data: Vec<FollowRelationship>,
            total: i64,
            #[serde(default)]
            pagination: helix::Pagination,
        }

        let response: InnerResponse = helix::parse_json(response, true).map_err(|e| {
            helix::HelixRequestGetError::DeserializeError(
                response.to_string(),
                e,
                uri.clone(),
                status,
            )
        })?;
        Ok(helix::Response {
            data: UsersFollows {
                total: response.total,
                follow_relationships: response.data,
            },
            pagination: response.pagination.cursor,
            request,
            total: Some(response.total),
            other: None,
        })
    }

Additional examples can be found in:

src/helix/endpoints/polls/create_poll.rs

src/helix/endpoints/predictions/create_prediction.rs

src/helix/endpoints/raids/start_a_raid.rs

src/helix/endpoints/subscriptions/check_user_subscription.rs

src/helix/endpoints/points/create_custom_rewards.rs

src/helix/endpoints/moderation/ban_user.rs

src/helix/request.rs

src/helix/endpoints/bits/get_bits_leaderboard.rs

src/helix/endpoints/chat/get_chat_settings.rs

src/eventsub/event.rs

src/helix/endpoints/eventsub/get_eventsub_subscriptions.rs

src/helix/endpoints/chat/update_chat_settings.rs

src/helix/endpoints/points/update_custom_reward.rs

src/helix/endpoints/points/update_redemption_status.rs

src/helix/endpoints/polls/end_poll.rs

src/helix/endpoints/predictions/end_prediction.rs

src/helix/client/custom.rs

src/helix/endpoints/eventsub/create_eventsub_subscription.rs

src/pubsub/mod.rs