Function twitch_api::parse_json

pub fn parse_json<'a, T: Deserialize<'a>>(
    s: &'a str,
    log_ignored: bool
) -> Result<T, DeserError>

Available on crate feature serde_json only.

Parse a string as T, logging ignored fields and giving a more detailed error message on parse errors

The log_ignored argument decides if a trace of ignored value should be emitted