twitch_api
Module helix

Available on crate feature helix only.

Helix endpoints or the New Twitch API

HelixClient provides convenience function for sending requests as described on the crate documentation.

Manual Usage

Aside from using HelixClient, you can decide to use this library without any specific client implementation.

use twitch_api::{helix::{self, Request, RequestGet, users::{GetUsersRequest, User}}, types};

let logins: &[&types::UserNameRef] = &["justintv123".into()];
let request = GetUsersRequest::logins(logins);

// Send it however you want
// Create a [`http::Response<hyper::body::Bytes>`] with RequestGet::create_request, which takes an access token and a client_id
let response = send_http_request(request.create_request("accesstoken", "client_id")?)?;

// then parse the response
let uri = request.get_uri()?;
let user: helix::Response<_, Vec<User>> = GetUsersRequest::parse_response(Some(request), &uri, response)?;
println!("{:#?}", user);

Implemented Endpoints

Ads 游릭 3/3

Endpoint	Helper	Module

Start Commercial	-	channels::start_commercial

Get Ad Schedule	-	channels::get_ad_schedule

Snooze Next Ad	-	channels::snooze_next_ad

Analytics 游댮 0/2

Endpoint	Helper	Module

Get Extension Analytics	-	-

Get Game Analytics	-	-

Bits 游리 2/3

Endpoint	Helper	Module

Get Bits Leaderboard	-	bits::get_bits_leaderboard

Get Cheermotes	-	bits::get_cheermotes

Get Extension Transactions	-	-

CCLs 游댮 0/1

Endpoint	Helper	Module

Get Content Classification Labels	-	-

Channel Points 游릭 6/6

Endpoint	Helper	Module

Create Custom Rewards	-	points::create_custom_rewards

Delete Custom Reward	-	points::delete_custom_reward

Get Custom Reward	-	points::get_custom_reward

Get Custom Reward Redemption	-	points::get_custom_reward_redemption

Update Custom Reward	-	points::update_custom_reward

Update Redemption Status	-	points::update_redemption_status

Channels 游릭 5/5

Endpoint	Helper	Module

Get Channel Information	-	channels::get_channel_information

Modify Channel Information	-	channels::modify_channel_information

Get Channel Editors	-	channels::get_channel_editors

Get Followed Channels	HelixClient::get_followed_channels	channels::get_followed_channels

Get Channel Followers	-	channels::get_channel_followers

Charity 游릭 2/2

Endpoint	Helper	Module

Get Charity Campaign	-	charity::get_charity_campaign

Get Charity Campaign Donations	-	charity::get_charity_campaign_donations

Chat 游리 13/15

Endpoint	Helper	Module

Get Chatters	HelixClient::get_chatters	chat::get_chatters

Get Channel Emotes	-	chat::get_channel_emotes

Get Global Emotes	HelixClient::get_global_emotes	chat::get_global_emotes

Get Emote Sets	HelixClient::get_emote_sets	chat::get_emote_sets

Get Channel Chat Badges	-	chat::get_channel_chat_badges

Get Global Chat Badges	-	chat::get_global_chat_badges

Get Chat Settings	HelixClient::get_chat_settings	chat::get_chat_settings

Get Shared Chat Session	-	-

Get User Emotes	-	-

Update Chat Settings	-	chat::update_chat_settings

Send Chat Announcement	HelixClient::send_chat_announcement	chat::send_chat_announcement

Send a Shoutout	-	chat::send_a_shoutout

Send Chat Message	-	chat::send_chat_message

Get User Chat Color	HelixClient::get_user_chat_color	chat::get_user_chat_color

Update User Chat Color	HelixClient::update_user_chat_color	chat::update_user_chat_color

Clips 游릭 2/2

Endpoint	Helper	Module

Create Clip	-	clips::create_clip

Get Clips	-	clips::get_clips

Conduits 游리 4/6

Endpoint	Helper	Module

Get Conduits	HelixClient::get_conduits	eventsub::get_conduits

Create Conduits	HelixClient::create_conduit	eventsub::create_conduit

Update Conduits	-	-

Delete Conduit	-	-

Get Conduit Shards	HelixClient::get_conduit_shards	eventsub::get_conduit_shards

Update Conduit Shards	HelixClient::update_conduit_shards	eventsub::update_conduit_shards

Entitlements 游댮 0/2

Endpoint	Helper	Module

Get Drops Entitlements	-	-

Update Drops Entitlements	-	-

EventSub 游릭 3/3

Endpoint	Helper	Module

Create EventSub Subscription	HelixClient::create_eventsub_subscription	eventsub::create_eventsub_subscription

Delete EventSub Subscription	HelixClient::delete_eventsub_subscription	eventsub::delete_eventsub_subscription

Get EventSub Subscriptions	HelixClient::get_eventsub_subscriptions	eventsub::get_eventsub_subscriptions

Extensions 游댮 0/12

Endpoint	Helper	Module

Get Extension Configuration Segment	-	-

Set Extension Configuration Segment	-	-

Set Extension Required Configuration	-	-

Send Extension PubSub Message	-	-

Get Extension Live Channels	-	-

Get Extension Secrets	-	-

Create Extension Secret	-	-

Send Extension Chat Message	-	-

Get Extensions	-	-

Get Released Extensions	-	-

Get Extension Bits Products	-	-

Update Extension Bits Product	-	-

Games 游릭 2/2

Endpoint	Helper	Module

Get Top Games	-	games::get_top_games

Get Games	-	games::get_games

Goals 游릭 1/1

Endpoint	Helper	Module

Get Creator Goals	-	goals::get_creator_goals

Guest Star 游댮 0/12

Endpoint	Helper	Module

Get Channel Guest Star Settings	-	-

Update Channel Guest Star Settings	-	-

Get Guest Star Session	-	-

Create Guest Star Session	-	-

End Guest Star Session	-	-

Get Guest Star Invites	-	-

Send Guest Star Invite	-	-

Delete Guest Star Invite	-	-

Assign Guest Star Slot	-	-

Update Guest Star Slot	-	-

Delete Guest Star Slot	-	-

Update Guest Star Slot Settings	-	-

Hype Train 游댮 0/1

Endpoint	Helper	Module

Get Hype Train Events	-	-

Moderation 游릭 23/23

Endpoint	Helper	Module

Check AutoMod Status	-	moderation::check_automod_status

Manage Held AutoMod Messages	-	moderation::manage_held_automod_messages

Get AutoMod Settings	-	moderation::get_automod_settings

Update AutoMod Settings	-	moderation::update_automod_settings

Get Banned Users	-	moderation::get_banned_users

Ban User	HelixClient::ban_user	moderation::ban_user

Unban User	HelixClient::unban_user	moderation::unban_user

Get Unban Requests	HelixClient::get_unban_requests	moderation::get_unban_requests

Resolve Unban Requests	-	moderation::resolve_unban_request

Get Blocked Terms	-	moderation::get_blocked_terms

Add Blocked Term	-	moderation::add_blocked_term

Remove Blocked Term	-	moderation::remove_blocked_term

Delete Chat Messages	-	moderation::delete_chat_messages

Get Moderated Channels	HelixClient::get_moderated_channels	moderation::get_moderated_channels

Get Moderators	-	moderation::get_moderators

Add Channel Moderator	HelixClient::add_channel_moderator	moderation::add_channel_moderator

Remove Channel Moderator	HelixClient::remove_channel_moderator	moderation::remove_channel_moderator

Get VIPs	-	channels::get_vips

Add Channel VIP	HelixClient::add_channel_vip	channels::add_channel_vip

Remove Channel VIP	HelixClient::remove_channel_vip	channels::remove_channel_vip

Update Shield Mode Status	-	moderation::update_shield_mode_status

Get Shield Mode Status	-	moderation::get_shield_mode_status

Warn Chat User	HelixClient::warn_chat_user	moderation::warn_chat_user

Polls 游릭 3/3

Endpoint	Helper	Module

Get Polls	-	polls::get_polls

Create Poll	-	polls::create_poll

End Poll	-	polls::end_poll

Predictions 游릭 3/3

Endpoint	Helper	Module

Get Predictions	-	predictions::get_predictions

Create Prediction	-	predictions::create_prediction

End Prediction	-	predictions::end_prediction

Raids 游릭 2/2

Endpoint	Helper	Module

Start a raid	HelixClient::start_a_raid	raids::start_a_raid

Cancel a raid	HelixClient::cancel_a_raid	raids::cancel_a_raid

Schedule 游리 5/6

Endpoint	Helper	Module

Get Channel Stream Schedule	-	schedule::get_channel_stream_schedule

Get Channel iCalendar	-	-

Update Channel Stream Schedule	-	schedule::update_channel_stream_schedule

Create Channel Stream Schedule Segment	-	schedule::create_channel_stream_schedule_segment

Update Channel Stream Schedule Segment	-	schedule::update_channel_stream_schedule_segment

Delete Channel Stream Schedule Segment	-	schedule::delete_channel_stream_schedule_segment

Search 游릭 2/2

Endpoint	Helper	Module

Search Categories	HelixClient::search_categories	search::search_categories

Search Channels	HelixClient::search_channels	search::search_channels

Streams 游리 2/5

Endpoint	Helper	Module

Get Stream Key	-	-

Get Streams	-	streams::get_streams

Get Followed Streams	HelixClient::get_followed_streams	streams::get_followed_streams

Create Stream Marker	-	-

Get Stream Markers	-	-

Subscriptions 游릭 2/2

Endpoint	Helper	Module

Get Broadcaster Subscriptions	HelixClient::get_broadcaster_subscriptions	subscriptions::get_broadcaster_subscriptions

Check User Subscription	-	subscriptions::check_user_subscription

Tags 游리 1/2

Endpoint	Helper	Module

Get All Stream Tags	-	tags::get_all_stream_tags

Get Stream Tags	-	-

Teams 游릭 2/2

Endpoint	Helper	Module

Get Channel Teams	-	teams::get_channel_teams

Get Teams	-	teams::get_teams

Users 游리 4/8

Endpoint	Helper	Module

Get Users	-	users::get_users

Update User	-	-

Get User Block List	-	users::get_user_block_list

Block User	HelixClient::block_user	users::block_user

Unblock User	HelixClient::unblock_user	users::unblock_user

Get User Extensions	-	-

Get User Active Extensions	-	-

Update User Extensions	-	-

Videos 游릭 2/2

Endpoint	Helper	Module

Get Videos	-	videos::get_videos

Delete Videos	-	videos::delete_videos

Whispers 游릭 1/1

Endpoint	Helper	Module

Send Whisper	HelixClient::send_whisper	whispers::send_whisper

Re-exports

pub use twitch_oauth2::Scope;

twitch_oauth2

Modules

bits

Helix endpoints regarding bits

channels

Helix endpoints regarding channels

charity

Helix endpoints regarding charities

chat

Helix endpoints regarding chat

clientclient

Client for Helix endpoints

clips

Helix endpoints regarding clips

eventsubeventsub

Helix endpoints regarding EventSub

games

Helix endpoints regarding games

goals

Helix endpoints regarding creator goals

hypetrain

Helix endpoints regarding hype trains

moderation

Helix endpoints regarding moderation

points

Helix endpoints regarding channel points/redeems

polls

Helix endpoints regarding channel polls

predictions

Helix endpoints regarding channel predictions

raids

Helix endpoints regarding channel raids

request

Requests for driving the API

response

Responses contains the return values of a request.

schedule

Helix endpoints regarding stream schedules

search

Helix endpoints regarding search

streams

Helix endpoints regarding streams

subscriptions

Helix endpoints regarding subscriptions

tags

Helix endpoints regarding tags

teams

Helix endpoints regarding subscriptions

users

Helix endpoints regarding users

videos

Helix endpoints regarding videos

whispers

Helix endpoints regarding whispers

Structs

Cursor

A cursor is a pointer to the current 랋age in the twitch api pagination

CursorRef

A cursor is a pointer to the current 랋age in the twitch api pagination

EmptyBody

An empty body.

HelixClientclient

Client for Helix or the New Twitch API

Response

Response retrieved from endpoint. Data is the type in Request::Response

Enums

BodyError

Errors that can happen when creating a body

ClientExtErrorclient

Error type to combine a http client error with a other error

ClientRequestErrorclient

Errors for HelixClient::req_get and similar functions.

CreateRequestError

Could not create request

HelixRequestDeleteError

Could not parse DELETE response

HelixRequestGetError

Could not parse GET response

HelixRequestPatchError

Could not parse PATCH response

HelixRequestPostError

Could not parse POST response

HelixRequestPutError

Could not parse PUT response

InvalidUri

Errors that can happen when creating http::Uri for Request

SerializeError

Errors from the query serializer

Traits

HelixRequestBody

Create a body. Used for specializing request bodies

Paginated

A request that can be paginated.

Request

A request is a Twitch endpoint, see New Twitch API reference

RequestDelete

Helix endpoint DELETEs information

RequestGet

Helix endpoint GETs information

RequestPatch

Helix endpoint PATCHs information

RequestPost

Helix endpoint POSTs information

RequestPut

Helix endpoint PUTs information

Functions

make_streamclient

Make a paginate-able request into a stream