twitch_api::helix
Module channels

Available on crate feature helix only.

Helix endpoints regarding channels

Examples

let client = HelixClient::new();
let ids: &[&twitch_types::UserIdRef] = &["1234".into()];
let req = GetChannelInformationRequest::broadcaster_ids(ids);

// Get Channel Information Request only returns one entry.
println!("{:?}", &client.req_get(req, &token).await?.data);

Modules

add_channel_vip

Adds a VIP to the broadcaster’s chat room. add-channel-vip

get_ad_schedule

Gets ad schedule related information, including snooze, when the last ad was run, when the next ad is scheduled, and if the channel is currently in pre-roll free time. get-ad-schedule

get_channel_editors

Gets a list of users who have editor permissions for a specific channel. get-channel-editors

get_channel_followers

Gets a list of users that follow the specified broadcaster. You can also use this endpoint to see whether a specific user follows the broadcaster. get-channel-followers

get_channel_information

Gets channel information for users. get-channel-information

get_followed_channels

Gets a list of broadcasters that the specified user follows. You can also use this endpoint to see whether a user follows a specific broadcaster. get-followed-channels

get_vips

Gets a list of the channel’s VIPs. get-vips

modify_channel_information

Modify channel information for users. modify-channel-information

remove_channel_vip

Removes a VIP from the broadcaster’s chat room. remove-channel-vip

snooze_next_ad

If available, pushes back the timestamp of the upcoming automatic mid-roll ad by 5 minutes. This endpoint duplicates the snooze functionality in the creator dashboard’s Ads Manager. snooze-next-ad

start_commercial

Starts a commercial on a specified channel. start-commercial

Structs

AdSchedule

Return Values for Get Ad Schedule

AddChannelVipRequest

Query Parameters for Add Channel Vip

ChannelInformation

Return Values for Get Channel Information

ContentClassificationLabel

List of labels that should be set as the Channel’s CCLs.

Editor

Return Values for Get Channel Editors

FollowedBroadcaster

Return Values for Get Followed Channels

Follower

Return Values for Get Channel Followers

GetAdScheduleRequest

Query Parameters for Get Ad Schedule

GetChannelEditorsRequest

Query Parameters for Get Channel Editors

GetChannelFollowersRequest

Query Parameters for Get Channel Followers

GetChannelInformationRequest

Query Parameters for Get Channel Information

GetFollowedChannels

Query Parameters for Get Followed Channels

GetVipsRequest

Query Parameters for Get VIPs

ModifyChannelInformationBody

Body Parameters for Modify Channel Information

ModifyChannelInformationRequest

Query Parameters for Modify Channel Information

RemoveChannelVipRequest

Query Parameters for Remove Channel VIP

SnoozeNextAdRequest

Query Parameters for Snooze Next Ad

SnoozedAdSchedule

Return Values for Snooze Next Ad

StartCommercial

Return Values for Start Commercial

StartCommercialBody

Body Parameters for Start Commercial

StartCommercialRequest

Query Parameters for Start Commercial

Vip

Return Values for Get VIPs

Enums

AddChannelVipResponse

Return Values for Add Channel Vip

ModifyChannelInformation

Return Values for Modify Channel Information

RemoveChannelVipResponse

Return Values for Remove Channel VIP