Module twitch_api::helix::chat

Available on crate feature helix only.

Helix endpoints regarding chat

Modules

get_channel_chat_badges

Gets a list of custom chat badges that can be used in chat for the specified channel. This includes subscriber badges and Bit badges. get-channel-chat-badges

get_channel_emotes

Gets all custom emotes for a specific Twitch channel including subscriber emotes, Bits tier emotes, and follower emotes. get-channel-emotes

get_chat_settings

Gets the broadcaster’s chat settings. get-chat-settings

get_chatters

Gets the list of users that are connected to the specified broadcaster’s chat session. get-chatters

get_emote_sets

Gets all Twitch emotes for one or more specific emote sets. get-emote-sets

get_global_chat_badges

Gets a list of chat badges that can be used in chat for any channel. get-global-chat-badges

get_global_emotes

Gets all global emotes. Global emotes are Twitch-specific emoticons that every user can use in Twitch chat. get-global-emotes

get_user_chat_color

Gets the color used for the user’s name in chat. get-user-chat-color

send_chat_announcement

Sends an announcement to the broadcaster’s chat room. send-chat-announcement

update_chat_settings

Updates the broadcaster’s chat settings. update-chat-settings

update_user_chat_color

Updates the color used for the user’s name in chat.

Structs

AnnouncementColorError

An error for an invalid AnnouncementColor

BadgeSet

A set of badges

ChannelEmote

A chat emote

ChatBadge

A chat Badge

ChatSettings

Chat settings

Chatter

Return Values for Get Chatters

GetChannelChatBadgesRequest

Query Parameters for Get Channel Chat Badges

GetChannelEmotesRequest

Query Parameters for Get Channel Emotes

GetChatSettingsRequest

Query Parameters for Get Chat Settings

GetChattersRequest

Query Parameters for Get Chatters

GetEmoteSetsRequest

Query Parameters for Get Channel Emotes

GetGlobalChatBadgesRequest

Query Parameters for Get Global Chat Badges

GetGlobalEmotesRequest

Query Parameters for Get Channel Emotes

GetUserChatColorRequest

Query Parameters for Get Chatters

GlobalEmote

A chat emote

SendChatAnnouncementBody

Body Parameters for Send Chat Announcement

SendChatAnnouncementRequest

Query Parameters for Send Chat Announcement

UpdateChatSettingsBody

Body Parameters for Update Chat Settings

UpdateChatSettingsRequest

Query Parameters for Update Chat Settings

UpdateUserChatColorRequest

Query Parameters for Update Chat Settings

UserChatColor

Return Values for Get Chatters

Enums

AnnouncementColor

Valid colors for announcements

SendChatAnnouncementResponse

Return Values for Send Chat Announcement

UpdateUserChatColorResponse

Return Values for Update Chat Settings