twitch_api::helix

Module chat

Available on crate feature helix only.

Helix endpoints regarding chat

Implemented endpoints

Chat 🟢 15/15

Endpoint	Helper	Module

Get Chatters	HelixClient::get_chatters	get_chatters

Get Channel Emotes	HelixClient::get_channel_emotes_from_id, HelixClient::get_channel_emotes_from_login	get_channel_emotes

Get Global Emotes	HelixClient::get_global_emotes	get_global_emotes

Get Emote Sets	HelixClient::get_emote_sets	get_emote_sets

Get Channel Chat Badges	-	get_channel_chat_badges

Get Global Chat Badges	-	get_global_chat_badges

Get Chat Settings	HelixClient::get_chat_settings	get_chat_settings

Get Shared Chat Session	HelixClient::get_shared_chat_session	get_shared_chat_session

Get User Emotes	HelixClient::get_user_emotes, HelixClient::get_user_emotes_in_channel	get_user_emotes

Update Chat Settings	-	update_chat_settings

Send Chat Announcement	HelixClient::send_chat_announcement	send_chat_announcement

Send a Shoutout	-	send_a_shoutout

Send Chat Message	HelixClient::send_chat_message, HelixClient::send_chat_message_reply	send_chat_message

Get User Chat Color	HelixClient::get_user_chat_color, HelixClient::get_users_chat_colors	get_user_chat_color

Update User Chat Color	HelixClient::update_user_chat_color	update_user_chat_color

Modules

get_channel_chat_badges

Gets a list of custom chat badges that can be used in chat for the specified channel. This includes subscriber badges and Bit badges. get-channel-chat-badges

get_channel_emotes

Gets all custom emotes for a specific Twitch channel including subscriber emotes, Bits tier emotes, and follower emotes. get-channel-emotes

get_chat_settings

Gets the broadcaster’s chat settings. get-chat-settings

get_chatters

Gets the list of users that are connected to the specified broadcaster’s chat session. get-chatters

get_emote_sets

Gets all Twitch emotes for one or more specific emote sets. get-emote-sets

get_global_chat_badges

Gets a list of chat badges that can be used in chat for any channel. get-global-chat-badges

get_global_emotes

Gets all global emotes. Global emotes are Twitch-specific emoticons that every user can use in Twitch chat. get-global-emotes

get_shared_chat_session

Retrieves the active shared chat session for a channel. get-shared-chat-session

get_user_chat_color

Gets the color used for the user’s name in chat. get-user-chat-color

get_user_emotes

Retrieves emotes available to the user across all channels. get-user-emotes

send_a_shoutout

Sends a Shoutout to the specified broadcaster. send-a-shoutout

send_chat_announcement

Sends an announcement to the broadcaster’s chat room. send-chat-announcement

send_chat_message

Sends an message in the broadcaster’s chat room. send-chat-message

update_chat_settings

Updates the broadcaster’s chat settings. update-chat-settings

update_user_chat_color

Updates the color used for the user’s name in chat.

Structs

BadgeSet

A set of badges

ChannelEmote

A chat emote

ChatBadge

A chat Badge

ChatMessageDropReason

A drop reason of a sent message.

ChatSettings

Chat settings

Chatter

Return Values for Get Chatters

GetChannelChatBadgesRequest

Query Parameters for Get Channel Chat Badges

GetChannelEmotesRequest

Query Parameters for Get Channel Emotes

GetChatSettingsRequest

Query Parameters for Get Chat Settings

GetChattersRequest

Query Parameters for Get Chatters

GetEmoteSetsRequest

Query Parameters for Get Channel Emotes

GetGlobalChatBadgesRequest

Query Parameters for Get Global Chat Badges

GetGlobalEmotesRequest

Query Parameters for Get Channel Emotes

GetSharedChatSessionRequest

Query Parameters for Get Shared Chat Session

GetUserChatColorRequest

Query Parameters for Get Chatters

GetUserEmotesRequest

Query Parameters for Get User Emotes

GlobalEmote

A chat emote

SendAShoutoutRequest

Query Parameters for Send A Shoutout

SendChatAnnouncementBody

Body Parameters for Send Chat Announcement

SendChatAnnouncementRequest

Query Parameters for Send Chat Announcement

SendChatMessageBody

Body Parameters for Send Chat message

SendChatMessageRequest

Query Parameters for Send Chat message

SendChatMessageResponse

Return Values for Send Chat message

SharedChatParticipant

A participant in a shared chat session

SharedChatSession

Return Values for Get Shared Chat Session

UpdateChatSettingsBody

Body Parameters for Update Chat Settings

UpdateChatSettingsRequest

Query Parameters for Update Chat Settings

UpdateUserChatColorRequest

Query Parameters for Update Chat Settings

UserChatColor

Return Values for Get Chatters

UserEmote

Return Values for Get User Emotes

Enums

AnnouncementColoreventsub or helix

Valid colors for announcements

ChatMessageDropCode

Code for why a message was dropped.

SendAShoutoutResponse

Return Values for Send A Shoutout

SendChatAnnouncementResponse

Return Values for Send Chat Announcement

UpdateUserChatColorResponse

Return Values for Update Chat Settings