Type Alias twitch_api::helix::chat::get_channel_emotes::GetChannelEmotesResponse

pub type GetChannelEmotesResponse = ChannelEmote;

Available on crate feature helix only.

Return Values for Get Channel Emotes

get-channel-emotes

Aliased Type

struct GetChannelEmotesResponse {
    pub id: EmoteId,
    pub name: String,
    pub images: Image,
    pub tier: Option<SubscriptionTier>,
    pub emote_type: String,
    pub emote_set_id: EmoteSetId,
    pub format: Vec<EmoteAnimationSetting>,
    pub scale: Vec<EmoteScale>,
    pub theme_mode: Vec<EmoteThemeMode>,
}

Fields
id: EmoteId

ID of the emote.

name: String

Name of the emote a viewer types into Twitch chat for the image to appear.

images: Image

Object of image URLs for the emote.

tier: Option<SubscriptionTier>

If the emote_type is “subscriptions”, this indicates the subscriber tier at which the emote is unlocked. Set to an empty string otherwise.

emote_type: String

The type of emote.

The most common values for custom channel emotes are

subscriptions: Indicates a custom subscriber emote.

bitstier: Indicates a custom Bits tier emote.

follower: Indicates a custom follower emote.

emote_set_id: EmoteSetId

ID of the emote set the emote belongs to.

format: Vec<EmoteAnimationSetting>

The formats that the emote is available in.

scale: Vec<EmoteScale>

The sizes that the emote is available in.

theme_mode: Vec<EmoteThemeMode>

The background themes that the emote is available in.