Module twitch_api::helix::chat::get_emote_sets

Available on crate feature helix only.

Gets all Twitch emotes for one or more specific emote sets. get-emote-sets

Accessing the endpoint

Request: GetEmoteSetsRequest

To use this endpoint, construct a GetEmoteSetsRequest with the GetEmoteSetsRequest::emote_set_ids() method.

use twitch_api::helix::chat::get_emote_sets;
let ids: &[&twitch_types::EmoteSetIdRef] = &["1234".into()];
let request = get_emote_sets::GetEmoteSetsRequest::emote_set_ids(ids);

Response: Emote

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, chat::get_emote_sets};
let ids: &[&twitch_types::EmoteSetIdRef] = &["1234".into()];
let request = get_emote_sets::GetEmoteSetsRequest::emote_set_ids(ids);
let response: Vec<helix::chat::get_emote_sets::Emote> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetEmoteSetsRequest::parse_response(None, &request.get_uri(), response)

Structs

Emote

Return Values for Get Channel Emotes

GetEmoteSetsRequest

Query Parameters for Get Channel Emotes