Module twitch_api::helix::chat::get_global_chat_badges

Available on crate feature helix only.

Gets a list of chat badges that can be used in chat for any channel. get-global-chat-badges

Accessing the endpoint

Request: GetGlobalChatBadgesRequest

To use this endpoint, construct a GetGlobalChatBadgesRequest with the GetGlobalChatBadgesRequest::new() method.

use twitch_api::helix::chat::get_global_chat_badges;
let request = get_global_chat_badges::GetGlobalChatBadgesRequest::new();

Response: BadgeSet

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, chat::get_global_chat_badges};
let request = get_global_chat_badges::GetGlobalChatBadgesRequest::new();
let response: Vec<helix::chat::BadgeSet> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetGlobalChatBadgesRequest::parse_response(None, &request.get_uri(), response)

Structs

GetGlobalChatBadgesRequest

Query Parameters for Get Global Chat Badges

Type Aliases

GetGlobalChatBadgesResponse

Return Values for Get Global Chat Badges