twitch_api::helix::chat

Module get_global_emotes

Available on crate feature helix only.

Gets all global emotes. Global emotes are Twitch-specific emoticons that every user can use in Twitch chat. get-global-emotes

Accessing the endpoint

Request: GetGlobalEmotesRequest

To use this endpoint, construct a GetGlobalEmotesRequest with the GetGlobalEmotesRequest::default() method.

use twitch_api::helix::chat::get_global_emotes;
let request = get_global_emotes::GetGlobalEmotesRequest::default();

Response: GlobalEmote

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix::{self, chat::get_global_emotes};
let request = get_global_emotes::GetGlobalEmotesRequest::default();
let response: Vec<helix::chat::GlobalEmote> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetGlobalEmotesRequest::parse_response(None, &request.get_uri(), response)

Structs

GetGlobalEmotesRequest

Query Parameters for Get Channel Emotes

Type Aliases

GetChannelEmotesResponse

Return Values for Get Channel Emotes