twitch_api::helix::chat

Module update_chat_settings

Available on crate feature helix only.

Updates the broadcaster’s chat settings. update-chat-settings

Accessing the endpoint

Request: UpdateChatSettingsRequest

To use this endpoint, construct an UpdateChatSettingsRequest with the UpdateChatSettingsRequest::new() method.

use twitch_api::helix::chat::update_chat_settings;
let request =
    update_chat_settings::UpdateChatSettingsRequest::new("1234", "5678");

Body: UpdateChatSettingsBody

We also need to provide a body to the request containing what we want to change.

let mut body = update_chat_settings::UpdateChatSettingsBody::default();
body.slow_mode = Some(true);
body.slow_mode_wait_time = Some(10);

Response: ChatSettings

Send the request to receive the response with HelixClient::req_patch().

use twitch_api::helix::{self, chat::update_chat_settings};
let request = update_chat_settings::UpdateChatSettingsRequest::new("1234", "5678");
let mut body = update_chat_settings::UpdateChatSettingsBody::default();
body.slow_mode = Some(true);
body.slow_mode_wait_time = Some(10);
let response: helix::chat::ChatSettings = client.req_patch(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with UpdateChatSettingsRequest::parse_response(None, &request.get_uri(), response)

Structs

UpdateChatSettingsBody

Body Parameters for Update Chat Settings

UpdateChatSettingsRequest

Query Parameters for Update Chat Settings

Type Aliases

UpdateChatSettingsResponse

Return Values for Update Chat Settings