twitch_api::helix::chat::update_chat_settings
Type Alias UpdateChatSettingsResponse

pub type UpdateChatSettingsResponse = ChatSettings;

Available on crate feature helix only.

Return Values for Update Chat Settings

update-chat-settings

Aliased Type

struct UpdateChatSettingsResponse {
    pub broadcaster_id: UserId,
    pub emote_mode: bool,
    pub follower_mode: bool,
    pub follower_mode_duration: Option<u64>,
    pub moderator_id: Option<UserId>,
    pub non_moderator_chat_delay: Option<bool>,
    pub non_moderator_chat_delay_duration: Option<u64>,
    pub slow_mode: bool,
    pub slow_mode_wait_time: Option<u64>,
    pub subscriber_mode: bool,
    pub unique_chat_mode: bool,
}

Fields
broadcaster_id: UserId

The ID of the broadcaster specified in the request.

emote_mode: bool

A Boolean value that determines whether chat messages must contain only emotes. Is true, if only messages that are 100% emotes are allowed; otherwise, false.

follower_mode: bool

A Boolean value that determines whether the broadcaster restricts the chat room to followers only, based on how long they’ve followed.

Is true, if the broadcaster restricts the chat room to followers only; otherwise, false. See follower_mode_duration for how long the followers must have followed the broadcaster to participate in the chat room.

follower_mode_duration: Option<u64>

The length of time, in minutes, that the followers must have followed the broadcaster to participate in the chat room. See follower_mode.

Is null if follower_mode is false.

moderator_id: Option<UserId>

The ID of the moderator specified in the request for chat settings.

non_moderator_chat_delay: Option<bool>

A Boolean value that determines whether the broadcaster adds a short delay before chat messages appear in the chat room. This gives chat moderators and bots a chance to remove them before viewers can see the message.

Is true, if the broadcaster applies a delay; otherwise, false. See non_moderator_chat_delay_duration for the length of the delay.

Notes

This field and non_moderator_chat_delay_duration are not received when the request is made without a specified moderator_id.

non_moderator_chat_delay_duration: Option<u64>

The amount of time, in seconds, that messages are delayed from appearing in chat. See non_moderator_chat_delay.

Is null if non_moderator_chat_delay is false.

slow_mode: bool

A Boolean value that determines whether the broadcaster limits how often users in the chat room are allowed to send messages.

Is true, if the broadcaster applies a delay; otherwise, false. See slow_mode_wait_time for the delay.

slow_mode_wait_time: Option<u64>

The amount of time, in seconds, that users need to wait between sending messages. See slow_mode.

Is null if slow_mode is false.

subscriber_mode: bool

A Boolean value that determines whether only users that subscribe to the broadcaster’s channel can talk in the chat room.

Is true, if the broadcaster restricts the chat room to subscribers only; otherwise, false.

unique_chat_mode: bool

A Boolean value that determines whether the broadcaster requires users to post only unique messages in the chat room.

Is true, if the broadcaster requires unique messages only; otherwise, false.