twitch_api::helix

Module moderation

Available on crate feature helix only.

Helix endpoints regarding moderation

Implemented endpoints

Moderation 🟢 23/23

Endpoint	Helper	Module

Check AutoMod Status	-	check_automod_status

Manage Held AutoMod Messages	-	manage_held_automod_messages

Get AutoMod Settings	-	get_automod_settings

Update AutoMod Settings	-	update_automod_settings

Get Banned Users	HelixClient::get_banned_users_in_channel_from_id	get_banned_users

Ban User	HelixClient::ban_user	ban_user

Unban User	HelixClient::unban_user	unban_user

Get Unban Requests	HelixClient::get_unban_requests	get_unban_requests

Resolve Unban Requests	-	resolve_unban_request

Get Blocked Terms	-	get_blocked_terms

Add Blocked Term	-	add_blocked_term

Remove Blocked Term	-	remove_blocked_term

Delete Chat Messages	HelixClient::delete_chat_message	delete_chat_messages

Get Moderated Channels	HelixClient::get_moderated_channels	get_moderated_channels

Get Moderators	HelixClient::get_moderators_in_channel_from_id	get_moderators

Add Channel Moderator	HelixClient::add_channel_moderator	add_channel_moderator

Remove Channel Moderator	HelixClient::remove_channel_moderator	remove_channel_moderator

Get VIPs	HelixClient::get_vips_in_channel	get_vips

Add Channel VIP	HelixClient::add_channel_vip	add_channel_vip

Remove Channel VIP	HelixClient::remove_channel_vip	remove_channel_vip

Update Shield Mode Status	-	update_shield_mode_status

Get Shield Mode Status	-	get_shield_mode_status

Warn Chat User	HelixClient::warn_chat_user	warn_chat_user

Modules

add_blocked_term

Adds a word or phrase to the broadcaster’s list of blocked terms. These are the terms that broadcasters don’t want used in their chat room. add-blocked-term

add_channel_moderator

Adds a moderator to the broadcaster’s moderation room. add-channel-moderator

add_channel_vip

Adds a VIP to the broadcaster’s chat room. add-channel-vip

ban_user

Bans a user from participating in a broadcaster’s chat room, or puts them in a timeout. ban-user

check_automod_status

Determines whether a string message meets the channel’s AutoMod requirements. check-automod-status

delete_chat_messages

Removes a single chat message or all chat messages from the broadcaster’s chat room. delete-chat-messages

get_automod_settings

Gets the broadcaster’s AutoMod settings. get-automod-settings

get_banned_users

Returns all banned and timed-out users in a channel. get-banned-users

get_blocked_terms

Gets the broadcaster’s list of non-private, blocked words or phrases. These are the terms that the broadcaster or moderator added manually, or that were denied by AutoMod. get-blocked-terms

get_moderated_channels

Gets a list of channels that the specified user has moderator privileges in. get-moderated-channels

get_moderators

Returns all moderators in a channel. get-moderators

get_shield_mode_status

Gets the broadcaster’s Shield Mode activation status. get-shield-mode-status

get_unban_requests

Gets a list of unban requests for a broadcaster’s channel. get-unban-requests

get_vips

Gets a list of the channel’s VIPs. get-vips

manage_held_automod_messages

Allow or deny a message that was held for review by AutoMod. manage-held-automod-messages

remove_blocked_term

Removes the word or phrase that the broadcaster is blocking users from using in their chat room. remove-blocked-term

remove_channel_moderator

Removes a moderator from the broadcaster’s chat room. remove-channel-moderator

remove_channel_vip

Removes a VIP from the broadcaster’s chat room. remove-channel-vip

resolve_unban_request

Resolves an unban request by approving or denying it.

unban_user

Removes the ban or timeout that was placed on the specified user. unban-user

update_automod_settings

Updates the broadcaster’s AutoMod settings.

update_shield_mode_status

Activates or deactivates the broadcaster’s Shield Mode.

warn_chat_userbeta

Warns a user in a broadcaster’s chat room, preventing them from chatting until the warning is acknowledged. warn-chat-user

Structs

AddBlockedTermBody

Body Parameters for Add Blocked Term

AddBlockedTermRequest

Query Parameters for Add Blocked Term

AddChannelModeratorRequest

Query Parameters for Add Channel Moderator

AddChannelVipRequest

Query Parameters for Add Channel Vip

AutoModSettings

Return Values for Get AutoMod Settings

BanUser

Return Values for Ban User

BanUserBody

Body Parameters for Ban User

BanUserRequest

Query Parameters for Ban User

BannedUser

Return Values for Get Banned Users

BlockedTerm

A blocked term in automod

CheckAutoModStatus

Return Values for Check AutoMod Status

CheckAutoModStatusBody

Body Parameters for Check AutoMod Status

CheckAutoModStatusRequest

Query Parameters for Check AutoMod Status

DeleteChatMessagesRequest

Query Parameters for Delete Chat Messages

GetAutoModSettingsRequest

Query Parameters for Get AutoMod Settings

GetBannedUsersRequest

Query Parameters for Get Banned Users

GetModeratedChannelsRequest

Query Parameters for Get Moderated Channels

GetModeratorsRequest

Query Parameters for Get Moderators

GetShieldModeStatusRequest

Query Parameters for Get Shield Mode Status

GetUnbanRequestsRequest

Query Parameters for Get Unban Requests

GetVipsRequest

Query Parameters for Get VIPs

LastShieldMode

Information about a shield mode

ManageHeldAutoModMessagesBody

Body Parameters for Manage Held AutoMod Messages

ManageHeldAutoModMessagesRequest

Query Parameters for Manage Held AutoMod Messages

ModeratedChannel

Return Values for Get Moderated Channels

Moderator

Return Values for Get Moderators

RemoveBlockedTermRequest

Query Parameters for Remove Blocked Terms

RemoveChannelModeratorRequest

Query Parameters for Remove Channel Moderator

RemoveChannelVipRequest

Query Parameters for Remove Channel VIP

ResolveUnbanRequest

Query Parameters for Resolve Unban Request

ShieldModeStatus

Return Values for Get Shield Mode Status

UnbanRequest

Return Values for Get Unban Requests

UnbanUserRequest

Query Parameters for Unban User

UpdateAutoModSettingsIndividual

Set the individual levels for each setting

UpdateAutoModSettingsRequest

Query Parameters for Update AutoMod Settings

Vip

Return Values for Get VIPs

WarnChatUserbeta

Return Values for Warn Chat User

WarnChatUserBodybeta

Body Parameters for Warn Chat User

WarnChatUserRequestbeta

Query Parameters for Warn Chat User

Enums

AddChannelModeratorResponse

Return Values for Add Channel Moderator

AddChannelVipResponse

Return Values for Add Channel Vip

AutoModAction

Action to take for a message.

DeleteChatMessagesResponse

Return Values for Delete Chat Messages

ManageHeldAutoModMessages

Return Values for Manage Held AutoMod Messages

RemoveBlockedTerm

Return Values for Remove Blocked Terms

RemoveChannelModeratorResponse

Return Values for Remove Channel Moderator

RemoveChannelVipResponse

Return Values for Remove Channel VIP

UnbanRequestStatus

The status of an unban request

UnbanUserResponse

Return Values for Unban User

UpdateAutoModSettingsBody

Body Parameters for Update AutoMod Settings