Module twitch_api::helix::moderation

Available on crate feature helix only.

Helix endpoints regarding moderation

Modules

add_blocked_term

Adds a word or phrase to the broadcaster’s list of blocked terms. These are the terms that broadcasters don’t want used in their chat room. add-blocked-term

add_channel_moderator

Adds a moderator to the broadcaster’s moderation room. add-channel-moderator

ban_user

Bans a user from participating in a broadcaster’s chat room, or puts them in a timeout. ban-user

check_automod_status

Determines whether a string message meets the channel’s AutoMod requirements. check-automod-status

delete_chat_messages

Removes a single chat message or all chat messages from the broadcaster’s chat room. delete-chat-messages

get_banned_users

Returns all banned and timed-out users in a channel. get-banned-users

get_blocked_terms

Gets the broadcaster’s list of non-private, blocked words or phrases. These are the terms that the broadcaster or moderator added manually, or that were denied by AutoMod. get-blocked-terms

get_moderators

Returns all moderators in a channel. get-moderators

get_shield_mode_status

Gets the broadcaster’s Shield Mode activation status. get-shield-mode-status

manage_held_automod_messages

Allow or deny a message that was held for review by AutoMod. manage-held-automod-messages

remove_blocked_term

Removes the word or phrase that the broadcaster is blocking users from using in their chat room. remove-blocked-term

remove_channel_moderator

Removes a moderator from the broadcaster’s chat room. remove-channel-moderator

unban_user

Removes the ban or timeout that was placed on the specified user. unban-user

update_shield_mode_status

Activates or deactivates the broadcaster’s Shield Mode.

Structs

AddBlockedTermBody

Body Parameters for Add Blocked Term

AddBlockedTermRequest

Query Parameters for Add Blocked Term

AddChannelModeratorRequest

Query Parameters for Add Channel Moderator

BanUser

Return Values for Ban User

BanUserBody

Body Parameters for Ban User

BanUserRequest

Query Parameters for Ban User

BannedUser

Return Values for Get Banned Users

BlockedTerm

A blocked term in automod

CheckAutoModStatus

Return Values for Check AutoMod Status

CheckAutoModStatusBody

Body Parameters for Check AutoMod Status

CheckAutoModStatusRequest

Query Parameters for Check AutoMod Status

DeleteChatMessagesRequest

Query Parameters for Delete Chat Messages

GetBannedUsersRequest

Query Parameters for Get Banned Users

GetModeratorsRequest

Query Parameters for Get Moderators

GetShieldModeStatusRequest

Query Parameters for Get Shield Mode Status

LastShieldMode

Information about a shield mode

ManageHeldAutoModMessagesBody

Body Parameters for Manage Held AutoMod Messages

ManageHeldAutoModMessagesRequest

Query Parameters for Manage Held AutoMod Messages

Moderator

Return Values for Get Moderators

RemoveBlockedTermRequest

Query Parameters for Remove Blocked Terms

RemoveChannelModeratorRequest

Query Parameters for Remove Channel Moderator

ShieldModeStatus

Return Values for Get Shield Mode Status

UnbanUserRequest

Query Parameters for Unban User

Enums

AddChannelModeratorResponse

Return Values for Add Channel Moderator

AutoModAction

Action to take for a message.

DeleteChatMessagesResponse

Return Values for Delete Chat Messages

ManageHeldAutoModMessages

Return Values for Manage Held AutoMod Messages

RemoveBlockedTerm

Return Values for Remove Blocked Terms

RemoveChannelModeratorResponse

Return Values for Remove Channel Moderator

UnbanUserResponse

Return Values for Unban User