twitch_api::helix
Module moderation

Available on crate feature helix only.

Helix endpoints regarding moderation

Modules

add_blocked_term

Adds a word or phrase to the broadcaster’s list of blocked terms. These are the terms that broadcasters don’t want used in their chat room. add-blocked-term

add_channel_moderator

Adds a moderator to the broadcaster’s moderation room. add-channel-moderator

ban_user

Bans a user from participating in a broadcaster’s chat room, or puts them in a timeout. ban-user

check_automod_status

Determines whether a string message meets the channel’s AutoMod requirements. check-automod-status

delete_chat_messages

Removes a single chat message or all chat messages from the broadcaster’s chat room. delete-chat-messages

get_automod_settings

Gets the broadcaster’s AutoMod settings. get-automod-settings

get_banned_users

Returns all banned and timed-out users in a channel. get-banned-users

get_blocked_terms

Gets the broadcaster’s list of non-private, blocked words or phrases. These are the terms that the broadcaster or moderator added manually, or that were denied by AutoMod. get-blocked-terms

get_moderated_channels

Gets a list of channels that the specified user has moderator privileges in. get-moderated-channels

get_moderators

Returns all moderators in a channel. get-moderators

get_shield_mode_status

Gets the broadcaster’s Shield Mode activation status. get-shield-mode-status

get_unban_requests

Gets a list of unban requests for a broadcaster’s channel. get-unban-requests

manage_held_automod_messages

Allow or deny a message that was held for review by AutoMod. manage-held-automod-messages

remove_blocked_term

Removes the word or phrase that the broadcaster is blocking users from using in their chat room. remove-blocked-term

remove_channel_moderator

Removes a moderator from the broadcaster’s chat room. remove-channel-moderator

resolve_unban_request

Resolves an unban request by approving or denying it.

unban_user

Removes the ban or timeout that was placed on the specified user. unban-user

update_automod_settings

Updates the broadcaster’s AutoMod settings.

update_shield_mode_status

Activates or deactivates the broadcaster’s Shield Mode.

warn_chat_userbeta

Warns a user in a broadcaster’s chat room, preventing them from chatting until the warning is acknowledged. warn-chat-user

Structs

AddBlockedTermBody

Body Parameters for Add Blocked Term

AddBlockedTermRequest

Query Parameters for Add Blocked Term

AddChannelModeratorRequest

Query Parameters for Add Channel Moderator

AutoModSettings

Return Values for Get AutoMod Settings

BanUser

Return Values for Ban User

BanUserBody

Body Parameters for Ban User

BanUserRequest

Query Parameters for Ban User

BannedUser

Return Values for Get Banned Users

BlockedTerm

A blocked term in automod

CheckAutoModStatus

Return Values for Check AutoMod Status

CheckAutoModStatusBody

Body Parameters for Check AutoMod Status

CheckAutoModStatusRequest

Query Parameters for Check AutoMod Status

DeleteChatMessagesRequest

Query Parameters for Delete Chat Messages

GetAutoModSettingsRequest

Query Parameters for Get AutoMod Settings

GetBannedUsersRequest

Query Parameters for Get Banned Users

GetModeratedChannelsRequest

Query Parameters for Get Moderated Channels

GetModeratorsRequest

Query Parameters for Get Moderators

GetShieldModeStatusRequest

Query Parameters for Get Shield Mode Status

GetUnbanRequestsRequest

Query Parameters for Get Unban Requests

LastShieldMode

Information about a shield mode

ManageHeldAutoModMessagesBody

Body Parameters for Manage Held AutoMod Messages

ManageHeldAutoModMessagesRequest

Query Parameters for Manage Held AutoMod Messages

ModeratedChannel

Return Values for Get Moderated Channels

Moderator

Return Values for Get Moderators

RemoveBlockedTermRequest

Query Parameters for Remove Blocked Terms

RemoveChannelModeratorRequest

Query Parameters for Remove Channel Moderator

ResolveUnbanRequest

Query Parameters for Resolve Unban Request

ShieldModeStatus

Return Values for Get Shield Mode Status

UnbanRequest

Return Values for Get Unban Requests

UnbanUserRequest

Query Parameters for Unban User

UpdateAutoModSettingsIndividual

Set the individual levels for each setting

UpdateAutoModSettingsRequest

Query Parameters for Update AutoMod Settings

WarnChatUserbeta

Return Values for Warn Chat User

WarnChatUserBodybeta

Body Parameters for Warn Chat User

WarnChatUserRequestbeta

Query Parameters for Warn Chat User

Enums

AddChannelModeratorResponse

Return Values for Add Channel Moderator

AutoModAction

Action to take for a message.

DeleteChatMessagesResponse

Return Values for Delete Chat Messages

ManageHeldAutoModMessages

Return Values for Manage Held AutoMod Messages

RemoveBlockedTerm

Return Values for Remove Blocked Terms

RemoveChannelModeratorResponse

Return Values for Remove Channel Moderator

UnbanRequestStatus

The status of an unban request

UnbanUserResponse

Return Values for Unban User

UpdateAutoModSettingsBody

Body Parameters for Update AutoMod Settings