twitch_api::helix::moderation

Module add_blocked_term

Available on crate feature helix only.

Adds a word or phrase to the broadcaster’s list of blocked terms. These are the terms that broadcasters don’t want used in their chat room. add-blocked-term

Accessing the endpoint

Request: AddBlockedTermRequest

To use this endpoint, construct a AddBlockedTermRequest with the AddBlockedTermRequest::new() method.

use twitch_api::helix::moderation::add_blocked_term;
let request = add_blocked_term::AddBlockedTermRequest::new("1234", "5678");

Body: AddBlockedTermBody

We also need to provide a body to the request containing what we want to change.

let body =
    add_blocked_term::AddBlockedTermBody::new("A phrase I'm not fond of");

Response: BlockedTerm

Send the request to receive the response with HelixClient::req_post().

use twitch_api::helix::{self, moderation::add_blocked_term};
let request = add_blocked_term::AddBlockedTermRequest::new("1234", "5678");
let body = add_blocked_term::AddBlockedTermBody::new("A phrase I'm not fond of");
let response: &helix::moderation::BlockedTerm = client.req_post(request, body, &token).await?.data.first().unwrap();

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with AddBlockedTermRequest::parse_response(None, &request.get_uri(), response)

Structs

AddBlockedTermBody

Body Parameters for Add Blocked Term

AddBlockedTermRequest

Query Parameters for Add Blocked Term

Type Aliases

AddBlockedTermResponse

Return Values for Add Blocked Term