twitch_api::helix::points

Module get_custom_reward_redemption

Available on crate feature helix only.

Returns Custom Reward Redemption objects for a Custom Reward on a channel that was created by the same client_id.

Developers only have access to get and update redemptions for the rewards they created. get-custom-reward-redemption

Accessing the endpoint

Request: GetCustomRewardRedemptionRequest

To use this endpoint, construct a GetCustomRewardRedemptionRequest with the GetCustomRewardRedemptionRequest::broadcaster_id() method.

use twitch_api::helix::points::{
    CustomRewardRedemptionStatus, GetCustomRewardRedemptionRequest,
};
let request = GetCustomRewardRedemptionRequest::broadcaster_id("274637212")
    .reward_id("92af127c-7326-4483-a52b-b0da0be61c01")
    .status(CustomRewardRedemptionStatus::Canceled);

Response: CustomRewardRedemption

Send the request to receive the response with HelixClient::req_get().

use twitch_api::helix;
use twitch_api::helix::points::{CustomRewardRedemptionStatus, CustomRewardRedemption, GetCustomRewardRedemptionRequest};
let request = GetCustomRewardRedemptionRequest::broadcaster_id("274637212")
    .reward_id("92af127c-7326-4483-a52b-b0da0be61c01")
    .status(CustomRewardRedemptionStatus::Canceled);
let response: Vec<CustomRewardRedemption> = client.req_get(request, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with GetCustomRewardRedemptionRequest::parse_response(None, &request.get_uri(), response)

Structs

CustomRewardRedemption

Return Values for Get Custom Reward Redemption

GetCustomRewardRedemptionRequest

Query Parameters for Get Custom Reward Redemption

Reward

Information about the reward involved

Enums

GetCustomRewardRedemptionSortOrder

The order to sort redemptions by.