Type Alias twitch_api::helix::points::CreateCustomRewardResponse

pub type CreateCustomRewardResponse = CustomReward;

Available on crate feature helix only.

Return Values for Create Custom Rewards

create-custom-rewards

Aliased Type

struct CreateCustomRewardResponse {

    pub broadcaster_id: UserId,
    pub broadcaster_login: Nickname,
    pub broadcaster_name: DisplayName,
    pub id: RewardId,
    pub title: String,
    pub prompt: String,
    pub cost: usize,
    pub image: Option<Image>,
    pub default_image: Option<Image>,
    pub background_color: String,
    pub is_enabled: bool,
    pub is_user_input_required: bool,
    pub max_per_stream_setting: Max,
    pub max_per_user_per_stream_setting: Max,
    pub global_cooldown_setting: GlobalCooldown,
    pub is_paused: bool,
    pub is_in_stock: bool,
    pub should_redemptions_skip_request_queue: bool,
    pub redemptions_redeemed_current_stream: Option<usize>,
    pub cooldown_expires_at: Option<Timestamp>,

}

Fields
broadcaster_id: UserId

ID of the channel the reward is for

broadcaster_login: Nickname

Login of the channel the reward is for

broadcaster_name: DisplayName

Display name of the channel the reward is for

id: RewardId

ID of the reward

title: String

The title of the reward

prompt: String

The prompt for the viewer when they are redeeming the reward

cost: usize

The cost of the reward

image: Option<Image>

Set of custom images of 1x, 2x and 4x sizes for the reward { url_1x: string, url_2x: string, url_4x: string }, can be null if no images have been uploaded

default_image: Option<Image>

Set of default images of 1x, 2x and 4x sizes for the reward { url_1x: string, url_2x: string, url_4x: string }

background_color: String

Custom background color for the reward. Format: Hex with # prefix. Example: #00E5CB.

is_enabled: bool

Is the reward currently enabled, if false the reward won’t show up to viewers

is_user_input_required: bool

Does the user need to enter information when redeeming the reward

max_per_stream_setting: Max

Whether a maximum per stream is enabled and what the maximum is. { is_enabled: bool, max_per_stream: int }

max_per_user_per_stream_setting: Max

Whether a maximum per user per stream is enabled and what the maximum is. { is_enabled: bool, max_per_user_per_stream: int }

global_cooldown_setting: GlobalCooldown

Whether a cooldown is enabled and what the cooldown is. { is_enabled: bool, global_cooldown_seconds: int }

is_paused: bool

Is the reward currently paused, if true viewers can’t redeem

is_in_stock: bool

Is the reward currently in stock, if false viewers can’t redeem

should_redemptions_skip_request_queue: bool

Should redemptions be set to FULFILLED status immediately when redeemed and skip the request queue instead of the normal UNFULFILLED status.

redemptions_redeemed_current_stream: Option<usize>

The number of redemptions redeemed during the current live stream. Counts against the max_per_stream_setting limit. Null if the broadcasters stream isn’t live or max_per_stream_setting isn’t enabled.

cooldown_expires_at: Option<Timestamp>

Timestamp of the cooldown expiration. Null if the reward isn’t on cooldown.