Module twitch_api::helix::points::update_custom_reward

Available on crate feature helix only.

Updates a Custom Reward created on a channel.

Only rewards created programmatically by the same client_id can be updated. update-custom-reward

Accessing the endpoint

Request: UpdateCustomRewardRequest

To use this endpoint, construct an UpdateCustomRewardRequest with the UpdateCustomRewardRequest::new() method.

use twitch_api::helix::points::update_custom_reward;
let request = update_custom_reward::UpdateCustomRewardRequest::new(
    "274637212",
    "reward-id",
);

Body: UpdateCustomRewardBody

We also need to provide a body to the request containing what we want to change.

let mut body = update_custom_reward::UpdateCustomRewardBody::default();
body.cost = Some(501);
body.title = Some("hydrate but differently now!".into());

Response: UpdateCustomReward

Send the request to receive the response with HelixClient::req_patch().

use twitch_api::helix::{self, points::update_custom_reward};
let request = update_custom_reward::UpdateCustomRewardRequest::new("274637212", "reward-id");
let mut body = update_custom_reward::UpdateCustomRewardBody::default();
body.cost = Some(501);
body.title = Some("hydrate but differently now!".into());
let response: update_custom_reward::UpdateCustomReward = client.req_patch(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with UpdateCustomRewardRequest::parse_response(None, &request.get_uri(), response)

Structs

UpdateCustomRewardBody

Body Parameters for Update Custom Rewards

UpdateCustomRewardRequest

Query Parameters for Update Custom Rewards

Enums

UpdateCustomReward

Return Values for Update CustomReward