twitch_api::helix::polls

Module end_poll

Available on crate feature helix only.

End a poll that is currently active.

end-poll

Accessing the endpoint

Request: EndPollRequest

To use this endpoint, construct an EndPollRequest with the EndPollRequest::new() method.

use twitch_api::helix::polls::end_poll;
let request = end_poll::EndPollRequest::new();

Body: EndPollBody

We also need to provide a body to the request containing what we want to change.

let body = end_poll::EndPollBody::new(
    "274637212",
    "92af127c-7326-4483-a52b-b0da0be61c01",
    end_poll::PollStatus::Terminated,
);

Response: EndPoll

Send the request to receive the response with HelixClient::req_patch().

use twitch_api::helix::{self, polls::end_poll};
let request = end_poll::EndPollRequest::new();
let body = end_poll::EndPollBody::new(
    "274637212",
    "92af127c-7326-4483-a52b-b0da0be61c01",
    end_poll::PollStatus::Terminated,
);
let response: end_poll::EndPoll = client.req_patch(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with EndPollRequest::parse_response(None, &request.get_uri(), response)

Structs

EndPollBody

Body Parameters for End Poll

EndPollRequest

Query Parameters for End Poll

Enums

EndPoll

Return Values for Update CustomReward

PollStatus

Status of a poll