Module twitch_api::helix::predictions

Available on crate feature helix only.

Helix endpoints regarding channel predictions

Modules

create_prediction

Create a Channel Points Prediction for a specific Twitch channel.

end_prediction

End a prediction that is currently active.

get_predictions

Get information about all predictions or specific predictions for a Twitch channel. Prediction information is available for 90 days. get-predictions

Structs

GetPredictionsRequest

Query Parameters for Get predictions

Prediction

Return Values for Get predictions