Struct twitch_api::helix::response::Response

#[non_exhaustive]
pub struct Response<R, D>
where
    R: Request,
    D: DeserializeOwned + PartialEq,
{
    pub data: D,
    pub pagination: Option<Cursor>,
    pub request: Option<R>,
    pub total: Option<i64>,
    pub other: Option<Map<String, Value>>,
}

Available on crate feature helix only.

Response retrieved from endpoint. Data is the type in Request::Response

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

data: D

Twitch’s response field for data.

pagination: Option<Cursor>

A cursor value, to be used in a subsequent request to specify the starting point of the next set of results.

request: Option<R>

The request that was sent, used for pagination.

total: Option<i64>

Response would return this many results if fully paginated. Sometimes this is not emmitted or correct for this purpose, in those cases, this value will be None.

other: Option<Map<String, Value>>

Fields which are not part of the data response, but are returned by the endpoint.

See for example Get Broadcaster Subscriptions which returns this.

Implementations

impl Response<GetBroadcasterSubscriptionsRequest<'_>, Vec<BroadcasterSubscription>>

pub fn points(&self) -> Result<i64, BroadcasterSubscriptionPointsError>

The current number of subscriber points earned by this broadcaster.

impl<R, D> Response<R, D>
where R: Request, D: DeserializeOwned + PartialEq,

pub fn new( data: D, pagination: Option<Cursor>, request: Option<R>, total: Option<i64>, other: Option<Map<String, Value>> ) -> Self

Create a new response.

Examples

use twitch_api::helix::{self, Request, RequestPost};

#[derive(serde::Serialize, Debug, PartialEq)]
pub struct MyTwitchRequest {}

#[derive(serde::Deserialize, Debug, PartialEq)]
pub enum MyTwitchResponse {
    Success,
}

impl Request for MyTwitchRequest {
    type Response = MyTwitchResponse;

    const PATH: &'static str = "my/request";
    const SCOPE: twitch_api::twitch_oauth2::Validator =
        twitch_api::twitch_oauth2::validator![];
}

impl RequestPost for MyTwitchRequest {
    type Body = helix::EmptyBody;

    fn parse_inner_response(
        request: Option<Self>,
        uri: &http::Uri,
        response: &str,
        status: http::StatusCode,
    ) -> Result<
        helix::Response<Self, Self::Response>,
        helix::HelixRequestPostError,
    >
    where
        Self: Sized,
    {
        match status {
            http::StatusCode::NO_CONTENT => Ok(helix::Response::new(
                MyTwitchResponse::Success,
                None,
                request,
                None,
                None,
            )),
            _ => Err(helix::HelixRequestPostError::InvalidResponse {
                reason: "unexpected status",
                response: response.to_string(),
                status,
                uri: uri.clone(),
            }),
        }
    }
}

pub fn with_data(data: D, request: Option<R>) -> Self

Create a new response without pagination, total or other fields.

Examples

use twitch_api::helix::{self, Request, RequestPut};

#[derive(serde::Serialize, Debug, PartialEq)]
pub struct MyTwitchRequest {}

#[derive(serde::Deserialize, Debug, PartialEq)]
pub enum MyTwitchResponse {
    Success,
}

impl Request for MyTwitchRequest {
    type Response = MyTwitchResponse;

    const PATH: &'static str = "my/request";
    const SCOPE: twitch_api::twitch_oauth2::Validator =
        twitch_api::twitch_oauth2::validator![];
}

impl RequestPut for MyTwitchRequest {
    type Body = helix::EmptyBody;

    fn parse_inner_response(
        request: Option<Self>,
        uri: &http::Uri,
        response: &str,
        status: http::StatusCode,
    ) -> Result<
        helix::Response<Self, Self::Response>,
        helix::HelixRequestPutError,
    >
    where
        Self: Sized,
    {
        match status {
            http::StatusCode::NO_CONTENT => Ok(helix::Response::with_data(
                MyTwitchResponse::Success,
                request,
            )),
            _ => Err(helix::HelixRequestPutError::InvalidResponse {
                reason: "unexpected status",
                response: response.to_string(),
                status,
                uri: uri.clone(),
            }),
        }
    }
}

pub fn get_other<Q, V>(&self, key: &Q) -> Result<Option<V>, Error>
where String: Borrow<Q>, Q: ?Sized + Ord + Eq + Hash, V: DeserializeOwned,

Get a field from the response that is not part of data.

impl<R, D, T> Response<R, D>
where R: Request, D: IntoIterator<Item = T> + PartialEq + DeserializeOwned,

pub fn first(self) -> Option<T>

Get first result of this response.

impl<R, D> Response<R, D>
where R: Request<Response = D> + Clone + Paginated + RequestGet + Debug, D: DeserializeOwned + Debug + PartialEq,

pub async fn get_next<'a, C: HttpClient + 'a>( self, client: &'a HelixClient<'a, C>, token: &(impl TwitchToken + ?Sized) ) -> Result<Option<Response<R, D>>, ClientRequestError<<C as HttpClient>::Error>>

Available on crate feature client only.

Get the next page in the responses.