twitch_api::helix
Module schedule

Available on crate feature helix only.

Helix endpoints regarding stream schedules

Modules

create_channel_stream_schedule_segment

Create a single scheduled broadcast or a recurring scheduled broadcast for a channel’s stream schedule. create-channel-stream-schedule-segment

delete_channel_stream_schedule_segment

Delete a single scheduled broadcast or a recurring scheduled broadcast for a channel’s stream schedule. delete-channel-stream-schedule-segment

get_channel_stream_schedule

Gets all scheduled broadcasts or specific scheduled broadcasts from a channel’s stream schedule. Scheduled broadcasts are defined as “stream segments” in the API. get-channel-stream-schedule

update_channel_stream_schedule

Update the settings for a channel’s stream schedule. This can be used for setting vacation details.

update_channel_stream_schedule_segment

Update a single scheduled broadcast or a recurring scheduled broadcast for a channel’s stream schedule. update-channel-stream-schedule-segment

Structs

Category

The category for a scheduled broadcast.

CreateChannelStreamScheduleSegmentBody

Body Parameters for Create Channel Stream Schedule Segment

CreateChannelStreamScheduleSegmentRequest

Query Parameters for Create Channel Stream Schedule Segment

DeleteChannelStreamScheduleSegmentRequest

Query Parameters for Delete Channel Stream Schedule Segment

GetChannelStreamScheduleRequest

Query Parameters for Get Channel Stream Schedule

ScheduledBroadcasts

Scheduled broadcasts for a specific channel given by Get Channel Stream Schedule

Segment

Scheduled broadcast for a stream schedule.

UpdateChannelStreamScheduleRequest

Query Parameters for Update Channel Stream Schedule

Vacation

Information on Vacation Mode

Enums

DeleteChannelStreamScheduleSegment

Return Values for Delete Channel Stream Schedule Segment

UpdateChannelStreamSchedule

Return Values for Update Channel Stream Schedule