twitch_api::helix::schedule
Module create_channel_stream_schedule_segment

Available on crate feature helix only.

Create a single scheduled broadcast or a recurring scheduled broadcast for a channel’s stream schedule. create-channel-stream-schedule-segment

Accessing the endpoint

Request: CreateChannelStreamScheduleSegmentRequest

To use this endpoint, construct a CreateChannelStreamScheduleSegmentRequest with the CreateChannelStreamScheduleSegmentRequest::broadcaster_id() method.

use twitch_api::helix::schedule::create_channel_stream_schedule_segment;
let request =
    create_channel_stream_schedule_segment::CreateChannelStreamScheduleSegmentRequest::broadcaster_id("141981764");

Body: CreateChannelStreamScheduleSegmentBody

We also need to provide a body to the request containing what we want to change.

let mut body =
    create_channel_stream_schedule_segment::CreateChannelStreamScheduleSegmentBody::new(
        twitch_api::types::Timestamp::try_from("2021-07-01T18:00:00Z").unwrap(),
        "America/New_York",
        false,
    );
body.duration = Some("60".into());
body.category_id = Some(twitch_types::CategoryIdRef::from_static("509670").as_cow());
body.title = Some("TwitchDev Monthly Update // July 1, 2021".into());

Response: ScheduledBroadcasts

Send the request to receive the response with HelixClient::req_post().

use twitch_api::helix::{self, schedule::create_channel_stream_schedule_segment};
let request =
    create_channel_stream_schedule_segment::CreateChannelStreamScheduleSegmentRequest::broadcaster_id("141981764");
let timestamp = twitch_api::types::Timestamp::try_from("2021-07-01T18:00:00Z")?;
let mut body =
    create_channel_stream_schedule_segment::CreateChannelStreamScheduleSegmentBody::new(
        twitch_api::types::Timestamp::try_from("2021-07-01T18:00:00Z").unwrap(),
        "America/New_York",
        false,
    );
body.duration = Some("60".into());
body.category_id = Some(twitch_types::CategoryIdRef::from_static("509670").as_cow());
body.title = Some("TwitchDev Monthly Update // July 1, 2021".into());
let response: create_channel_stream_schedule_segment::CreateChannelStreamScheduleSegmentResponse = client.req_post(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&body, &token, &client_id) and parse the http::Response with CreateChannelStreamScheduleSegmentRequest::parse_response(None, &request.get_uri(), response)

Structs

CreateChannelStreamScheduleSegmentBody

Body Parameters for Create Channel Stream Schedule Segment

CreateChannelStreamScheduleSegmentRequest

Query Parameters for Create Channel Stream Schedule Segment

Type Aliases

CreateChannelStreamScheduleSegmentResponse

Return Values for Create Channel Stream Schedule Segment