twitch_api::helix
Module search

Available on crate feature helix only.

Helix endpoints regarding search

Examples

let client = HelixClient::new();
let req = SearchCategoriesRequest::query("Pokémon");

println!("{:?}", &client.req_get(req, &token).await?.data);

Modules

search_categories

Returns a list of games or categories that match the query via name either entirely or partially. search-categories

search_channels

Returns a list of channels (users who have streamed within the past 6 months) that match the query via channel name or description either entirely or partially. search-channels

Structs

Category

A game or category as defined by Twitch

Channel

Return Values for Search Channels

SearchCategoriesRequest

Query Parameters for Search Categories

SearchChannelsRequest

Query Parameters for Search Channels