twitch_api::helix

Module streams

Available on crate feature helix only.

Helix endpoints regarding streams

Examples

See HelixClient::get_streams_from_logins and HelixClient::get_streams_from_ids

Manual request

let client = HelixClient::new();
let req = GetStreamsRequest::user_logins(&"justinfan1337");

// If this doesn't return a result, that would mean the stream is not live.
println!("{:?}", &client.req_get(req, &token).await?.data.first());

Implemented endpoints

Streams 🟢 5/5

Endpoint	Helper	Module

Get Stream Key	HelixClient::get_stream_key	get_stream_key

Get Streams	HelixClient::get_streams_from_ids, HelixClient::get_streams_from_logins	get_streams

Get Followed Streams	HelixClient::get_followed_streams	get_followed_streams

Create Stream Marker	HelixClient::create_stream_marker	create_stream_marker

Get Stream Markers	-	get_stream_markers

Modules

create_stream_marker

Adds a marker to a live stream. create-stream-marker

get_followed_streams

Gets information about active streams belonging to channels that the authenticated user follows. get-followed-streams

get_stream_key

Gets the channel’s stream key. get-stream-key

get_stream_markers

Gets a list of markers from the user’s most recent stream or from the specified VOD/video. get-stream-markers

get_stream_tags

Gets the list of tags for a specified stream (channel). get-stream-tags

get_streams

Gets information about active streams. get-streams

replace_stream_tags

Applies specified tags to a specified stream, overwriting any existing tags applied to that stream. If no tags are specified, all tags previously applied to the stream are removed. Automated tags are not affected by this operation. replace-stream-tags

Structs

CreateStreamMarkerBody

Body Parameters for Create Stream Marker

CreateStreamMarkerRequest

Query Parameters for Create Stream Marker

CreatedStreamMarker

Return Value for Create Stream Marker

GetFollowedStreamsRequest

Query Parameters for Get Followed Streams

GetStreamKeyRequest

Query Parameters for Get Stream Key

GetStreamKeyResponse

Return Values for Get Stream Key

GetStreamMarkersRequest

Query Parameters for Get Stream Markers

GetStreamTagsRequestDeprecated

Query Parameters for Get Stream Tags

GetStreamsRequest

Query Parameters for Get Streams

ReplaceStreamTagsBodyDeprecated

Body Parameters for Replace Stream Tags

ReplaceStreamTagsRequestDeprecated

Query Parameters for Replace Stream Tags

Stream

Return Values for Get Streams

StreamMarker

A marker on a video

StreamMarkerGroup

Return Values for Get Stream Markers

StreamMarkerVideo

A video with markers

Enums

ReplaceStreamTagsDeprecated

Return Values for Replace Stream Tags

StreamType

Gotten from Stream.type_

Type Aliases

Tag

Return Values for Get Stream Tags