Module twitch_api::helix::streams

Available on crate feature helix only.

Helix endpoints regarding streams

Examples

let client = HelixClient::new();
let logins: &[&types::UserNameRef] = &["justinfan1337".into()];
let req = GetStreamsRequest::builder().user_login(logins).build();

// If this doesn't return a result, that would mean the stream is not live.
println!("{:?}", &client.req_get(req, &token).await?.data.get(0));

Modules

get_followed_streams

Gets information about active streams belonging to channels that the authenticated user follows. get-followed-streams

get_stream_tags

Gets the list of tags for a specified stream (channel). get-stream-tags

get_streams

Gets information about active streams. get-streams

replace_stream_tags

Applies specified tags to a specified stream, overwriting any existing tags applied to that stream. If no tags are specified, all tags previously applied to the stream are removed. Automated tags are not affected by this operation. replace-stream-tags

Structs

GetFollowedStreamsRequest

Query Parameters for Get Followed Streams

GetStreamTagsRequestDeprecated

Query Parameters for Get Stream Tags

GetStreamsRequest

Query Parameters for Get Streams

ReplaceStreamTagsBodyDeprecated

Body Parameters for Replace Stream Tags

ReplaceStreamTagsRequestDeprecated

Query Parameters for Replace Stream Tags

Stream

Return Values for Get Streams

Enums

ReplaceStreamTagsDeprecated

Return Values for Replace Stream Tags

StreamType

Gotten from Stream.type_

Type Definitions

Tag

Return Values for Get Stream Tags