twitch_api::helix::streams::get_followed_streams
Type Alias GetFollowedStreamsResponse

pub type GetFollowedStreamsResponse = Stream;

Available on crate feature helix only.

Return Values for Get Followed Streams

get-followed-streams

Aliased Type

struct GetFollowedStreamsResponse {

    pub game_id: CategoryId,
    pub game_name: String,
    pub id: StreamId,
    pub language: String,
    pub is_mature: bool,
    pub started_at: Timestamp,
    pub tag_ids: Vec<TagId>,
    pub tags: Vec<String>,
    pub thumbnail_url: String,
    pub title: String,
    pub type_: StreamType,
    pub user_id: UserId,
    pub user_name: DisplayName,
    pub user_login: Nickname,
    pub viewer_count: usize,

}

Fields
game_id: CategoryId

ID of the game being played on the stream.

game_name: String

Name of the game being played.

id: StreamId

Stream ID.

language: String

Stream language.

is_mature: bool

Indicates if the broadcaster has specified their channel contains mature content that may be inappropriate for younger audiences.

started_at: Timestamp

UTC timestamp.

tag_ids: Vec<TagId>
๐Deprecated: use tags instead

Shows tag IDs that apply to the stream.

tags: Vec<String>

The tags applied to the stream.

thumbnail_url: String

Thumbnail URL of the stream. All image URLs have variable width and height. You can replace {width} and {height} with any values to get that size image

title: String

Stream title.

type_: StreamType

Stream type: โliveโ or โโ (in case of error).

user_id: UserId

ID of the user who is streaming.

user_name: DisplayName

Display name corresponding to user_id.

user_login: Nickname

Login of the user who is streaming.

viewer_count: usize

Number of viewers watching the stream at the time of the query.