twitch_api::helix
Trait Request

pub trait Request: Serialize {
    type Response: DeserializeOwned + PartialEq;

    const PATH: &'static str;
    const SCOPE: Validator;
    const OPT_SCOPE: &'static [Scope] = _;

    // Provided methods
    fn query(&self) -> Result<String, SerializeError> { ... }
    fn get_uri(&self) -> Result<Uri, InvalidUri> { ... }
    fn get_bare_uri() -> Result<Uri, InvalidUri> { ... }
}

Available on crate feature helix only.

A request is a Twitch endpoint, see New Twitch API reference

Required Associated Constants

const PATH: &'static str

The path to the endpoint relative to the helix root. eg. channels for Get Channel Information

const SCOPE: Validator

Available on crate feature twitch_oauth2 only.

Scopes needed for this endpoint

Provided Associated Constants

const OPT_SCOPE: &'static [Scope] = _

Available on crate feature twitch_oauth2 only.

Optional scopes needed by this endpoint

Required Associated Types

type Response: DeserializeOwned + PartialEq

Response type. twitch’s response will deserialize to this.

Provided Methods

fn query(&self) -> Result<String, SerializeError>

Defines layout of the url parameters.

fn get_uri(&self) -> Result<Uri, InvalidUri>

Returns full URI for the request, including query parameters.

fn get_bare_uri() -> Result<Uri, InvalidUri>

Returns bare URI for the request, NOT including query parameters.

Object Safety

This trait is not object safe.