Module twitch_api::helix::users

Available on crate feature helix only.

Helix endpoints regarding users

Examples

let client = HelixClient::new();
let req = GetUsersRequest::builder()
    .login(vec!["justinfan1337".into()])
    .build();

println!("{:?}", &client.req_get(req, &token).await?.data);

Modules

block_user

Blocks the specified user on behalf of the authenticated user. block-user

get_user_block_list

Gets a specified user’s block list. The list is sorted by when the block occurred in descending order (i.e. most recent block first). get-user-block-list

get_users

Gets information about one or more specified Twitch users. get-users

get_users_follows

Gets information on follow relationships between two Twitch users. get-users-follows

unblock_user

Unblocks the specified user on behalf of the authenticated user. unblock-user

Structs

BlockUserRequest

Query Parameters for Block User

FollowRelationship

Describes a follow relationship

GetUserBlockListRequest

Query Parameters for Get Users Block List

GetUsersFollowsRequest

Query Parameters for Get Users Follows

GetUsersRequest

Query Parameters for Get Users

UnblockUserRequest

Query Parameters for Unblock User

User

Return Values for Get Users

UserBlock

Return Values for Get Users Block List

UsersFollows

Return Values for Get Users Follows

Enums

BlockUser

Return Values for Block User

UnblockUser

Return Values for Unblock User