twitch_api::helix

Module users

Available on crate feature helix only.

Helix endpoints regarding users

Examples

let client = HelixClient::new();
let req = GetUsersRequest::logins(&["justintvfan"]);

println!("{:?}", &client.req_get(req, &token).await?.data);

Implemented endpoints

Users 🟢 8/8

Endpoint	Helper	Module

Get Users	HelixClient::get_user_from_id, HelixClient::get_user_from_login, HelixClient::get_users_from_ids	get_users

Update User	HelixClient::update_user_description	update_user

Get User Block List	-	get_user_block_list

Block User	HelixClient::block_user	block_user

Unblock User	HelixClient::unblock_user	unblock_user

Get User Extensions	HelixClient::get_user_extensions	get_user_extensions

Get User Active Extensions	HelixClient::get_user_active_extensions	get_user_active_extensions

Update User Extensions	-	update_user_extensions

Modules

block_user

Blocks the specified user on behalf of the authenticated user. block-user

get_user_active_extensions

Gets the active extensions that the broadcaster has installed for each configuration. get-user-active-extensions

get_user_block_list

Gets a specified user’s block list. The list is sorted by when the block occurred in descending order (i.e. most recent block first). get-user-block-list

get_user_extensions

Gets a list of all extensions (both active and inactive) that the broadcaster has installed. get-user-extensions

get_users

Gets information about one or more specified Twitch users. get-users

get_users_follows

Gets information on follow relationships between two Twitch users. get-users-follows

unblock_user

Unblocks the specified user on behalf of the authenticated user. unblock-user

update_user

Updates the specified user’s information.

update_user_extensions

Updates the specified user’s information.

Structs

BlockUserRequest

Query Parameters for Block User

Extension

Return Values for Get User Extensions

ExtensionConfiguration

Return Values for Get User Active Extensions

ExtensionSpecification

Inner body for Update User Extensions

FollowRelationship

Describes a follow relationship

GetUserActiveExtensionsRequest

Query Parameters for Get User Active Extensions

GetUserBlockListRequest

Query Parameters for Get Users Block List

GetUserExtensionsRequest

Query Parameters for Get User Extensions

GetUsersFollowsRequestDeprecated

Query Parameters for Get Users Follows

GetUsersRequest

Query Parameters for Get Users

UnblockUserRequest

Query Parameters for Unblock User

UpdateUserExtensionsBody

Body for Update User Extensions

UpdateUserExtensionsRequest

Query Parameters for Update User Extensions

UpdateUserRequest

Query Parameters for Update User

User

Return Values for Get Users

UserBlock

Return Values for Get Users Block List

UsersFollowsDeprecated

Return Values for Get Users Follows

Enums

BlockUser

Return Values for Block User

ExtensionSlot

A slot for an extension to be active in

ExtensionType

Where an extension can appear.

UnblockUser

Return Values for Unblock User