twitch_api::helix::users
Module block_user

Available on crate feature helix only.

Blocks the specified user on behalf of the authenticated user. block-user

Accessing the endpoint

Request: BlockUserRequest

To use this endpoint, construct a BlockUserRequest with the BlockUserRequest::block_user() method.

use twitch_api::helix::users::block_user::{self, Reason, SourceContext};
let request = block_user::BlockUserRequest::block_user("1234");
// Or, specifying a reason for the block
let request = block_user::BlockUserRequest::block_user("1234")
    .source_context(SourceContext::Chat)
    .reason(Reason::Spam);

Response: BlockUser

Send the request to receive the response with HelixClient::req_put().

use twitch_api::helix::{self, users::block_user};
let request = block_user::BlockUserRequest::block_user("1234");
let response: block_user::BlockUser = client.req_put(request, helix::EmptyBody, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with BlockUserRequest::parse_response(None, &request.get_uri(), response)

Structs

BlockUserRequest

Query Parameters for Block User

Enums

BlockUser

Return Values for Block User

Reason

Reason for blocking the user.

SourceContext

Source context for blocking the user.