twitch_api::helix::whispers

Module send_whisper

Available on crate feature helix only.

Sends a whisper message to the specified user. send-whisper

Accessing the endpoint

Request: SendWhisperRequest

To use this endpoint, construct a SendWhisperRequest with the SendWhisperRequest::new() method.

use twitch_api::helix::whispers::send_whisper;
let request = send_whisper::SendWhisperRequest::new("456", "123");

Body: SendWhisperBody

We also need to provide a body to the request containing what we want to change.

let body = send_whisper::SendWhisperBody::new("Hello!");

Response: SendWhisperResponse

Send the request to receive the response with HelixClient::req_post().

use twitch_api::helix::{self, whispers::send_whisper};
let request = send_whisper::SendWhisperRequest::new("456", "123");
let body = send_whisper::SendWhisperBody::new("Hello!");
let response: send_whisper::SendWhisperResponse = client.req_post(request, body, &token).await?.data;

You can also get the http::Request with request.create_request(&token, &client_id) and parse the http::Response with SendWhisperRequest::parse_response(None, &request.get_uri(), response)

Structs

SendWhisperBody

Body Parameters for Send Whisper

SendWhisperRequest

Query Parameters for Send Whisper

Enums

SendWhisperResponse

Return Values for Send Whisper