twitch_api
Module pubsub

👎Deprecated since 0.7.0: use EventSub instead, see https://discuss.dev.twitch.com/t/legacy-pubsub-deprecation-and-shutdown-timeline/58043

Available on crate feature pubsub only.

Holds serializable pubsub stuff

Use listen_command() to send subscription listen and parse the responses with Response::parse

Undocumented features

This crate has some pubsub topics that are not documented by twitch. These may stop working at any time. To enable these, use feature unsupported to use them. Note that this crate doesn’t try to keep changes to these pubsub topics semver compatible.

Modules

automod_queueDeprecated

PubSub messages for when AutoMod flags a message as potentially inappropriate, and when a moderator takes action on a message.

channel_bitsDeprecated

PubSub messages for bits

channel_bits_badgeDeprecated

PubSub messages for bits

channel_cheerDeprecatedunsupported

PubSub messages for cheer events

channel_pointsDeprecated

PubSub messages for channel points

channel_sub_giftsDeprecatedunsupported

PubSub messages for sub gifts

channel_subscriptionsDeprecated

PubSub messages for subscriptions

community_pointsDeprecatedunsupported

PubSub messages for community points.

followingDeprecatedunsupported

PubSub messages for follows

hypetrainDeprecatedunsupported

PubSub messages for hype-trains

moderationDeprecated

PubSub messages for moderator actions

raidDeprecatedunsupported

PubSub messages for raids

user_moderation_notificationsDeprecated

PubSub messages for user moderation notifications

video_playbackDeprecatedunsupported

PubSub messages for (live) stream playback information

Structs

TwitchResponseDeprecated

Response from twitch PubSub

Enums

ResponseDeprecated

Response from twitchs PubSub server. Either a response indicating status of something or a message from a topic

TopicDataDeprecated

Message response from twitch PubSub.

TopicsDeprecated

All possible topics

Traits

TopicDeprecated

A logical partition of messages that clients may subscribe to, to get messages.

Functions

listen_commandDeprecated

Create a listen command.

unlisten_commandDeprecated

Create a unlisten command.