Module twitch_api::pubsub

Available on crate feature pubsub only.

Holds serializable pubsub stuff

Use listen_command() to send subscription listen and parse the responses with Response::parse

Undocumented features

This crate has some pubsub topics that are not documented by twitch. These may stop working at any time. To enable these, use feature unsupported to use them. Note that this crate doesn’t try to keep changes to these pubsub topics semver compatible.

Modules

automod_queue

PubSub messages for when AutoMod flags a message as potentially inappropriate, and when a moderator takes action on a message.

channel_bits

PubSub messages for bits

channel_bits_badge

PubSub messages for bits

channel_cheerunsupported

PubSub messages for cheer events

channel_points

PubSub messages for channel points

channel_sub_giftsunsupported

PubSub messages for sub gifts

channel_subscriptions

PubSub messages for subscriptions

community_pointsunsupported

PubSub messages for community points.

followingunsupported

PubSub messages for follows

hypetrainunsupported

PubSub messages for hype-trains

moderation

PubSub messages for moderator actions

raidunsupported

PubSub messages for raids

user_moderation_notifications

PubSub messages for user moderation notifications

video_playbackunsupported

PubSub messages for (live) stream playback information

Structs

TwitchResponse

Response from twitch PubSub

Enums

Response

Response from twitchs PubSub server. Either a response indicating status of something or a message from a topic

TopicData

Message response from twitch PubSub.

Topics

All possible topics

Traits

Topic

A logical partition of messages that clients may subscribe to, to get messages.

Functions

listen_command

Create a listen command.

unlisten_command

Create a unlisten command.