Struct twitch_api::TmiClient

pub struct TmiClient<'a, C: HttpClient + 'a> { /* private fields */ }

๐Deprecated: All TMI functionality has been implemented with helix endpoint Get Chatters

Available on crate features client and tmi only.

Client for the twitch TMI endpoint, almost entirely undocumented and certainly not supported.

Examples

let client = TmiClient::new();
println!("{:?}", client.get_chatters("justinfan10".into()).await?);

Most clients will be able to use the 'static lifetime

pub struct MyStruct {
    twitch: TmiClient<'static, reqwest::Client>,
    token: twitch_oauth2::AppAccessToken,
}
// etc

See HttpClient for implemented http clients, you can also define your own if needed.

Implementations

impl<'a, C: HttpClient + 'a> TmiClient<'a, C>

pub fn with_client(client: C) -> TmiClient<'a, C>

Create a new client with an existing client

pub fn new() -> TmiClient<'a, C>where C: ClientDefault<'a>,

Create a new TmiClient with a default HttpClient

pub fn clone_client(&self) -> Cwhere C: Clone,

Retrieve a clone of the HttpClient inside this TmiClient

pub fn get_client(&self) -> &C

Retrieve a reference of the HttpClient inside this TmiClient

pub async fn get_chatters( &'a self, broadcaster: &UserNameRef ) -> Result<GetChatters, RequestError<<C as HttpClient>::Error>>

๐Deprecated: Has been implemented with helix endpoint Get Chatters

Get all the chatters in the chat

Notes

This function will aside from url sanitize the broadcasters username, will also remove any # and make it lowercase ascii